Change require_once to include_once in some cases and hack around double-includes for lang_util
