+++ htdocs/Language/fr.php
@@ -1,16 +1,16 @@
 <?php
-$LANG_ARRAY = array ( 
-		"general" => array ( 
-		"TIPS" => "Conseils", 
-		"USERNAME" => "Username", 
-		"PWD" => "Password", 
-		"PWD_TEMP" => "Mot de pasee Temp.", 
-		"PWD_RESET" => "Restaurer le Mot de Passe", 
+$LANG_ARRAY = array (
+		"general" => array (
+		"TIPS" => "Conseils",
+		"USERNAME" => "Username",
+		"PWD" => "Password",
+		"PWD_TEMP" => "Mot de pasee Temp.",
+		"PWD_RESET" => "Restaurer le Mot de Passe",
 		"CNFRMPWD" => "Confirm Password French",
 		"PASSWORDRESET_ERR" => "Password reset expired. Mail c4gbackup@gmail.com with the subject 'Password' to enable password recovery French",
 		"PASSWORDRESET_COMPLETE" => "Password Reset enabled. Visit login.php to reset your password French",
-		"NAME" => "Nom", 
-		"PHONE" => "No. de Tel", 
+		"NAME" => "Nom",
+		"PHONE" => "No. de Tel",
 		"EMAIL" => "Email",
 		"ADD_CURRENCY" => "Add a new currency >>",
 		"NEW_CURRENCY" => "New Currency",
@@ -22,138 +22,138 @@ $LANG_ARRAY = array (
 		"UPDATED_DATE" => "Updated Date",
 		"ACTIONS" => "Actes",
 		"ADD_CURRENCY_RATE" => "Add new currency rate",
-		"CMD_LOGIN" => "Ouverture", 
-		"CMD_SEARCH" => "Chercher", 
-		"CMD_SUBMIT" => "Soumettre", 
-		"CMD_SUBMITTING" => "En soumission", 
-		"CMD_UPDATE" => "Actualiser", 
-		"MSG_UPDATED" => "Actualisé", 
-		"CMD_CANCEL" => "Quitter", 
-		"CMD_SKIP" => "Sauter/passer", 
-		"CMD_HIDE" => "Masquer", 
-		"CMD_SHOW" => "Visualiser", 
-		"REPORT_TYPE" => "Report Type", 
-		"REPORT_TO" => "Report To", 
-		"MSG_REQDFIELD" => "Ce champs est indispensable", 
-		"REPORTED" => "Enregistré", 
-		"REPORTED_NOT" => "Pas de rapport", 
-		"SKIPPED" => "Sauté/passé", 
-		"CMD_BACK" => "Retour", 
-		"CMD_GETREPORT" => "Obtenir un Rapport", 
+		"CMD_LOGIN" => "Ouverture",
+		"CMD_SEARCH" => "Chercher",
+		"CMD_SUBMIT" => "Soumettre",
+		"CMD_SUBMITTING" => "En soumission",
+		"CMD_UPDATE" => "Actualiser",
+		"MSG_UPDATED" => "Actualisé",
+		"CMD_CANCEL" => "Quitter",
+		"CMD_SKIP" => "Sauter/passer",
+		"CMD_HIDE" => "Masquer",
+		"CMD_SHOW" => "Visualiser",
+		"REPORT_TYPE" => "Report Type",
+		"REPORT_TO" => "Report To",
+		"MSG_REQDFIELD" => "Ce champs est indispensable",
+		"REPORTED" => "Enregistré",
+		"REPORTED_NOT" => "Pas de rapport",
+		"SKIPPED" => "Sauté/passé",
+		"CMD_BACK" => "Retour",
+		"CMD_GETREPORT" => "Obtenir un Rapport",
 		"CMD_GET_SPECIMEN_REPORT" => "Avoir Echantillon Rapport",
 		"CMD_GET_TEST_REPORT" => "Avoir Examine Rapport",
-		"MSG_SELECT_TTYPE" => "Sélectioner Examen en priorité", 
-		"CMD_EDIT" => "Modifier", 
-		"CMD_EXPORTWORD" => "Exporter comme un document Word", 
-		"CMD_EXPORTTXT" => "Exporter comme un Fichier Texte", 
-		"CMD_EXPORTPDF" => "Exporter comme un Doc PDF", 
-		"CMD_SELECT" => "Selectionner", 
-		"CMD_DELETE" => "Effacer", 
-		"CMD_ENTER" => "Entrer", 
-		"CMD_OK" => "OK", 
-		"CMD_PRINT" => "Imprimer", 
-		"CMD_SEARCHING" => "En recherche", 
-		"CMD_FETCH" => "Chercher", 
-		"CMD_FETCHING" => "En recherche", 
-		"MSG_DELETED" => "Supprimé", 
-		"MSG_RESET" => "Réinitialiser", 
-		"SELECT_ONE" => "Sélectionner un", 
-		"CMD_CLOSEPAGE" => "Fermer cette page", 
-		"CMD_CLOSEPREVIEW" => "Fermeture précédente", 
-		"FACILITY" => "Facilité", 
-		"LOCATION" => "Localisation", 
-		"LAB_MGR" => "Chef de Lab.", 
-		"LAB_DIR" => "Country Director", 
-		"TYPE" => "Type", 
-		"LAB_STATUS" => "Staus du Lab.", 
-		"SP_STATUS" => "Status", 
-		"MANAGE" => "Chef", 
-		"CMD_ADDNEWACCOUNT" => "Ajouter un nouveuau copmte", 
-		"CMD_THISTORY" => "Historique de examens", 
-		"DOCTOR" => "Médecin/Infirmier", 
-		"PATIENT" => "Patient", 
-		"PATIENTS" => "Patients", 
-		"PATIENT_NAME" => "Patient Name", 
-		"PATIENT_ID" => "Patient ID", 
+		"MSG_SELECT_TTYPE" => "Sélectioner Examen en priorité",
+		"CMD_EDIT" => "Modifier",
+		"CMD_EXPORTWORD" => "Exporter comme un document Word",
+		"CMD_EXPORTTXT" => "Exporter comme un Fichier Texte",
+		"CMD_EXPORTPDF" => "Exporter comme un Doc PDF",
+		"CMD_SELECT" => "Selectionner",
+		"CMD_DELETE" => "Effacer",
+		"CMD_ENTER" => "Entrer",
+		"CMD_OK" => "OK",
+		"CMD_PRINT" => "Imprimer",
+		"CMD_SEARCHING" => "En recherche",
+		"CMD_FETCH" => "Chercher",
+		"CMD_FETCHING" => "En recherche",
+		"MSG_DELETED" => "Supprimé",
+		"MSG_RESET" => "Réinitialiser",
+		"SELECT_ONE" => "Sélectionner un",
+		"CMD_CLOSEPAGE" => "Fermer cette page",
+		"CMD_CLOSEPREVIEW" => "Fermeture précédente",
+		"FACILITY" => "Facilité",
+		"LOCATION" => "Localisation",
+		"LAB_MGR" => "Chef de Lab.",
+		"LAB_DIR" => "Country Director",
+		"TYPE" => "Type",
+		"LAB_STATUS" => "Staus du Lab.",
+		"SP_STATUS" => "Status",
+		"MANAGE" => "Chef",
+		"CMD_ADDNEWACCOUNT" => "Ajouter un nouveuau copmte",
+		"CMD_THISTORY" => "Historique de examens",
+		"DOCTOR" => "Médecin/Infirmier",
+		"PATIENT" => "Patient",
+		"PATIENTS" => "Patients",
+		"PATIENT_NAME" => "Patient Name",
+		"PATIENT_ID" => "Patient ID",
 		"PATIENT_BARCODE" => "Patient Barcode French",
 		"PATIENT_SIGNATURE" => "Patient Signature French",
-		"ADDL_ID" => "ID Additionel", 
-		"GENDER" => "Sexe", 
-		"MALE" => "Masculin", 
-		"FEMALE" => "Femelle", 
+		"ADDL_ID" => "ID Additionel",
+		"GENDER" => "Sexe",
+		"MALE" => "Masculin",
+		"FEMALE" => "Femelle",
 		"OTHER" => "Other",
-		"PATIENT_DAILYNUM" => "Patient Number", 
-		"AGE" => "Age", 
-		"OR" => "Ou", 
-		"DOB" => "Date de Naissance", 
-		"YEARS" => "Ans", 
-		"MONTHS" => "mois", 
+		"PATIENT_DAILYNUM" => "Patient Number",
+		"AGE" => "Age",
+		"OR" => "Ou",
+		"DOB" => "Date de Naissance",
+		"YEARS" => "Ans",
+		"MONTHS" => "mois",
 		"WEEKS" => "Semaines",
-		"DAYS" => "Jours", 
+		"DAYS" => "Jours",
 		"HOURS" => "Heures",
-		"MINUTES" => "Minutes", 
-		"DROPDOWN" => "Menu déroulant", 
-		"FREETEXT" => "Texte Libre", 
-		"NUMERIC_FIELD" => "Numerique", 
-		"MULTISELECT" => "Sélection Multiple", 
-		"ACCESSION_NUM" => "No Accées", 
-		"SPECIMEN" => "Echantillon", 
-		"IF_AVLBL" => "Si possible", 
-		"SPECIMENS" => "Echantillons", 
-		"SPECIMEN_ID" => "ID Echantillon", 
-		"ADDL_ID_S" => "ID Additionel", 
-		"SPECIMEN_TYPE" => "Type échantillon", 
-		"SPECIMEN_TYPES" => "Types échantillons", 
-		"TEST" => "Examen", 
-		"TESTS" => "Examens", 
-		"TEST_TYPE" => "Type Examen", 
-		"TEST_TYPES" => "Types Examens", 
-		"CONTROL_TESTING_RESULTS" => "Control Results", 
-		"DATE" => "Date", 
-		"DATE_FORMAT" => "Format de la date", 
-		"R_DATE" => "Lab Receipt Date", 
-		"C_DATE" => "Date de prélèvement", 
-		"C_TIME" => "Temps de prélèvement", 
-		"E_DATE" => "Date entrée des résultats", 
-		"RE_DATE" => "Date de report", 
-		"G_DATE" => "Date", 
-		"CUMULATIVE" => "Cumulatif", 
-		"REGD_BY" => "Enregistrer par:", 
-		"COMMENTS" => "Commentaires", 
-		"TECHNICIAN" => "Technicien", 
-		"TECHNICIANS" => "Techniciens", 
-		"RESULTS" => "Resultats du Patient", 
-		"RESULT_COMMENTS" => "Remarques", 
-		"RANGE" => "Valeur Normale", 
-		"RANGE_OUTOF" => "En dehors de la valeur", 
-		"DETAILS" => "Détails", 
-		"OPTIONAL" => "Optionel", 
-		"REF_OUT" => "Refferé à:", 
-		"PENDING_RESULTS" => "Resutats manquant", 
-		"PENDING_VER" => "Vérification en Attente", 
-		"DONE" => "Complété", 
-		"REF_RETURNED" => "Referré en dehors et retouné", 
-		"MEASURES" => "Mesures", 
-		"VALUES" => "Valeurs", 
-		"OPTIONS" => "Options", 
-		"UNIT" => "Unité", 
-		"RANGE_NUMERIC" => "Valeur Numérique", 
-		"RANGE_ALPHANUM" => "Valeur alphanumérique", 
-		"RANGE_AUTOCOMPLETE" => "Autocomplete", 
-		"YES" => "Oui", 
-		"NO" => "Non", 
-		"ERROR" => "Erreur", 
-		"INVALID" => "Invalide", 
-		"NOTKNOWN" => "Non connu", 
-		"REGD_BY" => "Enregistrer par:", 
-		"ENTERED_BY" => "Saisie par:", 
+		"MINUTES" => "Minutes",
+		"DROPDOWN" => "Menu déroulant",
+		"FREETEXT" => "Texte Libre",
+		"NUMERIC_FIELD" => "Numerique",
+		"MULTISELECT" => "Sélection Multiple",
+		"ACCESSION_NUM" => "No Accées",
+		"SPECIMEN" => "Echantillon",
+		"IF_AVLBL" => "Si possible",
+		"SPECIMENS" => "Echantillons",
+		"SPECIMEN_ID" => "ID Echantillon",
+		"ADDL_ID_S" => "ID Additionel",
+		"SPECIMEN_TYPE" => "Type échantillon",
+		"SPECIMEN_TYPES" => "Types échantillons",
+		"TEST" => "Examen",
+		"TESTS" => "Examens",
+		"TEST_TYPE" => "Type Examen",
+		"TEST_TYPES" => "Types Examens",
+		"CONTROL_TESTING_RESULTS" => "Control Results",
+		"DATE" => "Date",
+		"DATE_FORMAT" => "Format de la date",
+		"R_DATE" => "Lab Receipt Date",
+		"C_DATE" => "Date de prélèvement",
+		"C_TIME" => "Temps de prélèvement",
+		"E_DATE" => "Date entrée des résultats",
+		"RE_DATE" => "Date de report",
+		"G_DATE" => "Date",
+		"CUMULATIVE" => "Cumulatif",
+		"REGD_BY" => "Enregistrer par:",
+		"COMMENTS" => "Commentaires",
+		"TECHNICIAN" => "Technicien",
+		"TECHNICIANS" => "Techniciens",
+		"RESULTS" => "Resultats du Patient",
+		"RESULT_COMMENTS" => "Remarques",
+		"RANGE" => "Valeur Normale",
+		"RANGE_OUTOF" => "En dehors de la valeur",
+		"DETAILS" => "Détails",
+		"OPTIONAL" => "Optionel",
+		"REF_OUT" => "Refferé à:",
+		"PENDING_RESULTS" => "Resutats manquant",
+		"PENDING_VER" => "Vérification en Attente",
+		"DONE" => "Complété",
+		"REF_RETURNED" => "Referré en dehors et retouné",
+		"MEASURES" => "Mesures",
+		"VALUES" => "Valeurs",
+		"OPTIONS" => "Options",
+		"UNIT" => "Unité",
+		"RANGE_NUMERIC" => "Valeur Numérique",
+		"RANGE_ALPHANUM" => "Valeur alphanumérique",
+		"RANGE_AUTOCOMPLETE" => "Autocomplete",
+		"YES" => "Oui",
+		"NO" => "Non",
+		"ERROR" => "Erreur",
+		"INVALID" => "Invalide",
+		"NOTKNOWN" => "Non connu",
+		"REGD_BY" => "Enregistrer par:",
+		"ENTERED_BY" => "Saisie par:",
 		"VERIFIED_BY" => "Vérifié par:",
-		"ACTION" => "Action French", 		
-		"RECORDS" => "Bilans", 
-		"RECORDS_TEST" => "Bilans examens", 
-		"RECORDS_SPECIMEN" => "Bilans échantillons", 
-		"RECORDS_PATIENT" => "Bilant patients", 
-		"PATIENT_BARCODE" => "Patient Barcode French", 
+		"ACTION" => "Action French",
+		"RECORDS" => "Bilans",
+		"RECORDS_TEST" => "Bilans examens",
+		"RECORDS_SPECIMEN" => "Bilans échantillons",
+		"RECORDS_PATIENT" => "Bilant patients",
+		"PATIENT_BARCODE" => "Patient Barcode French",
 		"ADD_CURRENCY" => "Add a new currency >>",
 		"CURRENCY_DELIMITER" => "Currency Delimiter:",
 		"CURRENCY" => "Currency",
@@ -162,53 +162,53 @@ $LANG_ARRAY = array (
 		"ACTIONS" => "Actions",
 		"ADD_CURRENCY_RATE" => "Add new currency rate",
 		"NEW_CURRENCY" => "New Currency",
-		"LAB_SECTION" => "Section du Lab.", 
-		"CMD_VIEW" => "Présenter", 
-		"CMD_ADD" => "Ajouter", 
-		"PREVIEW" => "Précédent", 
+		"LAB_SECTION" => "Section du Lab.",
+		"CMD_VIEW" => "Présenter",
+		"CMD_ADD" => "Ajouter",
+		"PREVIEW" => "Précédent",
 		"ALL" => "Tout",
-		"WIDTH" => "Width French", 
-		"HEIGHT" => "Height French", 
-		"FONT_SIZE" => "Font Size French", 
-		"COL_COUNT" => "Column-Count French", 
-		"ALL_TESTS" => "Tous les Examens French", 
-		"CMD_TRACK" => "Pistes des Actions", 
-		"MSG_NOTFOUND" => "Non Retrouvé", 
-		"RELATED_TESTS_NOTFOUND" => "Related Tests Not Found", 
-		"MSG_SIMILARNOTFOUND" => "Aucune entrée similaire retrouvée", 
-		"REF_TO" => "Referer à:", 
-		"CMD_REMOVE" => "Supprimer", 
-		"CMD_VERIFY" => "Vérifier", 
+		"WIDTH" => "Width French",
+		"HEIGHT" => "Height French",
+		"FONT_SIZE" => "Font Size French",
+		"COL_COUNT" => "Column-Count French",
+		"ALL_TESTS" => "Tous les Examens French",
+		"CMD_TRACK" => "Pistes des Actions",
+		"MSG_NOTFOUND" => "Non Retrouvé",
+		"RELATED_TESTS_NOTFOUND" => "Related Tests Not Found",
+		"MSG_SIMILARNOTFOUND" => "Aucune entrée similaire retrouvée",
+		"REF_TO" => "Referer à:",
+		"CMD_REMOVE" => "Supprimer",
+		"CMD_VERIFY" => "Vérifier",
 		"CMD_REFRESH" => "Refresh French",
-		"ENTER_RESULTS" => "Entrer les Résultats", 
-		"MSG_ADDED" => "Ajouter", 
-		"FROM_DATE" => "De", 
-		"TO_DATE" => "A", 
-		"DESCRIPTION" => "Description", 
-		"PANEL_TEST" => "Liste des Examens", 
-		"MSG_MANDATORYFIELD" => "Champs Obligatoire", 
-		"SWITCH_TOTECH" => "Travailler comme Technicien", 
-		"TIPS_SELECTSITE" => "Prière de sélectionner un site de localisation.", 
-		"TIPS_INCOMPLETEINFO" => "Information incomplète. Prière de ressaisir et entrer à nouveau.", 
-		"TIPS_DATEINVALID" => "Erreur: Date invalide.", 
-		"ADDNEW" => "Ajouter un nouveau.", 
-		"ADDANOTHER" => "Ajouter un autre.", 
-		"COMPATIBLE_SPECIMENS" => "Echantillons Compatibles", 
-		"COMPATIBLE_TESTS" => "Examens Compatibles", 
-		"TIPS_ACC_CONFIRMDELETE" => "Etes vous sure de vouloir supprimer ce Compte?", 
-		"LAB_TECH" => "Techn de Lab", 
-		"LAB_RECEPTIONIST" => "Receptioniste de Lab.", 
-		"MSG_ACC_ADDED" => "Compte ajouté", 
-		"MSG_ACC_DELETED" => "Compte supprimé", 
-		"MSG_ACC_UPDATED" => "Compte actualisé", 
-		"TAT" => "Delai exécution", 
-		"TAT_AVG" => "Delai Exsécution Moyen", 
-		"TAT_TARGET" => "Cible du Felai Exécution Moyen", 
-		"TAT_ILE" => "90th percentile Delai Exécution Moyen", 
-		"PREVALENCE_RATE" => "Prévalence", 
-		"LANGUAGE" => "Language", 
-		"USER_FUNCTIONS" => "Fonctions", 
-		"APPROX" => "Approximatif", 
+		"ENTER_RESULTS" => "Entrer les Résultats",
+		"MSG_ADDED" => "Ajouter",
+		"FROM_DATE" => "De",
+		"TO_DATE" => "A",
+		"DESCRIPTION" => "Description",
+		"PANEL_TEST" => "Liste des Examens",
+		"MSG_MANDATORYFIELD" => "Champs Obligatoire",
+		"SWITCH_TOTECH" => "Travailler comme Technicien",
+		"TIPS_SELECTSITE" => "Prière de sélectionner un site de localisation.",
+		"TIPS_INCOMPLETEINFO" => "Information incomplète. Prière de ressaisir et entrer à nouveau.",
+		"TIPS_DATEINVALID" => "Erreur: Date invalide.",
+		"ADDNEW" => "Ajouter un nouveau.",
+		"ADDANOTHER" => "Ajouter un autre.",
+		"COMPATIBLE_SPECIMENS" => "Echantillons Compatibles",
+		"COMPATIBLE_TESTS" => "Examens Compatibles",
+		"TIPS_ACC_CONFIRMDELETE" => "Etes vous sure de vouloir supprimer ce Compte?",
+		"LAB_TECH" => "Techn de Lab",
+		"LAB_RECEPTIONIST" => "Receptioniste de Lab.",
+		"MSG_ACC_ADDED" => "Compte ajouté",
+		"MSG_ACC_DELETED" => "Compte supprimé",
+		"MSG_ACC_UPDATED" => "Compte actualisé",
+		"TAT" => "Delai exécution",
+		"TAT_AVG" => "Delai Exsécution Moyen",
+		"TAT_TARGET" => "Cible du Felai Exécution Moyen",
+		"TAT_ILE" => "90th percentile Delai Exécution Moyen",
+		"PREVALENCE_RATE" => "Prévalence",
+		"LANGUAGE" => "Language",
+		"USER_FUNCTIONS" => "Fonctions",
+		"APPROX" => "Approximatif",
 		"LANDSCAPE" => "Mode paysage",
          "TIPS_PATIENTID_EXISTS"=>"Patient ID existent déjà. Consulter pour le patient pour confirmer et enregistrer spécimen",
         "LAB_STAFF" => "Le personnel de laboratoire",
@@ -217,50 +217,50 @@ $LANG_ARRAY = array (
             "DISTRICT" => "District",
             "REGION" => "Région"
         ) ,
-	"header" => array ( 
-		"TITLE" => "", 
-		"LOGGEDINAS" => "Cliquer comme", 
-		"EDITPROFILE" => "Modifier votre profil", 
-		"LOGOUT" => "Quitter", 
-		"COMMENTS" => "Commentaires", 
-		"SWITCH_TOMGR" => "Travailler comme Biologiste", 
-		"MENU_BACKUP" => "Sauvegarder les Données", 
-		"SWITCH_TOTECH" => "Travailler comme Technicien", 
+	"header" => array (
+		"TITLE" => "",
+		"LOGGEDINAS" => "Cliquer comme",
+		"EDITPROFILE" => "Modifier votre profil",
+		"LOGOUT" => "Quitter",
+		"COMMENTS" => "Commentaires",
+		"SWITCH_TOMGR" => "Travailler comme Biologiste",
+		"MENU_BACKUP" => "Sauvegarder les Données",
+		"SWITCH_TOTECH" => "Travailler comme Technicien",
 		"SWITCH_TODIR" => "Travailler comme Directeur."
-	) , 
-	"footer" => array ( 
-		"TITLE" => "", 
-		"FOOTER_MSG" => "Une initiative conjointe de C4G @ Georgia Tech, la CDC et les pays participants", 
+	) ,
+	"footer" => array (
+		"TITLE" => "",
+		"FOOTER_MSG" => "Une initiative conjointe de C4G @ Georgia Tech, la CDC et les pays participants",
 		"SWITCHLANG" => "Traduire en Français"
-	) , 
-	"login" => array ( 
-		"TITLE" => "", 
-		"MSG_PLSLOGIN" => "Prière de cliquer dessus prioritairement", 
-		"MSG_TIMED_OUT" => "Temps de session suspendue à cause de inactivité", 
-		"MSG_ERR_PWD" => "Mot de passe/ Nom Utisateur incorecte", 
-		"MSG_ERR_PWDRST" => "Password cannot be reset now. Contact your admin or mail c4gbackup@gmail.com with the subject 'Password' French", 
-		"MSG_CAPSLOCK" => "Majuscule activée", 
-		"MSG_NEWPWD" => "Demande un nouveau mot de Passe", 
+	) ,
+	"login" => array (
+		"TITLE" => "",
+		"MSG_PLSLOGIN" => "Prière de cliquer dessus prioritairement",
+		"MSG_TIMED_OUT" => "Temps de session suspendue à cause de inactivité",
+		"MSG_ERR_PWD" => "Mot de passe/ Nom Utisateur incorecte",
+		"MSG_ERR_PWDRST" => "Password cannot be reset now. Contact your admin or mail c4gbackup@gmail.com with the subject 'Password' French",
+		"MSG_CAPSLOCK" => "Majuscule activée",
+		"MSG_NEWPWD" => "Demande un nouveau mot de Passe",
 		"MSG_RESETPWD" => "Reset Password French",
 		"TIPS_NEWPWD" => "Si vous avez oublié votre mot de passe, cliquer sur Demande un nouveau mot de passe et suivez le lien Nouveau, Un mot de passe temporaire vous sera communiquer à travers votre adtresse E-mail."
-	) , 
-	"home" => array ( 
-		"TITLE" => "Acceuil", 
-		"WELCOME" => "Bienvenue", 
-		"TIPS_BLISINTRO" => "Le Basic Laboratory Information System (BLIS) pistes échantillons du patient et des résultats de laboratoire.", 
+	) ,
+	"home" => array (
+		"TITLE" => "Acceuil",
+		"WELCOME" => "Bienvenue",
+		"TIPS_BLISINTRO" => "Le Basic Laboratory Information System (BLIS) pistes échantillons du patient et des résultats de laboratoire.",
 		"TIPS_PWD" => "Vous pouver mettre à jour votre profile en cliquant sur Modifier votre profile."
-	) , 
-	"regn" => array ( 
+	) ,
+	"regn" => array (
 		"TITLE" => "Enregistrement"
-	) , 
-	"find_patient" => array ( 
-		"TITLE" => "Surveillance du Patient", 
-		"ADD_NEW_PATIENT" => "Ajouter un Nouveau Patient", 
-		"TIPS_REGISTRATION_1" => "Cette page nous permet d'enregistrer de nouveaux patients ou de patients de recherche existants basés sur le nom, l'identité du patient ou le numéro.", 
-		"TIPS_REGISTRATION_2" => "Une fois qu'un patient a été enregistré, nous pouvons utiliser cette page pour visualiser, modifier le profil des patients.", 
-		"TIPS_PATIENT_LOOKUP" => "Pour enregistrer un nouvel échantillon, entrer le patient par son ID/Nom ou ajouter un nouvel échantillon.", 
-		"MSG_NOMATCH" => "Pas de sélection retrouvée", 
-		"CMD_REGISTERSPECIMEN" => "Enregistrer Echantillon", 
+	) ,
+	"find_patient" => array (
+		"TITLE" => "Surveillance du Patient",
+		"ADD_NEW_PATIENT" => "Ajouter un Nouveau Patient",
+		"TIPS_REGISTRATION_1" => "Cette page nous permet d'enregistrer de nouveaux patients ou de patients de recherche existants basés sur le nom, l'identité du patient ou le numéro.",
+		"TIPS_REGISTRATION_2" => "Une fois qu'un patient a été enregistré, nous pouvons utiliser cette page pour visualiser, modifier le profil des patients.",
+		"TIPS_PATIENT_LOOKUP" => "Pour enregistrer un nouvel échantillon, entrer le patient par son ID/Nom ou ajouter un nouvel échantillon.",
+		"MSG_NOMATCH" => "Pas de sélection retrouvée",
+		"CMD_REGISTERSPECIMEN" => "Enregistrer Echantillon",
 		"CMD_VIEWPROFILE" => "Voir le profil",
 		"CMD_DELPROFILE" => "Delete le profil",
 		"CMD_UPDPROFILE" => "Update Profile Fr"
@@ -277,189 +277,189 @@ $LANG_ARRAY = array (
 				"CMD_DELPROFILE" => "Delete le profil",
 				"CMD_UPDPROFILE" => "Update Profile Fr"
 		) ,
-	"new_patient" => array ( 
-		"TITLE" => "Nouveau Patient", 
-		"MSG_BACKTOLOOKUP" => "Retour à Idetification du Patient", 
-		"MSG_ADDNEWENTRY" => "Priére de saisir une nouvelle entrée.", 
-		"CMD_SELECTFORREGN" => "Sélectionner pour enregistrer", 
-		"SIMILAR_NAMES" => "Noms similaires", 
-		"MSG_PID_INVALID" => "ID du patient Invalide/existe déjà.", 
-		"TIPS_REGN" => "Patient ID, Name and Gender are mandatory fields.", 
-		"TIPS_REGN_NEW" => "Patient ID, nom et le sexe sont des champs obligatoires.", 
+	"new_patient" => array (
+		"TITLE" => "Nouveau Patient",
+		"MSG_BACKTOLOOKUP" => "Retour à Idetification du Patient",
+		"MSG_ADDNEWENTRY" => "Priére de saisir une nouvelle entrée.",
+		"CMD_SELECTFORREGN" => "Sélectionner pour enregistrer",
+		"SIMILAR_NAMES" => "Noms similaires",
+		"MSG_PID_INVALID" => "ID du patient Invalide/existe déjà.",
+		"TIPS_REGN" => "Patient ID, Name and Gender are mandatory fields.",
+		"TIPS_REGN_NEW" => "Patient ID, nom et le sexe sont des champs obligatoires.",
 		"TIPS_DOB_AGE" => "Seule Age ou Date de naissance est requis pour cette entrée."
-	) , 
-	"new_specimen" => array ( 
-		"TITLE" => "Enregistrement échantillon", 
-		"ADD_ANOTHER_SPECIMEN" => "Ajouter un autre échantillon", 
-		"MSG_SELECT_STYPE" => "Sélectionner le type échantillon premièrement", 
-		"MSG_SURETO_REMOVE" => "Etes vous sure de vouloir supprimer cette forme échantillon?", 
-		"MSG_ADD_PATIENT" => "Prière sélectionner ou ajouter un nouveau patient en priorité.", 
-		"MSG_STYPE_MISSING" => "Type échantillon manquant", 
-		"MSG_NOTESTS_SELECTED" => "Aucun examen sélecionné", 
-		"MSG_SID_MISSING" => "Id échantillon manquant", 
-		"MSG_SID_INVALID" => "ID échantiloin manquant /Existe déja", 
-		"MSG_RDATE_INVALID" => "Date de reception invalide", 
-		"MSG_CDATE_INVALID" => "Date de prélèvement invalide", 
-		"MSG_NOTESTMATCH" => "Aucun examen incompatible", 
-		"TIPS_SURETOABORT" => "Etes vous sue de vouloir intrrumpre cette enregistrement?", 
-		"TIPS_REGISTRATION_SPECIMEN" => "S'il vous plaît sélectionner un type d'échantillon pour voir les essais de ce spécimen.", 
+	) ,
+	"new_specimen" => array (
+		"TITLE" => "Enregistrement échantillon",
+		"ADD_ANOTHER_SPECIMEN" => "Ajouter un autre échantillon",
+		"MSG_SELECT_STYPE" => "Sélectionner le type échantillon premièrement",
+		"MSG_SURETO_REMOVE" => "Etes vous sure de vouloir supprimer cette forme échantillon?",
+		"MSG_ADD_PATIENT" => "Prière sélectionner ou ajouter un nouveau patient en priorité.",
+		"MSG_STYPE_MISSING" => "Type échantillon manquant",
+		"MSG_NOTESTS_SELECTED" => "Aucun examen sélecionné",
+		"MSG_SID_MISSING" => "Id échantillon manquant",
+		"MSG_SID_INVALID" => "ID échantiloin manquant /Existe déja",
+		"MSG_RDATE_INVALID" => "Date de reception invalide",
+		"MSG_CDATE_INVALID" => "Date de prélèvement invalide",
+		"MSG_NOTESTMATCH" => "Aucun examen incompatible",
+		"TIPS_SURETOABORT" => "Etes vous sue de vouloir intrrumpre cette enregistrement?",
+		"TIPS_REGISTRATION_SPECIMEN" => "S'il vous plaît sélectionner un type d'échantillon pour voir les essais de ce spécimen.",
 		"TIPS_REGISTRATION_SPECIMEN_1" => "Pour ajouter plus d'un spécimen cliquez sur Ajouter une autre option échantillons."
-	) , 
-	"specimen_added" => array ( 
-		"TITLE" => "Echantillons Enregistrés", 
+	) ,
+	"specimen_added" => array (
+		"TITLE" => "Echantillons Enregistrés",
 		"MSG_NEXTSPECIMEN" => "Enregistrer Echantillon suivant"
-	) , 
-	"results_entry" => array ( 
-		"TITLE" => "Résultats", 
-		"MENU_SINGLESPECIMEN" => "Résultas pour un seul Echantillon", 
+	) ,
+	"results_entry" => array (
+		"TITLE" => "Résultats",
+		"MENU_SINGLESPECIMEN" => "Résultas pour un seul Echantillon",
 		"MENU_LABSECTION" => "Lab Section-wise Results French",
-		"MENU_BATCHRESULTS" => "Résultats Groupés", 
-		"MENU_VERIFYRESULTS" => "Vérifier les Résultats", 
-		"CONTROL_TESTING_RESULTS" => "Control Results", 
+		"MENU_BATCHRESULTS" => "Résultats Groupés",
+		"MENU_VERIFYRESULTS" => "Vérifier les Résultats",
+		"CONTROL_TESTING_RESULTS" => "Control Results",
 		"RELATED_RESULTS" => "Related Test Results Entry",
 		"RELATED_RESULTS_EMPTY" => "No related tests for this specimen",
-		"MENU_REPORTRESULTS" => "Report Results", 
-		"MENU_WORKSHEET" => "Paillasse", 
-		"BLANK_WORKSHEET" => "Keep Blank", 
-		"NUM_ROWS" => "Nombres de Colonnes", 
-		"MSG_PENDINGNOTFOUND" => "Aucun Echantillon manquant retrouvé", 
-		"MSG_ALREADYENTERED" => "Resultats Présent", 
-		"MSG_RESULTSUBMITTED" => "Résultats Soumis", 
-		"MSG_RESULTMISSING" => "Valeurs manquantes", 
-		"TIPS_VERIFYCONFIRM" => "Résultats final, etes vous sure?", 
-		"TIPS_VERIFYNOTFOUND" => "Aucun résultat non vérifié retrouvé.", 
-		"TIPS_VERIFYDONE" => "Résultats marqué comme vérifié.", 
-		"TIPS_NOSPECIMENSTOTEST" => "Pas échantillon pour examen.", 
-		"TIPS_NOUNREPORTEDFOUND" => "Aucun échantiollon non reporté retrouvé.", 
-		"CMD_MARKASREPORTED" => "Marqué comme reporté.", 
-		"TIPS_MARKEDREPORTED" => "Selected specimens marked as Reported", 
-		"MSG_RESULTSUBMITTEDALL" => "Tous ls résultats soumis pour ce patient.", 
+		"MENU_REPORTRESULTS" => "Report Results",
+		"MENU_WORKSHEET" => "Paillasse",
+		"BLANK_WORKSHEET" => "Keep Blank",
+		"NUM_ROWS" => "Nombres de Colonnes",
+		"MSG_PENDINGNOTFOUND" => "Aucun Echantillon manquant retrouvé",
+		"MSG_ALREADYENTERED" => "Resultats Présent",
+		"MSG_RESULTSUBMITTED" => "Résultats Soumis",
+		"MSG_RESULTMISSING" => "Valeurs manquantes",
+		"TIPS_VERIFYCONFIRM" => "Résultats final, etes vous sure?",
+		"TIPS_VERIFYNOTFOUND" => "Aucun résultat non vérifié retrouvé.",
+		"TIPS_VERIFYDONE" => "Résultats marqué comme vérifié.",
+		"TIPS_NOSPECIMENSTOTEST" => "Pas échantillon pour examen.",
+		"TIPS_NOUNREPORTEDFOUND" => "Aucun échantiollon non reporté retrouvé.",
+		"CMD_MARKASREPORTED" => "Marqué comme reporté.",
+		"TIPS_MARKEDREPORTED" => "Selected specimens marked as Reported",
+		"MSG_RESULTSUBMITTEDALL" => "Tous ls résultats soumis pour ce patient.",
 		"CUSTOM_WORKSHEET" => "Champ commercial."
-	) , 
-	"search" => array ( 
-		"TITLE" => "Rechercher", 
-		"TIPS_SEARCH_1" => "Vous pouvez utiliser cette page pour rechercher des patients inscrits. Entrez un nom partiel de générer une liste de patients contenant la chaîne.", 
-		"TIPS_SEARCH_2" => "Une fois que vous sélectionnez le patient, vous pouvez visualiser le profil du patient, rechercher les tests qu'il a subis et les résultats de ces tests.", 
+	) ,
+	"search" => array (
+		"TITLE" => "Rechercher",
+		"TIPS_SEARCH_1" => "Vous pouvez utiliser cette page pour rechercher des patients inscrits. Entrez un nom partiel de générer une liste de patients contenant la chaîne.",
+		"TIPS_SEARCH_2" => "Une fois que vous sélectionnez le patient, vous pouvez visualiser le profil du patient, rechercher les tests qu'il a subis et les résultats de ces tests.",
 		"TIPS_SEARCH" => "Rechercher votre échantillon par son ID ou par Nom"
-	) , 
-	"patient_profile" => array ( 
-		"TITLE" => "Profil du patient", 
-		"MSG_REGNEWSPECIMEN" => "Enregistrer un nouvel échantillon", 
-		"MSG_UPDATEPROFILE" => "Actualiser le profil", 
-		"MSG_PRINTHISTORY" => "Imprimer le Bilant du Patient", 
+	) ,
+	"patient_profile" => array (
+		"TITLE" => "Profil du patient",
+		"MSG_REGNEWSPECIMEN" => "Enregistrer un nouvel échantillon",
+		"MSG_UPDATEPROFILE" => "Actualiser le profil",
+		"MSG_PRINTHISTORY" => "Imprimer le Bilant du Patient",
 		"TIPS_AGEORDOB" => "Prière de saisir Age ou date de naissance du patient uniquement."
-	) , 
-	"specimen_info" => array ( 
-		"TITLE" => "Info sur échantillon", 
-		"REGDTESTS" => "Examen(s) enregistré(s)", 
-		"TIPS_NOTESTSREGD" => "Aucun examen enregistré pour cet éhantillon.", 
-		"TIPS_VERIFYDONE" => "Résultas Echantillon vérifié et mis à jour.", 
-		"TIPS_ENTRYDONE" => "Résultats Echantillons saisies", 
-		"TITLE_VERIFY" => "Vérifier les résultas Echantillons", 
-		"TITLE_ENTRY" => "Résultats Echantillons Saisis", 
-		"TIPS_VERIFY" => "Vérifier ces résultats pour cet Echantillon et faire la correction si possible.", 
+	) ,
+	"specimen_info" => array (
+		"TITLE" => "Info sur échantillon",
+		"REGDTESTS" => "Examen(s) enregistré(s)",
+		"TIPS_NOTESTSREGD" => "Aucun examen enregistré pour cet éhantillon.",
+		"TIPS_VERIFYDONE" => "Résultas Echantillon vérifié et mis à jour.",
+		"TIPS_ENTRYDONE" => "Résultats Echantillons saisies",
+		"TITLE_VERIFY" => "Vérifier les résultas Echantillons",
+		"TITLE_ENTRY" => "Résultats Echantillons Saisis",
+		"TIPS_VERIFY" => "Vérifier ces résultats pour cet Echantillon et faire la correction si possible.",
 		"TIPS_ENTRY" => "Enter results for this specimen and make corrections to already existing results if any"
-	) , 
-	"lab_configs" => array ( 
-		"TITLE" => "Configurations du Lab.", 
-		"CMD_ADDNEWLAB" => "Ajouter un nouveau Lab.", 
-		"CMD_VIEWALL" => "Voir tout", 
-		"NEW_LAB_CONFIGURATION" => "Nouvelle configuration du Lab.", 
-		"MENU_SITEINFO" => "Site Information", 
-		"MENU_FIELDS" => "Confirmation", 
-		"TIPS_CUSTOM" => "N.B: Un champs commercial peut être ajouter aprés la configuration du Laboratoire, utiliser option de configuration de l\'administrateur.", 
-		"TIPS_CONFIRMNEWLAB" => "Prière de revoir toutes les mises à jour et cliquer sur AJOUTER pour créer une configuration du Laboratoire.", 
-		"TIPS_CREATINGLAB" => "Prière de patienter que le système crée une configuration du Lab, celà peut prendre quelques minutes.", 
-		"TIPS_MISSING_FACILITY" => "Erreur: Non du Facilitateur manquant.", 
-		"TIPS_MISSING_LOCATION" => "Erreur: Non du Facilitateur manquant.", 
-		"TIPS_MISSING_MGR" => "Erreur: Aministrateur de Laboratoire non sélectionner.", 
-		"TIPS_MISSING_STYPES" => "Erreur: Aucun échantillon sélectionné.", 
+	) ,
+	"lab_configs" => array (
+		"TITLE" => "Configurations du Lab.",
+		"CMD_ADDNEWLAB" => "Ajouter un nouveau Lab.",
+		"CMD_VIEWALL" => "Voir tout",
+		"NEW_LAB_CONFIGURATION" => "Nouvelle configuration du Lab.",
+		"MENU_SITEINFO" => "Site Information",
+		"MENU_FIELDS" => "Confirmation",
+		"TIPS_CUSTOM" => "N.B: Un champs commercial peut être ajouter aprés la configuration du Laboratoire, utiliser option de configuration de l\'administrateur.",
+		"TIPS_CONFIRMNEWLAB" => "Prière de revoir toutes les mises à jour et cliquer sur AJOUTER pour créer une configuration du Laboratoire.",
+		"TIPS_CREATINGLAB" => "Prière de patienter que le système crée une configuration du Lab, celà peut prendre quelques minutes.",
+		"TIPS_MISSING_FACILITY" => "Erreur: Non du Facilitateur manquant.",
+		"TIPS_MISSING_LOCATION" => "Erreur: Non du Facilitateur manquant.",
+		"TIPS_MISSING_MGR" => "Erreur: Aministrateur de Laboratoire non sélectionner.",
+		"TIPS_MISSING_STYPES" => "Erreur: Aucun échantillon sélectionné.",
 		"TIPS_MISSING_TTYPES" => "Erreur: Aucun Examen selectionné"
-	) , 
-	"lab_config_home" => array ( 
-		"TITLE" => "Configuration du Lab.", 
-		"MENU_SUMMARY" => "Résumer", 
-		"TIPS_SUMMARY_1" => "Donne un résumé de l'installation de laboratoire, y compris les comptes personnel de laboratoire, les essais et les types d'échantillons.", 
-		"TIPS_SUMMARY_2" => "-", 
-		"TIPS_SUMMARY_3" => "-", 
-		"MENU_BACKUP_REVERT" => "Retour à la sauvegarde", 
+	) ,
+	"lab_config_home" => array (
+		"TITLE" => "Configuration du Lab.",
+		"MENU_SUMMARY" => "Résumer",
+		"TIPS_SUMMARY_1" => "Donne un résumé de l'installation de laboratoire, y compris les comptes personnel de laboratoire, les essais et les types d'échantillons.",
+		"TIPS_SUMMARY_2" => "-",
+		"TIPS_SUMMARY_3" => "-",
+		"MENU_BACKUP_REVERT" => "Retour à la sauvegarde",
 		"MENU_KEY_MANAGEMENT" => "gérer les clés de sauvegarde",
-		"TIPS_REVERT" => "En cas de défaillance du système, si nous voulons revenir à une copie précédemment sauvegardé des données, nous pouvons utiliser cette option. Vous pouvez choisir la version actuelle de sauvegarde avant de faire cela. Vous pouvez également choisir d'inclure de revenir les paramètres de langue.", 
-		"BACKUP_LOCATION" => "Version de sauvegarde", 
-		"INCLUDE_LANGUAGE_SETTINGS" => "Inclure la mise à jour de la langue", 
-		"MENU_ST_TYPES" => "Echantillon/Types Examens", 
-		"TIPS_SPECIMENTESTTYPES" => "Spécimen / Types de test - Cette option nous permet de configurer les échantillons peuvent être collectés et des essais qui sont effectués au laboratoire particulier.", 
-		"TIPS_TARGETTAT" => "Cible TAT - Cette liste d'options le délai d'exécution pour chaque type de test. Nous pouvons modifier le TAT en utilisant l'option Modifier et modifier la valeur.", 
-		"TIPS_RESULTINTERPRETATION" => "Interprétation des résultats - Cette option nous permet de préciser l'interprétation de plusieurs plages de valeurs pour chaque type de test.", 
-		"MENU_USERS" => "Compte Utilisateur", 
+		"TIPS_REVERT" => "En cas de défaillance du système, si nous voulons revenir à une copie précédemment sauvegardé des données, nous pouvons utiliser cette option. Vous pouvez choisir la version actuelle de sauvegarde avant de faire cela. Vous pouvez également choisir d'inclure de revenir les paramètres de langue.",
+		"BACKUP_LOCATION" => "Version de sauvegarde",
+		"INCLUDE_LANGUAGE_SETTINGS" => "Inclure la mise à jour de la langue",
+		"MENU_ST_TYPES" => "Echantillon/Types Examens",
+		"TIPS_SPECIMENTESTTYPES" => "Spécimen / Types de test - Cette option nous permet de configurer les échantillons peuvent être collectés et des essais qui sont effectués au laboratoire particulier.",
+		"TIPS_TARGETTAT" => "Cible TAT - Cette liste d'options le délai d'exécution pour chaque type de test. Nous pouvons modifier le TAT en utilisant l'option Modifier et modifier la valeur.",
+		"TIPS_RESULTINTERPRETATION" => "Interprétation des résultats - Cette option nous permet de préciser l'interprétation de plusieurs plages de valeurs pour chaque type de test.",
+		"MENU_USERS" => "Compte Utilisateur",
         "ADD_USER_TYPE" => "Ajouter Nouveau Utilisateur Type",
-		"TIPS_USERACCOUNTS_1" => "Cette page donne un résumé des différents utilisateurs ont accès au système.", 
-		"TIPS_USERACCOUNTS_2" => "Il permet aussi la création de nouveaux comptes utilisateurs.", 
-		"TIPS_USERACCOUNTS_3" => "Il permet de modifier les paramètres du compte, supprimer des comptes et de surveiller l'activité du compte.", 
-		"MENU_CUSTOM" => "Champs Enregistrement", 
-		"TIPS_REGISTRATIONFIELDS_1" => "Cette page permet la configuration de la page d'inscription des patients. Il permet de créer des champs obligatoires et cache certains domaines si elles ne sont pas souhaités à pourvoir.", 
-		"TIPS_REGISTRATIONFIELDS_2" => "Il permet aussi la création de certains des champs personnalisés pour l'enregistrement des patients et de nouveaux addition de l'échantillon qui peut être requis par certains laboratoires seulement.", 
-		"TIPS_REGISTRATIONFIELDS_3" => "-", 
-		"MENU_TAT" => "Target TAT", 
-		"MENU_INFECTION" => "Bilan Infection", 
-		"TIPS_INFECTIONREPORT" => "Ceci est une page modifiable pour le réglage des différents paramètres associaated chaque type d'épreuve individuelle.", 
-		"MENU_GENERAL" => "General settings", 
-		"Reports" => "Rapports", 
+		"TIPS_USERACCOUNTS_1" => "Cette page donne un résumé des différents utilisateurs ont accès au système.",
+		"TIPS_USERACCOUNTS_2" => "Il permet aussi la création de nouveaux comptes utilisateurs.",
+		"TIPS_USERACCOUNTS_3" => "Il permet de modifier les paramètres du compte, supprimer des comptes et de surveiller l'activité du compte.",
+		"MENU_CUSTOM" => "Champs Enregistrement",
+		"TIPS_REGISTRATIONFIELDS_1" => "Cette page permet la configuration de la page d'inscription des patients. Il permet de créer des champs obligatoires et cache certains domaines si elles ne sont pas souhaités à pourvoir.",
+		"TIPS_REGISTRATIONFIELDS_2" => "Il permet aussi la création de certains des champs personnalisés pour l'enregistrement des patients et de nouveaux addition de l'échantillon qui peut être requis par certains laboratoires seulement.",
+		"TIPS_REGISTRATIONFIELDS_3" => "-",
+		"MENU_TAT" => "Target TAT",
+		"MENU_INFECTION" => "Bilan Infection",
+		"TIPS_INFECTIONREPORT" => "Ceci est une page modifiable pour le réglage des différents paramètres associaated chaque type d'épreuve individuelle.",
+		"MENU_GENERAL" => "General settings",
+		"Reports" => "Rapports",
 		"Results" => "Results",
-		"Tests" => "Tests", 
-		"New Stock" => "New Stock", 
-		"Edit Stock" => "Modifier Stock", 
-		"Inventory" => "Inventaire", 
-		"MENU_REPORTCONFIG" => "Daily Report Settings", 
-		"TIPS_DAILYREPORTSETTINGS" => "Ceci est une page modifiable pour la mise la mise en page du journal de patient, journal de log des échantillons et quotidiennes des patients.", 
-		"MENU_WORKSHEETCONFIG" => "Worksheet Settings", 
-		"TIPS_WORKSHEETS" => "Ceci est une page modifiable pour créer des feuilles de calcul personnalisées.", 
-		"MENU_MGR" => "Change Manager", 
-		"MENU_DEL" => "Effacer la Configuration", 
-		"MENU_EXPORTCONFIG" => "Exporter la Configuration.", 
-		"RECENT_ACTIVITY" => "Activité Récente.", 
-		"SP_REGNS" => "Enregistrement des échantillons", 
-		"RE_ENTRIES" => "Result Entries", 
-		"TIPS_RECENTACT" => "Aucune activité recente retrouvée.", 
-		"NEW_LAB_USER" => "Nouvel Utilisateur du Lab.", 
-		"EDIT_LAB_USER" => "Nouvel Utilisateur du Lab.", 
-		"TIPS_CUSTOMFIELD_NOTFOUND" => "Pas de champs commercial existant.", 
-		"CUSTOMFIELDS" => "Champs Commerciaux.", 
-		"NEW_CUSTOMFIELD" => "Nouveau Champ Commercial.", 
-		"EDIT_CUSTOMFIELD" => "Modifier le Champ commercial.", 
-		"ASSIGN_TO" => "Tranmettre à:", 
-		"TIPS_LABDELETE" => "Etes vous sure de vouloir effacer cette configuration du Lab? Toutes les données et les comptes des techniciens seront supprimés définitivement.", 
-		"TIPS_SPECIMENSNOTSELECTED" => "Erreur: Aucun Echantillon sélectionné.", 
-		"TIPS_TESTSNOTSELECTED" => "Erreur:Aucun Examen sélectionner.", 
-		"TIPS_USERACC" => "Modifier le compte le détail utilisateur ou restaurer le mot de passe en saisissant un nouveau.", 
-		"GROUP_BYGENDER" => "Grouper par Sexe.", 
-		"GROUP_BYAGE" => "Grouper par Age", 
-		"RANGE_AGE" => "Valeur Age", 
-		"RANGE_NOTREQD" => "Pas de valeur configurée nécessaire", 
-		"MSG_PENDINGTAT" => "Default TAT for pending tests", 
-		"MODIFYLANG" => "Modifier la Langue", 
-		"MSG_LANGUPDATED" => "Language strings updated.", 
-		"OTHER_FIELDS" => "Autres Champs", 
-		"MARGINS" => "Marges", 
-		"COLUMN_WIDTH" => "Column Width", 
-		"TOP" => "Entète", 
-		"BOTTOM" => "Bas de page", 
-		"LEFT" => "Gauche", 
-		"RIGHT" => "Droit", 
-		"DAILY" => "Journalier", 
-		"WEEKLY" => "Hebdomadaire", 
-		"MONTHLY" => "Mensuel", 
-		"YEARLY" => "Annuel", 
-		"USE_PNAME_RESULTS" => "Display Name at Results Entry", 
-		"CUSTOM_WORKSHEETS" => "Custom Worksheets", 
-		"NEW_CUSTOMWORKSHEET" => "Add Custom Worksheet", 
-		"ADDED_CUSTOMWORKSHEET" => "Custom Worksheet Added", 
-		"EDIT_CUSTOMWORKSHEET" => "Edit Custom Worksheet", 
-		"UPDATED_CUSTOMWORKSHEET" => "Custom Worksheet Updated", 
-		"BACK_TOCONFIG" => "Retourner à la configuration", 
-		"BACKUP_CURRENT_VERSION" => "Sauvegarder la version courante avant de retourner.", 
-		"TIPS_REVERTDONE" => "La Sauvegarde a réussite. Prière de redemarer pour achener le procedé inverse.", 
-		"TIPS_SETUPNETWORK_3" => "Vous allez maintenant être en mesure d'accéder en cliquant sur ??BLIS BlisSetup.html et en entrant votre nom d'utilisateur et mot de passe.", 
+		"Tests" => "Tests",
+		"New Stock" => "New Stock",
+		"Edit Stock" => "Modifier Stock",
+		"Inventory" => "Inventaire",
+		"MENU_REPORTCONFIG" => "Daily Report Settings",
+		"TIPS_DAILYREPORTSETTINGS" => "Ceci est une page modifiable pour la mise la mise en page du journal de patient, journal de log des échantillons et quotidiennes des patients.",
+		"MENU_WORKSHEETCONFIG" => "Worksheet Settings",
+		"TIPS_WORKSHEETS" => "Ceci est une page modifiable pour créer des feuilles de calcul personnalisées.",
+		"MENU_MGR" => "Change Manager",
+		"MENU_DEL" => "Effacer la Configuration",
+		"MENU_EXPORTCONFIG" => "Exporter la Configuration.",
+		"RECENT_ACTIVITY" => "Activité Récente.",
+		"SP_REGNS" => "Enregistrement des échantillons",
+		"RE_ENTRIES" => "Result Entries",
+		"TIPS_RECENTACT" => "Aucune activité recente retrouvée.",
+		"NEW_LAB_USER" => "Nouvel Utilisateur du Lab.",
+		"EDIT_LAB_USER" => "Nouvel Utilisateur du Lab.",
+		"TIPS_CUSTOMFIELD_NOTFOUND" => "Pas de champs commercial existant.",
+		"CUSTOMFIELDS" => "Champs Commerciaux.",
+		"NEW_CUSTOMFIELD" => "Nouveau Champ Commercial.",
+		"EDIT_CUSTOMFIELD" => "Modifier le Champ commercial.",
+		"ASSIGN_TO" => "Tranmettre à:",
+		"TIPS_LABDELETE" => "Etes vous sure de vouloir effacer cette configuration du Lab? Toutes les données et les comptes des techniciens seront supprimés définitivement.",
+		"TIPS_SPECIMENSNOTSELECTED" => "Erreur: Aucun Echantillon sélectionné.",
+		"TIPS_TESTSNOTSELECTED" => "Erreur:Aucun Examen sélectionner.",
+		"TIPS_USERACC" => "Modifier le compte le détail utilisateur ou restaurer le mot de passe en saisissant un nouveau.",
+		"GROUP_BYGENDER" => "Grouper par Sexe.",
+		"GROUP_BYAGE" => "Grouper par Age",
+		"RANGE_AGE" => "Valeur Age",
+		"RANGE_NOTREQD" => "Pas de valeur configurée nécessaire",
+		"MSG_PENDINGTAT" => "Default TAT for pending tests",
+		"MODIFYLANG" => "Modifier la Langue",
+		"MSG_LANGUPDATED" => "Language strings updated.",
+		"OTHER_FIELDS" => "Autres Champs",
+		"MARGINS" => "Marges",
+		"COLUMN_WIDTH" => "Column Width",
+		"TOP" => "Entète",
+		"BOTTOM" => "Bas de page",
+		"LEFT" => "Gauche",
+		"RIGHT" => "Droit",
+		"DAILY" => "Journalier",
+		"WEEKLY" => "Hebdomadaire",
+		"MONTHLY" => "Mensuel",
+		"YEARLY" => "Annuel",
+		"USE_PNAME_RESULTS" => "Display Name at Results Entry",
+		"CUSTOM_WORKSHEETS" => "Custom Worksheets",
+		"NEW_CUSTOMWORKSHEET" => "Add Custom Worksheet",
+		"ADDED_CUSTOMWORKSHEET" => "Custom Worksheet Added",
+		"EDIT_CUSTOMWORKSHEET" => "Edit Custom Worksheet",
+		"UPDATED_CUSTOMWORKSHEET" => "Custom Worksheet Updated",
+		"BACK_TOCONFIG" => "Retourner à la configuration",
+		"BACKUP_CURRENT_VERSION" => "Sauvegarder la version courante avant de retourner.",
+		"TIPS_REVERTDONE" => "La Sauvegarde a réussite. Prière de redemarer pour achener le procedé inverse.",
+		"TIPS_SETUPNETWORK_3" => "Vous allez maintenant être en mesure d'accéder en cliquant sur ??BLIS BlisSetup.html et en entrant votre nom d'utilisateur et mot de passe.",
 		"TIPS_SETUPNETWORK_4" => "Dans le cas d'un redémarrage de l'ordinateur ou cliquez sur une défaillance du réseau d'installation de réseau option à nouveau.",
 		"RPT_ITEMS_ON_ROW"=>"Articles sur une ligne",
 		"RPT_SHOW_DEMO_BORDER"=>"Afficher Démographie frontière",
@@ -489,112 +489,112 @@ $LANG_ARRAY = array (
         "PRINT_UNVERIFIED" => "Autoriser l'impression de tests non vérifiés"
 
     ) ,
-	"lab_admins" => array ( 
-		"TITLE" => "Biologistes", 
-		"LABS_ASSIGNED" => "Labs Assigned", 
-		"TIPS_MANAGELAB" => "Pour modifier le compte des techniciens, utiliser Option administrateur dans la page de Configuration du Lab.", 
-		"TIPS_ACC_LABSEXIST" => "Lab Configurations exist under this admin account. Please delete or re-assign them before deleting this account.", 
-		"ADD_NEW_ADMIN" => "Ajouter un Nouvel Administrateur.", 
+	"lab_admins" => array (
+		"TITLE" => "Biologistes",
+		"LABS_ASSIGNED" => "Labs Assigned",
+		"TIPS_MANAGELAB" => "Pour modifier le compte des techniciens, utiliser Option administrateur dans la page de Configuration du Lab.",
+		"TIPS_ACC_LABSEXIST" => "Lab Configurations exist under this admin account. Please delete or re-assign them before deleting this account.",
+		"ADD_NEW_ADMIN" => "Ajouter un Nouvel Administrateur.",
 		"EDIT_ADMIN_ACC" => "Modifier le Chef de Lab."
-	) , 
-	"catalog" => array ( 
-		"TITLE" => "Catalogue des Examens", 
-		"MENU_REMOVEDATA" => "Effacer les données du catalogue", 
-		"TIPS_CATALOG" => "Vous pouvez ajouter, supprimer ou modifier les types Echantillons, lles types Examens etintervale de validité dans cette page.", 
-		"TIPS_TESTSNOTFOUND" => "Pas de type Examen correspondant dans le catalogue.", 
-		"TIPS_SPECIMENSNOTFOUND" => "Pas de type Echantillon dans le Catalogue.", 
-		"TIPS_REMOVEDATA" => "Sure to remove catalog data? All specimen/test types will be deleted and would have to be entered again.", 
-		"NEW_TEST_TYPE" => "Nouveau Type Examen", 
-		"NEW_SPECIMEN_TYPE" => "Nouveau Type Echantillon", 
-		"EDIT_TEST_TYPE" => "Modifier le type Examen", 
-		"EDIT_SPECIMEN_TYPE" => "Modifier le type Echantillon", 
-		"NEW_LAB_SECTION" => "Nouvelle Section du Lab.", 
-		"TIPS_MISSING_TESTNAME" => "Erreur: Nom Examen manquant.", 
-		"TIPS_MISSING_SPECIMENNAME" => "Erreur: Nom Echantillon manquant.", 
-		"TIPS_MISSING_CATNAME" => "Erreur: Section du Lab. manquant.", 
-		"TIPS_MISSING_SELECTEDSPECIMEN" => "Erreur: Aucun Echantillon compatible sélectionné.", 
-		"TIPS_MISSING_SELECTEDTESTS" => "Erreur:Aucun examen compatible sélectinné.", 
-		"TIPS_MISSING_SELECTEDMEASURES" => "Erreur: Aucune Mésure sélectionner pour cet Examen.", 
-		"TEST_TYPE_ADDED" => "Examen ajouté", 
-		"SPECIMEN_TYPE_ADDED" => "Echantillon Ajouté", 
-		"TEST_TYPE_UPDATED" => "Examen mis-à-jour", 
-		"SPECIMEN_TYPE_UPDATED" => "Echantillon mis-à-jour.", 
-		"CMD_BACK_TOCATALOG" => "Retour au Catalogue.", 
-		"TIPS_MEASURES" => "Seuls les numéros et alphabets doivent être utilisés tout en ajoutant l'option measures.For gamme s'il vous plaît spécifier la limite inférieure et la borne supérieure.", 
-		"TIPS_TC_TESTTYPE_1" => "Donne une liste complète des types de tests disponibles.", 
-		"TIPS_TC_TESTTYPE_2" => "Utilisez cette page pour modifier les types de tests existants.", 
-		"TIPS_TC_TESTTYPE_3" => "Vous pouvez également ajouter de nouveaux types d'essai utilisant l'option 'Ajouter' situé près du haut de la page.", 
-		"TIPS_TC_SPECIMENTYPE_1" => "Donne une liste complète des types d'échantillons disponibles.", 
-		"TIPS_TC_SPECIMENTYPE_2" => "Utilisez cette page pour modifier les types d'échantillons existants. Tests compatible avec un type de spécimen particulier peut ajouter / supprimer en cochant les cases appropriées trouve ici.", 
+	) ,
+	"catalog" => array (
+		"TITLE" => "Catalogue des Examens",
+		"MENU_REMOVEDATA" => "Effacer les données du catalogue",
+		"TIPS_CATALOG" => "Vous pouvez ajouter, supprimer ou modifier les types Echantillons, lles types Examens etintervale de validité dans cette page.",
+		"TIPS_TESTSNOTFOUND" => "Pas de type Examen correspondant dans le catalogue.",
+		"TIPS_SPECIMENSNOTFOUND" => "Pas de type Echantillon dans le Catalogue.",
+		"TIPS_REMOVEDATA" => "Sure to remove catalog data? All specimen/test types will be deleted and would have to be entered again.",
+		"NEW_TEST_TYPE" => "Nouveau Type Examen",
+		"NEW_SPECIMEN_TYPE" => "Nouveau Type Echantillon",
+		"EDIT_TEST_TYPE" => "Modifier le type Examen",
+		"EDIT_SPECIMEN_TYPE" => "Modifier le type Echantillon",
+		"NEW_LAB_SECTION" => "Nouvelle Section du Lab.",
+		"TIPS_MISSING_TESTNAME" => "Erreur: Nom Examen manquant.",
+		"TIPS_MISSING_SPECIMENNAME" => "Erreur: Nom Echantillon manquant.",
+		"TIPS_MISSING_CATNAME" => "Erreur: Section du Lab. manquant.",
+		"TIPS_MISSING_SELECTEDSPECIMEN" => "Erreur: Aucun Echantillon compatible sélectionné.",
+		"TIPS_MISSING_SELECTEDTESTS" => "Erreur:Aucun examen compatible sélectinné.",
+		"TIPS_MISSING_SELECTEDMEASURES" => "Erreur: Aucune Mésure sélectionner pour cet Examen.",
+		"TEST_TYPE_ADDED" => "Examen ajouté",
+		"SPECIMEN_TYPE_ADDED" => "Echantillon Ajouté",
+		"TEST_TYPE_UPDATED" => "Examen mis-à-jour",
+		"SPECIMEN_TYPE_UPDATED" => "Echantillon mis-à-jour.",
+		"CMD_BACK_TOCATALOG" => "Retour au Catalogue.",
+		"TIPS_MEASURES" => "Seuls les numéros et alphabets doivent être utilisés tout en ajoutant l'option measures.For gamme s'il vous plaît spécifier la limite inférieure et la borne supérieure.",
+		"TIPS_TC_TESTTYPE_1" => "Donne une liste complète des types de tests disponibles.",
+		"TIPS_TC_TESTTYPE_2" => "Utilisez cette page pour modifier les types de tests existants.",
+		"TIPS_TC_TESTTYPE_3" => "Vous pouvez également ajouter de nouveaux types d'essai utilisant l'option 'Ajouter' situé près du haut de la page.",
+		"TIPS_TC_SPECIMENTYPE_1" => "Donne une liste complète des types d'échantillons disponibles.",
+		"TIPS_TC_SPECIMENTYPE_2" => "Utilisez cette page pour modifier les types d'échantillons existants. Tests compatible avec un type de spécimen particulier peut ajouter / supprimer en cochant les cases appropriées trouve ici.",
 		"TIPS_TC_SPECIMENTYPE_3" => "Vous pouvez également ajouter de nouveaux types des échantillons en utilisant l'option 'Ajouter' situé près du haut de la page."
-	) , 
-	"reports" => array ( 
+	) ,
+	"reports" => array (
 		"AGE_UNIT" => "Unite de Age",
 		"GROUP_BYGENDER" => "Groupe De Genre",
 		"GROUP_BYAGE" => "Groupe De Age",
-		"TITLE" => "Bilans", 
-		"MENU_DAILY" => "Bilans journalier", 
-		"MENU_PATIENT" => "Bilan du Patient", 
-		"MENU_SPECIMEN" => "Echantillon individuel", 
-		"MENU_PHISTORY" => "Historique du patient", 
-		"MENU_TESTRECORDS" => "Bilan des examens", 
-		"MENU_PENDINGTESTS" => "Examents en attentes", 
-		"MENU_DAILYLOGS" => "Registre Journalier", 
-		"MENU_AGGREPORTS" => "Rapport Groupés", 
+		"TITLE" => "Bilans",
+		"MENU_DAILY" => "Bilans journalier",
+		"MENU_PATIENT" => "Bilan du Patient",
+		"MENU_SPECIMEN" => "Echantillon individuel",
+		"MENU_PHISTORY" => "Historique du patient",
+		"MENU_TESTRECORDS" => "Bilan des examens",
+		"MENU_PENDINGTESTS" => "Examents en attentes",
+		"MENU_DAILYLOGS" => "Registre Journalier",
+		"MENU_AGGREPORTS" => "Rapport Groupés",
 		"MENU_TEST_SPECIMEN_AGGREGATE_REPORT_CONFIG" => "Examine/Echantillon Aggrege Configuration",
 		"MENU_SPECIMEN_AGGREGATE_REPORT_FORM" => "Echantillon Aggrege Rapport",
 		"MENU_TEST_AGGREGATE_REPORT_FORM" => "Examnie Aggrege Rapport",
-		"MENU_INFECTIONSUMMARY" => "Prévalence", 
-		"MENU_COUNTS" => "Dénombrement", 
-		"MENU_CONTROLREPORT" => "Control Report", 
-		"MENU_TAT" => "Delais de Rendu des Résultats", 
-		"MENU_INFECTIONREPORT" => "Compte Rendu des Infections", 
-		"COUNT_TYPE" => "Count Type", 
-		"COUNT_TEST" => "Comptabilité des Examens", 
-		"COUNT_SPECIMEN" => "Comptabilité Echantillon", 
-		"TIPS_PHISTORY" => "Sélectionner le site et ID du Patient pour imprimer Historique des examens.", 
-		"TIPS_SPECIMEN" => "Choisir Echantillon pour le bilan de tous les examens fait par le patient.", 
-		"TIPS_TESTRECORDS" => "Selectionner le site, le type examen, la date intervale, et le champ pour imprimer le bilant des examens.", 
-		"TIPS_PENDINGTESTS" => "Choisir la palllaisse et le type d\\\'examen pour voir la liste des examens en attente sur cette paillasse.", 
-		"TIPS_DAILYLOGS" => "Imprimer tous les rapports manuels pour la journée indiquée.", 
-		"TIPS_INFECTIONSUMMARY" => "Sélectionner le facilitateur et Intervale de la Date pour voir le graphe des infections et la prévalence.", 
-		"TIPS_COUNTS" => "Sélectionner le site et Intervale de la Date pour voir le nombre des échantillons ou des examens réalisés pendant un temps spécifié.", 
-		"TIPS_TAT" => "Sélectionner le site et la période pour voir le delai moyen de rendu de résultat pendant la période spécifiée.", 
-		"TIPS_INFECTIONREPORT" => "Sélectionner le type d'examen pour genérer la prévalence en fonction de qge et du sexe.", 
-		"TIPS_GETACCESSIONREPORT" => "Obtenir le bilan entier.", 
-		"RANGE_NORMAL" => "Valeur Normale", 
-		"REPORT_SAMPLE" => "Complete Sample Report", 
-		"TIPS_RECNOTFOUND" => "Pas de bilan retrouvé.", 
-		"TOTAL" => "Total", 
-		"TOTAL_TESTS" => "Examens Totaux", 
-		"TOTAL_PATIENTS" => "Patients Totaux", 
-		"TOTAL_SPECIMENS" => "Echantillons Totaux", 
-		"TESTED_N" => "Examen Négatifs", 
-		"TESTED_P" => "Examen Positif", 
-		"TOTAL_MF" => "M/F Total", 
-		"RANGE_AGE" => "Valeur Age", 
-		"MSG_BACKTOREPORTS" => "Retourner au Rapport", 
-		"TIPS_NONEWTESTS" => "Aucun nouveeau examen enregistré aujourdhui", 
-		"TIPS_NONEWPATIENTS" => "Aucun nouveau patient enregistré aujourdhui.", 
-		"MSG_SHOWTABLE" => "Présenter sous forme de tableau.", 
-		"MSG_HIDETABLE" => "Masquer le tableau.", 
-		"MSG_SHOWGRAPH" => "Présenter sur forme de Graphe.", 
-		"MSG_HIDEGRAPH" => "Masquer le Graphe.", 
-		"TIPS_NODISCRETE" => "Erreur: Aucune valeur numérique pour Examen Spécifié.", 
-		"PROGRESSION_M" => "Mensuel", 
-		"PROGRESSION_W" => "Hebdomadaire", 
-		"PROGRESSION_D" => "Journalier", 
-		"REPORT_PERIOD" => "Période de Enregistrement", 
-		"TIPS_NOTATTESTS" => "Aucun examen correspondant pour la période indiquée.", 
-		"TIPS_TATEXCEED" => "Aucun echantillon a exédé le delai de rendu des resultats.", 
-		"TIPS_TATEXCEEDNUM" => "Echantillons eyant exédés le delai de rendu des resultats.", 
-		"TIPS_TATNOPENDING" => "No specimens pending in the supplied date interval.", 
-		"TIPS_TATPENDING" => "Specimens pending in the supplied date interval.", 
-		"TIPS_CONFIGINFECTION" => "Erreur: Le bilan des infections non configuré pour cette facilité. Prière de contacter le Chef de Lab.", 
-		"TIPS_NOMATCHSPECIMENS" => "Aucun échantillon retrouvé.", 
-		"TIPS_MULTPATIENTS" => "Sélectionner les échantillons pour différents patients. Prière de recommencer.", 
-		"MSG_INCLUDEPENDING" => "Inclure les Examens manquants.", 
-		"TIPS_TABLESORT" => "Cliquer sur la colonne du haut de page pour voir par champ.", 
+		"MENU_INFECTIONSUMMARY" => "Prévalence",
+		"MENU_COUNTS" => "Dénombrement",
+		"MENU_CONTROLREPORT" => "Control Report",
+		"MENU_TAT" => "Delais de Rendu des Résultats",
+		"MENU_INFECTIONREPORT" => "Compte Rendu des Infections",
+		"COUNT_TYPE" => "Count Type",
+		"COUNT_TEST" => "Comptabilité des Examens",
+		"COUNT_SPECIMEN" => "Comptabilité Echantillon",
+		"TIPS_PHISTORY" => "Sélectionner le site et ID du Patient pour imprimer Historique des examens.",
+		"TIPS_SPECIMEN" => "Choisir Echantillon pour le bilan de tous les examens fait par le patient.",
+		"TIPS_TESTRECORDS" => "Selectionner le site, le type examen, la date intervale, et le champ pour imprimer le bilant des examens.",
+		"TIPS_PENDINGTESTS" => "Choisir la palllaisse et le type d\\\'examen pour voir la liste des examens en attente sur cette paillasse.",
+		"TIPS_DAILYLOGS" => "Imprimer tous les rapports manuels pour la journée indiquée.",
+		"TIPS_INFECTIONSUMMARY" => "Sélectionner le facilitateur et Intervale de la Date pour voir le graphe des infections et la prévalence.",
+		"TIPS_COUNTS" => "Sélectionner le site et Intervale de la Date pour voir le nombre des échantillons ou des examens réalisés pendant un temps spécifié.",
+		"TIPS_TAT" => "Sélectionner le site et la période pour voir le delai moyen de rendu de résultat pendant la période spécifiée.",
+		"TIPS_INFECTIONREPORT" => "Sélectionner le type d'examen pour genérer la prévalence en fonction de qge et du sexe.",
+		"TIPS_GETACCESSIONREPORT" => "Obtenir le bilan entier.",
+		"RANGE_NORMAL" => "Valeur Normale",
+		"REPORT_SAMPLE" => "Complete Sample Report",
+		"TIPS_RECNOTFOUND" => "Pas de bilan retrouvé.",
+		"TOTAL" => "Total",
+		"TOTAL_TESTS" => "Examens Totaux",
+		"TOTAL_PATIENTS" => "Patients Totaux",
+		"TOTAL_SPECIMENS" => "Echantillons Totaux",
+		"TESTED_N" => "Examen Négatifs",
+		"TESTED_P" => "Examen Positif",
+		"TOTAL_MF" => "M/F Total",
+		"RANGE_AGE" => "Valeur Age",
+		"MSG_BACKTOREPORTS" => "Retourner au Rapport",
+		"TIPS_NONEWTESTS" => "Aucun nouveeau examen enregistré aujourdhui",
+		"TIPS_NONEWPATIENTS" => "Aucun nouveau patient enregistré aujourdhui.",
+		"MSG_SHOWTABLE" => "Présenter sous forme de tableau.",
+		"MSG_HIDETABLE" => "Masquer le tableau.",
+		"MSG_SHOWGRAPH" => "Présenter sur forme de Graphe.",
+		"MSG_HIDEGRAPH" => "Masquer le Graphe.",
+		"TIPS_NODISCRETE" => "Erreur: Aucune valeur numérique pour Examen Spécifié.",
+		"PROGRESSION_M" => "Mensuel",
+		"PROGRESSION_W" => "Hebdomadaire",
+		"PROGRESSION_D" => "Journalier",
+		"REPORT_PERIOD" => "Période de Enregistrement",
+		"TIPS_NOTATTESTS" => "Aucun examen correspondant pour la période indiquée.",
+		"TIPS_TATEXCEED" => "Aucun echantillon a exédé le delai de rendu des resultats.",
+		"TIPS_TATEXCEEDNUM" => "Echantillons eyant exédés le delai de rendu des resultats.",
+		"TIPS_TATNOPENDING" => "No specimens pending in the supplied date interval.",
+		"TIPS_TATPENDING" => "Specimens pending in the supplied date interval.",
+		"TIPS_CONFIGINFECTION" => "Erreur: Le bilan des infections non configuré pour cette facilité. Prière de contacter le Chef de Lab.",
+		"TIPS_NOMATCHSPECIMENS" => "Aucun échantillon retrouvé.",
+		"TIPS_MULTPATIENTS" => "Sélectionner les échantillons pour différents patients. Prière de recommencer.",
+		"MSG_INCLUDEPENDING" => "Inclure les Examens manquants.",
+		"TIPS_TABLESORT" => "Cliquer sur la colonne du haut de page pour voir par champ.",
 		"MSG_ALLTESTSCOMPLETED" => "Tous les examens sont complets.",
         "MENU_TESTWISE_REPORTS" => "Rapports spécifiques au test",
         "SITE" => "Site",
@@ -606,147 +606,147 @@ $LANG_ARRAY = array (
         "SAMPLES_RECEIVED" => "Echantillon reçu",
          "OUT_OF_RANGE" => "hors de portée",
         "INTERPRETATION" => "Interprétation"
-	) , 
-	"edit_profile" => array ( 
-		"LINK_CHANGEPWD" => "Changer votre Mot de passe.", 
-		"TIPS_CASEPWD" => "NB: Le mot de passe est très sensible.", 
-		"MSG_NOFIELDUPDATE" => "Pas de champs à actualiser.", 
-		"CURRPWD" => "Mot de Passe courant.", 
-		"NEWPWD" => "Noveau Mot de Passe.", 
-		"REENTERPWD" => "Ressaisir votre Mot de Passe.", 
-		"MSG_REQDFIELD" => "Champs Indispensable.", 
-		"MSG_PWDNOMATCH" => "Mot de passe non spécifié.", 
-		"MSG_PWDMINLENGTH" => "Mot passe avec Min 3 Lettres/et ou chiffres.", 
-		"MSG_PROFILEUPDATED" => "Profil mis à jour.", 
-		"MSG_PWDUPDATEERROR" => "Erreur pendant la mise à jour du Mot de Passe.", 
-		"MSG_PWDMATCHERROR" => "Erreur: Ancien Mot de Passe non spécifié.", 
+	) ,
+	"edit_profile" => array (
+		"LINK_CHANGEPWD" => "Changer votre Mot de passe.",
+		"TIPS_CASEPWD" => "NB: Le mot de passe est très sensible.",
+		"MSG_NOFIELDUPDATE" => "Pas de champs à actualiser.",
+		"CURRPWD" => "Mot de Passe courant.",
+		"NEWPWD" => "Noveau Mot de Passe.",
+		"REENTERPWD" => "Ressaisir votre Mot de Passe.",
+		"MSG_REQDFIELD" => "Champs Indispensable.",
+		"MSG_PWDNOMATCH" => "Mot de passe non spécifié.",
+		"MSG_PWDMINLENGTH" => "Mot passe avec Min 3 Lettres/et ou chiffres.",
+		"MSG_PROFILEUPDATED" => "Profil mis à jour.",
+		"MSG_PWDUPDATEERROR" => "Erreur pendant la mise à jour du Mot de Passe.",
+		"MSG_PWDMATCHERROR" => "Erreur: Ancien Mot de Passe non spécifié.",
 		"MSG_PWDUPDATED" => "Mot de Passe mis à jour."
-	) , 
-	"help" => array ( 
+	) ,
+	"help" => array (
 		"TITLE" => "Aide"
-	) , 
-	"stocks" => array ( 
-		"Reagent" => "Réactif", 
-		"Lot_Number" => "Numéro de lot", 
-		"Expiry_Date" => "Date d'expiration", 
-		"Manufacturer" => "Fabricant", 
-		"Supplier" => "Fournisseur", 
-		"Quantity_Supplied" => "Quantité fournie", 
-		"Units" => "Unités", 
-		"Cost_per_Unit" => "Coût unitaire", 
-		"Date_received" => "Date de réception de", 
-		"Quantity" => "Quantité", 
-		"Edit" => "Modifier", 
-		"Inventory" => "Inventaire", 
-		"Current_Inventory" => "Inventaire en cours", 
-		"Quantity_Signed_Out" => "Quantité Signé Out", 
-		"Receiver" => "Récepteur", 
-		"Remarks" => "Remarques", 
-		"View_Current_Inventory" => "Voir l'inventaire actuel", 
-		"Current_Stock" => "Stock actuel", 
-		"Add_Stock" => "Add Stock", 
-		"Edit_Stock" => "Modifier Stock", 
-		"Total_Quantity" => "", 
-		"Quantity_In_Stock" => "Quantité totale", 
-		"TIPS_ADD" => "Ajouter les détails du stock sous la forme proposée. Pour ajouter plus d'un des articles en stock, vous pouvez sélectionner Ajouter un autre.", 
+	) ,
+	"stocks" => array (
+		"Reagent" => "Réactif",
+		"Lot_Number" => "Numéro de lot",
+		"Expiry_Date" => "Date d'expiration",
+		"Manufacturer" => "Fabricant",
+		"Supplier" => "Fournisseur",
+		"Quantity_Supplied" => "Quantité fournie",
+		"Units" => "Unités",
+		"Cost_per_Unit" => "Coût unitaire",
+		"Date_received" => "Date de réception de",
+		"Quantity" => "Quantité",
+		"Edit" => "Modifier",
+		"Inventory" => "Inventaire",
+		"Current_Inventory" => "Inventaire en cours",
+		"Quantity_Signed_Out" => "Quantité Signé Out",
+		"Receiver" => "Récepteur",
+		"Remarks" => "Remarques",
+		"View_Current_Inventory" => "Voir l'inventaire actuel",
+		"Current_Stock" => "Stock actuel",
+		"Add_Stock" => "Add Stock",
+		"Edit_Stock" => "Modifier Stock",
+		"Total_Quantity" => "",
+		"Quantity_In_Stock" => "Quantité totale",
+		"TIPS_ADD" => "Ajouter les détails du stock sous la forme proposée. Pour ajouter plus d'un des articles en stock, vous pouvez sélectionner Ajouter un autre.",
 		"TIPS_UPDATE" => "Entrez la quantité du stock en cours d'extraction. Le tableau ci-dessous donne compte de l'inventaire actuel."
-	) , 
-	"lang_modify" => array ( 
-		"TITLE" => "Modifier la Langue", 
-		"CATEGORY" => "Catégorie", 
-		"TIPS_MODIFYLANG_1" => "Cette page permet de basculer entre les langues. Il vous permet également de changer le contenu affiché.", 
+	) ,
+	"lang_modify" => array (
+		"TITLE" => "Modifier la Langue",
+		"CATEGORY" => "Catégorie",
+		"TIPS_MODIFYLANG_1" => "Cette page permet de basculer entre les langues. Il vous permet également de changer le contenu affiché.",
 		"TIPS_MODIFYLANG_2" => "Vous pouvez également choisir de changer la langue pour quelques pages en utilisant cette option. Les pages sont répertoriées sous forme de menu déroulant."
-	) , 
-	"blis_help_page" => array ( 
-		"TITLE" => "Aide page", 
-		"MENU_LABCONFIG" => "Lab Configuration", 
-		"MENU_SUMMARY" => "Sommaire", 
-		"TIPS_SUMMARY_1" => "Donne un résumé de l'installation de laboratoire, y compris les comptes personnel de laboratoire, les essais et les types d'échantillons.", 
-		"TIPS_SUMMARY_2" => "-", 
-		"TIPS_SUMMARY_3" => "-", 
-		"MENU_TESTS" => "Tests", 
-		"TIPS_SPECIMENTESTTYPES" => "Spécimen / Types de test - Cette option nous permet de configurer les échantillons peuvent être collectés et des essais qui sont effectués au laboratoire particulier.", 
-		"TIPS_TARGETTAT" => "Cible TAT - Cette liste d'options le délai d'exécution pour chaque type de test. Nous pouvons modifier le TAT en utilisant l'option Modifier et modifier la valeur.", 
-		"TIPS_RESULTINTERPRETATION" => "Interprétation des résultats - Cette option nous permet de préciser l'interprétation de plusieurs plages de valeurs pour chaque type de test.", 
-		"MENU_L_REPORTS" => "Rapports", 
-		"MENU_R_INFECTIONREPORT" => "Infection Rapport", 
-		"TIPS_INFECTIONREPORT" => "Ceci est une page modifiable pour le réglage des différents paramètres associaated chaque type d'épreuve individuelle.", 
-		"MENU_R_DRS" => "Paramètres Rapport journalier", 
-		"TIPS_DAILYREPORTSETTINGS" => "Ceci est une page modifiable pour la mise la mise en page du journal de patient, journal de log des échantillons et quotidiennes des patients.", 
-		"MENU_R_WORKSHEETS" => "Feuilles de travail", 
-		"TIPS_WORKSHEETS" => "Ceci est une page modifiable pour créer des feuilles de calcul personnalisées.", 
-		"MENU_USERACCOUNTS" => "Comptes d'utilisateurs", 
-		"TIPS_USERACCOUNTS_1" => "Cette page donne un résumé des différents utilisateurs ont accès au système.", 
-		"TIPS_USERACCOUNTS_2" => "Il permet aussi la création de nouveaux comptes utilisateurs.", 
-		"TIPS_USERACCOUNTS_3" => "Il permet de modifier les paramètres du compte, supprimer des comptes et de surveiller l'activité du compte.", 
-		"MENU_REGISTRATIONFIELDS" => "Registration Fields", 
-		"TIPS_REGISTRATIONFIELDS_1" => "Cette page permet la configuration de la page d'inscription des patients. Il permet de créer des champs obligatoires et cache certains domaines si elles ne sont pas souhaités à pourvoir.", 
-		"TIPS_REGISTRATIONFIELDS_2" => "Il permet aussi la création de certains des champs personnalisés pour l'enregistrement des patients et de nouveaux addition de l'échantillon qui peut être requis par certains laboratoires seulement.", 
-		"TIPS_REGISTRATIONFIELDS_3" => "-", 
-		"MENU_MODIFYLANG" => "Modifier la langue", 
-		"TIPS_MODIFYLANG_1" => "Cette page permet de basculer entre les langues.", 
-		"TIPS_MODIFYLANG_2" => "Vous pouvez également choisir de changer la langue pour quelques pages en utilisant cette option. Les pages sont répertoriées sous forme de menu déroulant.", 
-		"MENU_SETUPNETWORK" => "Configuration du réseau", 
-		"TIPS_SETUPNETWORK_1" => "Connectez-vous en tant qu'administrateur.", 
-		"TIPS_SETUPNETWORK_2" => "Cliquez sur l'option de configuration du réseau dans l'onglet Configuration Lab.", 
-		"TIPS_SETUPNETWORK_3" => "Vous allez maintenant être en mesure d'accéder en cliquant sur ??BLIS BlisSetup.html et en entrant votre nom d'utilisateur et mot de passe.", 
-		"TIPS_SETUPNETWORK_4" => "Dans le cas d'un redémarrage de l'ordinateur ou cliquez sur une défaillance du réseau d'installation de réseau option à nouveau.", 
-		"MENU_EXPORTCONFIG" => "Exporter la configuration", 
-		"TIPS_EXPORTCONFIG" => "Cette option donne le résumé de la configuration pour l'ensemble du système en tant que document Microsoft Word.", 
-		"MENU_BACKUP" => "Sauvegarde des données", 
-		"TIPS_BACKUP" => "Cette option permet de faire une sauvegarde de l'ensemble du système pour éviter la perte de données.", 
-		"MENU_REVERT" => "Revenir à la sauvegarde", 
-		"TIPS_REVERT" => "En cas de défaillance du système, si nous voulons revenir à une copie précédemment backedup des données, nous pouvons utiliser cette option.", 
-		"MENU_TESTCATALOG" => "Catalogue Test", 
-		"MENU_TC_TESTTYPE" => "Type d'essai", 
-		"TIPS_TC_TESTTYPE_1" => "Donne une liste complète des types de tests disponibles.", 
-		"TIPS_TC_TESTTYPE_2" => "Nous pouvons utiliser cette page pour modifier des types de tests actuellement.", 
-		"TIPS_TC_TESTTYPE_3" => "Nouveaux types de tests peuvent également être ajoutés en utilisant l'option Ajouter situé près du haut de la page.", 
-		"MENU_TC_SPECIMENTYPE" => "Type d'échantillon", 
-		"TIPS_TC_SPECIMENTYPE_1" => "Donne une liste complète des types d'échantillons disponibles.", 
-		"TIPS_TC_SPECIMENTYPE_2" => "Nous pouvons utiliser cette page pour modifier les types d'échantillons existants. Tests compatible avec un type de spécimen particulier peut ajouter / supprimer en cochant les cases appropriées trouve ici.", 
-		"TIPS_TC_SPECIMENTYPE_3" => "Types d'échantillons New peuvent également être ajoutés en utilisant l'option Ajouter situé près du haut de la page.", 
-		"MENU_BACKUPDATA" => "Sauvegarde des données", 
-		"TIPS_BACKUPDATA" => "Cette option prend une sauvegarde du système et le stocke sur le disque local. Cette option n'est disponible que pour l'administrateur.", 
-		"MENU_REGISTRATION" => "Enregistrement", 
-		"TIPS_REGISTRATION_1" => "Cette page nous permet d'enregistrer de nouveaux patients ou de patients de recherche existants basés sur le nom, l'identité du patient ou le numéro.", 
-		"TIPS_REGISTRATION_2" => "Une fois qu'un patient a été enregistré, nous pouvons utiliser cette page pour visualiser, modifier le profil des patients.", 
-		"TIPS_REGISTRATION_3" => "Nous pouvons également enregistrer un spécimen du patient a fourni pour un test donné.", 
-		"MENU_RESULTS" => "Résultats", 
-		"MENU_R_SINGLESPECIMENRESULTS" => "Résultats seul spécimen", 
-		"TIPS_R_SINGLESPECIMENRESULTS" => "Cette option nous permet d'ajouter de résultats pour un patient particulier sur la base des spécimens qu'il a fournies.", 
-		"MENU_R_BATCHRESULTS" => "Résultats lot", 
-		"TIPS_R_BATCHRESULTS" => "Cette option nous permet d'ajouter les résultats d'un essai de type particulier.", 
-		"MENU_R_VERIFYRESULTS" => "Vérifier les résultats", 
-		"TIPS_R_VERIFYRESULTS" => "Cette option nous permet de vérifier le résultat en fonction du type de test. Il affiche la liste des résultats pour tous les patients dont les résultats n'ont pas été vérifiées. Vous pouvez modifier les résultats et entrez remarques avant de vérifier les résultats.", 
-		"MENU_R_WORKSHEET" => "Feuille de travail", 
-		"TIPS_R_WORKSHEET_1" => "Cette option génère une feuille de calcul basée sur la section de laboratoire et d'essai de type. Vous pouvez également utiliser une feuille de calcul personnalisé qui peut être créé par les Admins en utilisant Lab Configuration> Tests> Rapports> Feuille de travail.", 
-		"TIPS_R_WORKSHEET_2" => "Vous pouvez créer une feuille de calcul vide en choisissant l'option Conserver les blancs et précisant le nombre de lignes dont vous avez besoin.", 
-		"MENU_SEARCH" => "Rechercher", 
-		"TIPS_SEARCH_1" => "Vous pouvez utiliser cette page pour rechercher des patients inscrits. Entrez un nom partiel de générer une liste de patients contenant la chaîne.", 
-		"TIPS_SEARCH_2" => "Une fois que vous sélectionnez le patient, vous pouvez visualiser le profil du patient, rechercher les tests qu'il a subis et les résultats de ces tests.", 
-		"TIPS_SEARCH_3" => "-", 
-		"MENU_INVENTORY" => "Inventaire", 
-		"TIPS_INVENTORY_1" => "Cette page affiche les quantités actuelles réactif en stock.", 
-		"TIPS_INVENTORY_2" => "Vous pouvez mettre à jour des stocks que vous acquérez de nouveaux réactifs en ajoutant le nom du réactif, la quantité reçue, le nom du destinataire et remarques.", 
-		"TIPS_INVENTORY_3" => "-", 
-		"MENU_REPORTS" => "Rapports", 
-		"MENU_DAILYREPORTS" => "Rapports quotidiens", 
-		"TIPS_PATIENTREPORT" => "Patient Rapport: Cette option nous permet de générer le rapport pour un patient particulier sur la base soit, le nom du patient, le numéro du patient ou de l'identité du patient.", 
-		"TIPS_DAILYLOG" => "Daily news: Cette option génère un jour ou l'autre des dossiers des patients ou de test à partir de la date de début particulière à la date de fin. Vous pouvez personnaliser le type de rapport que vous souhaitez en choisissant une option de la section (par exemple, sérologie, hématologie) et / ou du type d'essai (par exemple ALT / SGPT, HGB) menus déroulants.", 
-		"MENU_AGGREGATEREPORTS" => "Des rapports agrégés", 
-		"TIPS_PREVALENCERATE" => "Taux de prévalence: Ce rapport donne la prévalence de l'infection par un type particulier basé sur le nombre de tests effectués et les résultats. Vous pouvez entrer dans la période déterminée pour laquelle vous voulez vérifier le taux de prévalence d'une infection particulière. Vous pouvez également vues sur les tendances de l'infection pendant la période définie, comme un graphique, en cliquant sur l'option Tendances lorsque le rapport est affiché.", 
-		"TIPS_COUNTS" => "Chiffres: Cette option génère un rapport pour une période donnée de temps basé sur le nombre de tests, le spécimen ou statistiques médecin.", 
-		"TIPS_TURNAROUNDTIME" => "Délai d'exécution: Cette option génère le délai d'exécution pour la totalité ou d'essai choisie pour une période de temps donnée. Il génère également un graphique des statistiques. Vous pouvez choisir d'inclure les deux essais en cours ou terminées ou tout simplement les tests effectués.", 
-		"TIPS_INFECTIONREPORT" => "Infection Rapport: Cette option génère un rapport global d'infection pour une période donnée pour une ou toutes les sections de laboratoire. Il propose également une option pour créer un document Word du rapport généré.", 
+	) ,
+	"blis_help_page" => array (
+		"TITLE" => "Aide page",
+		"MENU_LABCONFIG" => "Lab Configuration",
+		"MENU_SUMMARY" => "Sommaire",
+		"TIPS_SUMMARY_1" => "Donne un résumé de l'installation de laboratoire, y compris les comptes personnel de laboratoire, les essais et les types d'échantillons.",
+		"TIPS_SUMMARY_2" => "-",
+		"TIPS_SUMMARY_3" => "-",
+		"MENU_TESTS" => "Tests",
+		"TIPS_SPECIMENTESTTYPES" => "Spécimen / Types de test - Cette option nous permet de configurer les échantillons peuvent être collectés et des essais qui sont effectués au laboratoire particulier.",
+		"TIPS_TARGETTAT" => "Cible TAT - Cette liste d'options le délai d'exécution pour chaque type de test. Nous pouvons modifier le TAT en utilisant l'option Modifier et modifier la valeur.",
+		"TIPS_RESULTINTERPRETATION" => "Interprétation des résultats - Cette option nous permet de préciser l'interprétation de plusieurs plages de valeurs pour chaque type de test.",
+		"MENU_L_REPORTS" => "Rapports",
+		"MENU_R_INFECTIONREPORT" => "Infection Rapport",
+		"TIPS_INFECTIONREPORT" => "Ceci est une page modifiable pour le réglage des différents paramètres associaated chaque type d'épreuve individuelle.",
+		"MENU_R_DRS" => "Paramètres Rapport journalier",
+		"TIPS_DAILYREPORTSETTINGS" => "Ceci est une page modifiable pour la mise la mise en page du journal de patient, journal de log des échantillons et quotidiennes des patients.",
+		"MENU_R_WORKSHEETS" => "Feuilles de travail",
+		"TIPS_WORKSHEETS" => "Ceci est une page modifiable pour créer des feuilles de calcul personnalisées.",
+		"MENU_USERACCOUNTS" => "Comptes d'utilisateurs",
+		"TIPS_USERACCOUNTS_1" => "Cette page donne un résumé des différents utilisateurs ont accès au système.",
+		"TIPS_USERACCOUNTS_2" => "Il permet aussi la création de nouveaux comptes utilisateurs.",
+		"TIPS_USERACCOUNTS_3" => "Il permet de modifier les paramètres du compte, supprimer des comptes et de surveiller l'activité du compte.",
+		"MENU_REGISTRATIONFIELDS" => "Registration Fields",
+		"TIPS_REGISTRATIONFIELDS_1" => "Cette page permet la configuration de la page d'inscription des patients. Il permet de créer des champs obligatoires et cache certains domaines si elles ne sont pas souhaités à pourvoir.",
+		"TIPS_REGISTRATIONFIELDS_2" => "Il permet aussi la création de certains des champs personnalisés pour l'enregistrement des patients et de nouveaux addition de l'échantillon qui peut être requis par certains laboratoires seulement.",
+		"TIPS_REGISTRATIONFIELDS_3" => "-",
+		"MENU_MODIFYLANG" => "Modifier la langue",
+		"TIPS_MODIFYLANG_1" => "Cette page permet de basculer entre les langues.",
+		"TIPS_MODIFYLANG_2" => "Vous pouvez également choisir de changer la langue pour quelques pages en utilisant cette option. Les pages sont répertoriées sous forme de menu déroulant.",
+		"MENU_SETUPNETWORK" => "Configuration du réseau",
+		"TIPS_SETUPNETWORK_1" => "Connectez-vous en tant qu'administrateur.",
+		"TIPS_SETUPNETWORK_2" => "Cliquez sur l'option de configuration du réseau dans l'onglet Configuration Lab.",
+		"TIPS_SETUPNETWORK_3" => "Vous allez maintenant être en mesure d'accéder en cliquant sur ??BLIS BlisSetup.html et en entrant votre nom d'utilisateur et mot de passe.",
+		"TIPS_SETUPNETWORK_4" => "Dans le cas d'un redémarrage de l'ordinateur ou cliquez sur une défaillance du réseau d'installation de réseau option à nouveau.",
+		"MENU_EXPORTCONFIG" => "Exporter la configuration",
+		"TIPS_EXPORTCONFIG" => "Cette option donne le résumé de la configuration pour l'ensemble du système en tant que document Microsoft Word.",
+		"MENU_BACKUP" => "Sauvegarde des données",
+		"TIPS_BACKUP" => "Cette option permet de faire une sauvegarde de l'ensemble du système pour éviter la perte de données.",
+		"MENU_REVERT" => "Revenir à la sauvegarde",
+		"TIPS_REVERT" => "En cas de défaillance du système, si nous voulons revenir à une copie précédemment backedup des données, nous pouvons utiliser cette option.",
+		"MENU_TESTCATALOG" => "Catalogue Test",
+		"MENU_TC_TESTTYPE" => "Type d'essai",
+		"TIPS_TC_TESTTYPE_1" => "Donne une liste complète des types de tests disponibles.",
+		"TIPS_TC_TESTTYPE_2" => "Nous pouvons utiliser cette page pour modifier des types de tests actuellement.",
+		"TIPS_TC_TESTTYPE_3" => "Nouveaux types de tests peuvent également être ajoutés en utilisant l'option Ajouter situé près du haut de la page.",
+		"MENU_TC_SPECIMENTYPE" => "Type d'échantillon",
+		"TIPS_TC_SPECIMENTYPE_1" => "Donne une liste complète des types d'échantillons disponibles.",
+		"TIPS_TC_SPECIMENTYPE_2" => "Nous pouvons utiliser cette page pour modifier les types d'échantillons existants. Tests compatible avec un type de spécimen particulier peut ajouter / supprimer en cochant les cases appropriées trouve ici.",
+		"TIPS_TC_SPECIMENTYPE_3" => "Types d'échantillons New peuvent également être ajoutés en utilisant l'option Ajouter situé près du haut de la page.",
+		"MENU_BACKUPDATA" => "Sauvegarde des données",
+		"TIPS_BACKUPDATA" => "Cette option prend une sauvegarde du système et le stocke sur le disque local. Cette option n'est disponible que pour l'administrateur.",
+		"MENU_REGISTRATION" => "Enregistrement",
+		"TIPS_REGISTRATION_1" => "Cette page nous permet d'enregistrer de nouveaux patients ou de patients de recherche existants basés sur le nom, l'identité du patient ou le numéro.",
+		"TIPS_REGISTRATION_2" => "Une fois qu'un patient a été enregistré, nous pouvons utiliser cette page pour visualiser, modifier le profil des patients.",
+		"TIPS_REGISTRATION_3" => "Nous pouvons également enregistrer un spécimen du patient a fourni pour un test donné.",
+		"MENU_RESULTS" => "Résultats",
+		"MENU_R_SINGLESPECIMENRESULTS" => "Résultats seul spécimen",
+		"TIPS_R_SINGLESPECIMENRESULTS" => "Cette option nous permet d'ajouter de résultats pour un patient particulier sur la base des spécimens qu'il a fournies.",
+		"MENU_R_BATCHRESULTS" => "Résultats lot",
+		"TIPS_R_BATCHRESULTS" => "Cette option nous permet d'ajouter les résultats d'un essai de type particulier.",
+		"MENU_R_VERIFYRESULTS" => "Vérifier les résultats",
+		"TIPS_R_VERIFYRESULTS" => "Cette option nous permet de vérifier le résultat en fonction du type de test. Il affiche la liste des résultats pour tous les patients dont les résultats n'ont pas été vérifiées. Vous pouvez modifier les résultats et entrez remarques avant de vérifier les résultats.",
+		"MENU_R_WORKSHEET" => "Feuille de travail",
+		"TIPS_R_WORKSHEET_1" => "Cette option génère une feuille de calcul basée sur la section de laboratoire et d'essai de type. Vous pouvez également utiliser une feuille de calcul personnalisé qui peut être créé par les Admins en utilisant Lab Configuration> Tests> Rapports> Feuille de travail.",
+		"TIPS_R_WORKSHEET_2" => "Vous pouvez créer une feuille de calcul vide en choisissant l'option Conserver les blancs et précisant le nombre de lignes dont vous avez besoin.",
+		"MENU_SEARCH" => "Rechercher",
+		"TIPS_SEARCH_1" => "Vous pouvez utiliser cette page pour rechercher des patients inscrits. Entrez un nom partiel de générer une liste de patients contenant la chaîne.",
+		"TIPS_SEARCH_2" => "Une fois que vous sélectionnez le patient, vous pouvez visualiser le profil du patient, rechercher les tests qu'il a subis et les résultats de ces tests.",
+		"TIPS_SEARCH_3" => "-",
+		"MENU_INVENTORY" => "Inventaire",
+		"TIPS_INVENTORY_1" => "Cette page affiche les quantités actuelles réactif en stock.",
+		"TIPS_INVENTORY_2" => "Vous pouvez mettre à jour des stocks que vous acquérez de nouveaux réactifs en ajoutant le nom du réactif, la quantité reçue, le nom du destinataire et remarques.",
+		"TIPS_INVENTORY_3" => "-",
+		"MENU_REPORTS" => "Rapports",
+		"MENU_DAILYREPORTS" => "Rapports quotidiens",
+		"TIPS_PATIENTREPORT" => "Patient Rapport: Cette option nous permet de générer le rapport pour un patient particulier sur la base soit, le nom du patient, le numéro du patient ou de l'identité du patient.",
+		"TIPS_DAILYLOG" => "Daily news: Cette option génère un jour ou l'autre des dossiers des patients ou de test à partir de la date de début particulière à la date de fin. Vous pouvez personnaliser le type de rapport que vous souhaitez en choisissant une option de la section (par exemple, sérologie, hématologie) et / ou du type d'essai (par exemple ALT / SGPT, HGB) menus déroulants.",
+		"MENU_AGGREGATEREPORTS" => "Des rapports agrégés",
+		"TIPS_PREVALENCERATE" => "Taux de prévalence: Ce rapport donne la prévalence de l'infection par un type particulier basé sur le nombre de tests effectués et les résultats. Vous pouvez entrer dans la période déterminée pour laquelle vous voulez vérifier le taux de prévalence d'une infection particulière. Vous pouvez également vues sur les tendances de l'infection pendant la période définie, comme un graphique, en cliquant sur l'option Tendances lorsque le rapport est affiché.",
+		"TIPS_COUNTS" => "Chiffres: Cette option génère un rapport pour une période donnée de temps basé sur le nombre de tests, le spécimen ou statistiques médecin.",
+		"TIPS_TURNAROUNDTIME" => "Délai d'exécution: Cette option génère le délai d'exécution pour la totalité ou d'essai choisie pour une période de temps donnée. Il génère également un graphique des statistiques. Vous pouvez choisir d'inclure les deux essais en cours ou terminées ou tout simplement les tests effectués.",
+		"TIPS_INFECTIONREPORT" => "Infection Rapport: Cette option génère un rapport global d'infection pour une période donnée pour une ou toutes les sections de laboratoire. Il propose également une option pour créer un document Word du rapport généré.",
 		"TIPS_INVENTORY" => "Inventaire: Cette option génère une liste de l'inventaire actuel."
        )  ,
 	"search_condition" => array (
 	"SEARCH_BEGIN_WITH"=>"Commence par",
 	"SEARCH_END_WITH"=>"Fin d'",
 	"SEARCH_CONTAINS"=>"Contient"
-	) 
+	)
 );
 
-include("../lang/lang_util.php");
+include_once("../lang/lang_util.php");
 ?>
