+++ htdocs/Language/remarks.php
@@ -1,519 +1,519 @@
 <?php
-$REMARKS_ARRAY = array ( 
-		"2" => array ( 
+$REMARKS_ARRAY = array (
+		"2" => array (
 		"4.0:10.0" => ""
-	) , 
-	"3" => array ( 
-		"N" => "", 
+	) ,
+	"3" => array (
+		"N" => "",
 		"P" => ""
-	) , 
-	"4" => array ( 
+	) ,
+	"4" => array (
 		"0:3000" => ""
-	) , 
-	"5" => array ( 
+	) ,
+	"5" => array (
 		"0:90" => ""
-	) , 
-	"6" => array ( 
+	) ,
+	"6" => array (
 		"0:1000" => ""
-	) , 
-	"7" => array ( 
+	) ,
+	"7" => array (
 		"60:110" => ""
-	) , 
-	"8" => array ( 
+	) ,
+	"8" => array (
 		"5:1000" => ""
-	) , 
-	"9" => array ( 
+	) ,
+	"9" => array (
 		"3.4:7" => ""
-	) , 
-	"10" => array ( 
+	) ,
+	"10" => array (
 		"8.1:10.5" => ""
-	) , 
-	"11" => array ( 
+	) ,
+	"11" => array (
 		"1:10" => ""
-	) , 
-	"12" => array ( 
+	) ,
+	"12" => array (
 		"1.8:2.4" => ""
-	) , 
-	"13" => array ( 
+	) ,
+	"13" => array (
 		"0.73:1.34" => ""
-	) , 
-	"14" => array ( 
+	) ,
+	"14" => array (
 		"0:1000" => ""
-	) , 
-	"15" => array ( 
+	) ,
+	"15" => array (
 		"0:1000" => ""
-	) , 
-	"17" => array ( 
+	) ,
+	"17" => array (
 		"0:1000" => ""
-	) , 
-	"18" => array ( 
-	) , 
-	"19" => array ( 
+	) ,
+	"18" => array (
+	) ,
+	"19" => array (
 		"0:50" => ""
-	) , 
-	"74" => array ( 
-		"P" => "", 
+	) ,
+	"74" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"21" => array ( 
+	) ,
+	"21" => array (
 		"136:146" => ""
-	) , 
-	"22" => array ( 
+	) ,
+	"22" => array (
 		"3.5:5.1" => ""
-	) , 
-	"23" => array ( 
+	) ,
+	"23" => array (
 		"96:106" => ""
-	) , 
-	"24" => array ( 
+	) ,
+	"24" => array (
 		"0:100" => ""
-	) , 
-	"25" => array ( 
+	) ,
+	"25" => array (
 		"130:250" => ""
-	) , 
-	"26" => array ( 
+	) ,
+	"26" => array (
 		"44:184" => ""
-	) , 
-	"27" => array ( 
+	) ,
+	"27" => array (
 		"45:" => ""
-	) , 
-	"28" => array ( 
+	) ,
+	"28" => array (
 		":158" => ""
-	) , 
-	"29" => array ( 
+	) ,
+	"29" => array (
 		"0:1000" => ""
-	) , 
-	"30" => array ( 
+	) ,
+	"30" => array (
 		"0:1000" => ""
-	) , 
-	"31" => array ( 
+	) ,
+	"31" => array (
 		"0:1000" => ""
-	) , 
-	"33" => array ( 
+	) ,
+	"33" => array (
 		"3.5:5.5" => ""
-	) , 
-	"34" => array ( 
+	) ,
+	"34" => array (
 		"0.5:15.0" => ""
-	) , 
-	"35" => array ( 
+	) ,
+	"35" => array (
 		"5:75" => ""
-	) , 
-	"36" => array ( 
+	) ,
+	"36" => array (
 		"80.0:99.0" => ""
-	) , 
-	"37" => array ( 
+	) ,
+	"37" => array (
 		"26.0:32.0" => ""
-	) , 
-	"38" => array ( 
+	) ,
+	"38" => array (
 		"32.0:36.0" => ""
-	) , 
-	"39" => array ( 
+	) ,
+	"39" => array (
 		"100:300" => ""
-	) , 
-	"40" => array ( 
-		"P" => "", 
+	) ,
+	"40" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"41" => array ( 
+	) ,
+	"41" => array (
 		"400:1000000" => ""
-	) , 
-	"42" => array ( 
-		"P" => "", 
+	) ,
+	"42" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"43" => array ( 
-		"P" => "", 
+	) ,
+	"43" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"44" => array ( 
-		"N" => "", 
+	) ,
+	"44" => array (
+		"N" => "",
 		"P" => ""
-	) , 
-	"45" => array ( 
-		"R" => "", 
+	) ,
+	"45" => array (
+		"R" => "",
 		"NR" => ""
-	) , 
-	"46" => array ( 
-		"AA" => "", 
-		"AF" => "", 
-		"AS" => "", 
+	) ,
+	"46" => array (
+		"AA" => "",
+		"AF" => "",
+		"AS" => "",
 		"SS" => ""
-	) , 
-	"47" => array ( 
+	) ,
+	"47" => array (
 		"0:1000" => ""
-	) , 
-	"48" => array ( 
-		"P" => "", 
+	) ,
+	"48" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"49" => array ( 
+	) ,
+	"49" => array (
 		"0:100" => ""
-	) , 
-	"50" => array ( 
+	) ,
+	"50" => array (
 		"10:45" => ""
-	) , 
-	"51" => array ( 
-		"Straw" => "", 
-		"Pale Yellow" => "", 
-		"Deep Yellow" => "", 
+	) ,
+	"51" => array (
+		"Straw" => "",
+		"Pale Yellow" => "",
+		"Deep Yellow" => "",
 		"Amber" => ""
-	) , 
-	"52" => array ( 
-		"Clear" => "", 
+	) ,
+	"52" => array (
+		"Clear" => "",
 		"Cloudy" => ""
-	) , 
-	"105" => array ( 
-		"No ova or protozoa seen" => "", 
+	) ,
+	"105" => array (
+		"No ova or protozoa seen" => "",
 		"Protozoa|ova seen" => ""
-	) , 
-	"54" => array ( 
-		"Formed(solid)" => "", 
-		"Semi-formed(not solid)" => "", 
+	) ,
+	"54" => array (
+		"Formed(solid)" => "",
+		"Semi-formed(not solid)" => "",
 		"Watery" => ""
-	) , 
-	"55" => array ( 
-		"P" => "", 
+	) ,
+	"55" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"56" => array ( 
-		"P" => "", 
+	) ,
+	"56" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"57" => array ( 
-		"Non reactive" => "", 
+	) ,
+	"57" => array (
+		"Non reactive" => "",
 		"Reactive" => ""
-	) , 
-	"58" => array ( 
-		"P" => "", 
+	) ,
+	"58" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"59" => array ( 
+	) ,
+	"59" => array (
 		"45:72" => ""
-	) , 
-	"60" => array ( 
+	) ,
+	"60" => array (
 		"0.15:0.45" => ""
-	) , 
-	"61" => array ( 
+	) ,
+	"61" => array (
 		"120:126" => ""
-	) , 
-	"62" => array ( 
-		"N" => "", 
+	) ,
+	"62" => array (
+		"N" => "",
 		"P" => ""
-	) , 
-	"63" => array ( 
+	) ,
+	"63" => array (
 		"0:1000" => ""
-	) , 
-	"66" => array ( 
-		"P" => "", 
+	) ,
+	"66" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"67" => array ( 
+	) ,
+	"67" => array (
 		"0:1000" => ""
-	) , 
-	"68" => array ( 
-		"Positive" => "", 
+	) ,
+	"68" => array (
+		"Positive" => "",
 		"Negative" => ""
-	) , 
-	"69" => array ( 
-	) , 
-	"70" => array ( 
+	) ,
+	"69" => array (
+	) ,
+	"70" => array (
 		":6" => ""
-	) , 
-	"71" => array ( 
-		"P" => "", 
+	) ,
+	"71" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"72" => array ( 
-		"AAFB seen" => "", 
+	) ,
+	"72" => array (
+		"AAFB seen" => "",
 		"AAFB not seen" => ""
-	) , 
-	"73" => array ( 
+	) ,
+	"73" => array (
 		"0:100" => ""
-	) , 
-	"75" => array ( 
+	) ,
+	"75" => array (
 		"0:1000" => ""
-	) , 
-	"76" => array ( 
+	) ,
+	"76" => array (
 		"0:1000" => ""
-	) , 
-	"77" => array ( 
+	) ,
+	"77" => array (
 		"0:1000" => ""
-	) , 
-	"88" => array ( 
+	) ,
+	"88" => array (
 		"3.6:12" => ""
-	) , 
-	"87" => array ( 
-		"NR" => "", 
-		"1+" => "", 
-		"2+" => "", 
-		"3+" => "", 
+	) ,
+	"87" => array (
+		"NR" => "",
+		"1+" => "",
+		"2+" => "",
+		"3+" => "",
 		"4+" => ""
-	) , 
-	"85" => array ( 
-		"P" => "", 
+	) ,
+	"85" => array (
+		"P" => "",
 		"N" => ""
-	) , 
-	"84" => array ( 
+	) ,
+	"84" => array (
 		"12:30" => ""
-	) , 
-	"89" => array ( 
-		"N" => "", 
+	) ,
+	"89" => array (
+		"N" => "",
 		"P" => ""
-	) , 
-	"90" => array ( 
+	) ,
+	"90" => array (
 		"350:1610" => ""
-	) , 
-	"91" => array ( 
+	) ,
+	"91" => array (
 		"130:1118" => ""
-	) , 
-	"92" => array ( 
+	) ,
+	"92" => array (
 		"190:1955" => ""
-	) , 
-	"93" => array ( 
+	) ,
+	"93" => array (
 		":" => ""
-	) , 
-	"94" => array ( 
-		"positif" => "", 
+	) ,
+	"94" => array (
+		"positif" => "",
 		"negatif" => ""
-	) , 
-	"95" => array ( 
-		"YES" => "", 
+	) ,
+	"95" => array (
+		"YES" => "",
 		"NO" => ""
-	) , 
-	"97" => array ( 
-		"positive" => "", 
+	) ,
+	"97" => array (
+		"positive" => "",
 		"negative" => ""
-	) , 
-	"98" => array ( 
+	) ,
+	"98" => array (
 		":200" => ""
-	) , 
-	"99" => array ( 
+	) ,
+	"99" => array (
 		":" => ""
-	) , 
-	"100" => array ( 
+	) ,
+	"100" => array (
 		":4" => ""
-	) , 
-	"101" => array ( 
+	) ,
+	"101" => array (
 		"16:256" => ""
-	) , 
-	"102" => array ( 
-	) , 
-	"103" => array ( 
+	) ,
+	"102" => array (
+	) ,
+	"103" => array (
 		":" => ""
-	) , 
-	"104" => array ( 
+	) ,
+	"104" => array (
 		":" => ""
-	) , 
-	"106" => array ( 
-	) , 
-	"107" => array ( 
-		"Offensive" => "", 
+	) ,
+	"106" => array (
+	) ,
+	"107" => array (
+		"Offensive" => "",
 		"Foul" => ""
-	) , 
-	"108" => array ( 
-		"Present" => "", 
+	) ,
+	"108" => array (
+		"Present" => "",
 		"Not present" => ""
-	) , 
-	"109" => array ( 
-		"Present" => "", 
+	) ,
+	"109" => array (
+		"Present" => "",
 		"Not present" => ""
-	) , 
-	"110" => array ( 
-		"Whitish" => "", 
-		"Cream White" => "", 
-		"Yellowish" => "", 
+	) ,
+	"110" => array (
+		"Whitish" => "",
+		"Cream White" => "",
+		"Yellowish" => "",
 		"Pale Yellow" => ""
-	) , 
-	"111" => array ( 
-		"Discharge" => "", 
+	) ,
+	"111" => array (
+		"Discharge" => "",
 		"No discharge" => ""
-	) , 
-	"112" => array ( 
-		"Offensive" => "", 
+	) ,
+	"112" => array (
+		"Offensive" => "",
 		"Non offensive" => ""
-	) , 
-	"113" => array ( 
-		"Discharge" => "", 
+	) ,
+	"113" => array (
+		"Discharge" => "",
 		"N" => ""
-	) , 
-	"114" => array ( 
-		"Normal" => "", 
+	) ,
+	"114" => array (
+		"Normal" => "",
 		"Pregnant" => ""
-	) , 
-	"115" => array ( 
-	) , 
-	"116" => array ( 
-		"few" => "", 
-		"moderate" => "", 
-		"rare" => "", 
+	) ,
+	"115" => array (
+	) ,
+	"116" => array (
+		"few" => "",
+		"moderate" => "",
+		"rare" => "",
 		"many" => ""
-	) , 
-	"117" => array ( 
-		"few" => "", 
-		"moderate" => "", 
-		"rare" => "", 
+	) ,
+	"117" => array (
+		"few" => "",
+		"moderate" => "",
+		"rare" => "",
 		"many" => ""
-	) , 
-	"118" => array ( 
-	) , 
-	"119" => array ( 
-	) , 
-	"120" => array ( 
-		"few" => "", 
-		"moderate" => "", 
-		"rare" => "", 
+	) ,
+	"118" => array (
+	) ,
+	"119" => array (
+	) ,
+	"120" => array (
+		"few" => "",
+		"moderate" => "",
+		"rare" => "",
 		"many" => ""
-	) , 
-	"121" => array ( 
-		"Type I" => "", 
-		"Type II" => "", 
-		"Type III" => "", 
+	) ,
+	"121" => array (
+		"Type I" => "",
+		"Type II" => "",
+		"Type III" => "",
 		"Type IV" => ""
-	) , 
-	"122" => array ( 
-		"Yellow" => "", 
+	) ,
+	"122" => array (
+		"Yellow" => "",
 		"Brown" => ""
-	) , 
-	"123" => array ( 
-		"Yes" => "", 
+	) ,
+	"123" => array (
+		"Yes" => "",
 		"No" => ""
-	) , 
-	"124" => array ( 
-	) , 
-	"125" => array ( 
-		"Present" => "", 
+	) ,
+	"124" => array (
+	) ,
+	"125" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"126" => array ( 
-		"Present" => "", 
+	) ,
+	"126" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"127" => array ( 
-		"Present" => "", 
+	) ,
+	"127" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"128" => array ( 
-		"Present" => "", 
+	) ,
+	"128" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"129" => array ( 
-		"Present" => "", 
+	) ,
+	"129" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"130" => array ( 
-		"Present" => "", 
+	) ,
+	"130" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"131" => array ( 
-		"Present" => "", 
+	) ,
+	"131" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"132" => array ( 
-		"Present" => "", 
+	) ,
+	"132" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"133" => array ( 
-		"Present" => "", 
+	) ,
+	"133" => array (
+		"Present" => "",
 		"Absent" => ""
-	) , 
-	"134" => array ( 
-		"Negative" => "", 
-		"+" => "", 
-		"++" => "", 
+	) ,
+	"134" => array (
+		"Negative" => "",
+		"+" => "",
+		"++" => "",
 		"+++" => ""
-	) , 
-	"135" => array ( 
-		"Normal" => "", 
-		"2(35)" => "", 
-		"4(70)" => "", 
-		"8(140)" => "", 
+	) ,
+	"135" => array (
+		"Normal" => "",
+		"2(35)" => "",
+		"4(70)" => "",
+		"8(140)" => "",
 		"12(200)" => ""
-	) , 
-	"136" => array ( 
-		"Negative" => "", 
-		"+" => "", 
-		"++" => "", 
+	) ,
+	"136" => array (
+		"Negative" => "",
+		"+" => "",
+		"++" => "",
 		"+++" => ""
-	) , 
-	"137" => array ( 
-		"Negative" => "", 
-		"+" => "", 
+	) ,
+	"137" => array (
+		"Negative" => "",
+		"+" => "",
 		"++" => ""
-	) , 
-	"138" => array ( 
-		"Normal" => "", 
-		"50(2.8)" => "", 
-		"150(8.3)" => "", 
-		"500(28)" => "", 
+	) ,
+	"138" => array (
+		"Normal" => "",
+		"50(2.8)" => "",
+		"150(8.3)" => "",
+		"500(28)" => "",
 		">=1000(56)" => ""
-	) , 
-	"139" => array ( 
-		"Negative" => "", 
-		"30(0.3)" => "", 
-		"100(1)" => "", 
+	) ,
+	"139" => array (
+		"Negative" => "",
+		"30(0.3)" => "",
+		"100(1)" => "",
 		"500(5)" => ""
-	) , 
-	"140" => array ( 
-		"Negative" => "", 
-		"+ca 5-10" => "", 
-		"++ca 50" => "", 
-		"+++ca 300" => "", 
-		"ca 5-10" => "", 
-		"ca 50" => "", 
+	) ,
+	"140" => array (
+		"Negative" => "",
+		"+ca 5-10" => "",
+		"++ca 50" => "",
+		"+++ca 300" => "",
+		"ca 5-10" => "",
+		"ca 50" => "",
 		"Ca 300" => ""
-	) , 
-	"141" => array ( 
-		"5" => "", 
-		"6" => "", 
-		"7" => "", 
-		"8" => "", 
+	) ,
+	"141" => array (
+		"5" => "",
+		"6" => "",
+		"7" => "",
+		"8" => "",
 		"9" => ""
-	) , 
-	"142" => array ( 
-		"Negative" => "", 
-		"Positive" => "", 
+	) ,
+	"142" => array (
+		"Negative" => "",
+		"Positive" => "",
 		"Pink colouration" => ""
-	) , 
-	"143" => array ( 
-		"Negative" => "", 
-		"Ca 25" => "", 
-		"Ca 75" => "", 
+	) ,
+	"143" => array (
+		"Negative" => "",
+		"Ca 25" => "",
+		"Ca 75" => "",
 		"Ca 500" => ""
-	) , 
-	"144" => array ( 
-		"1.00" => "", 
-		"1.005" => "", 
-		"1.010" => "", 
-		"1.015" => "", 
-		"1.020" => "", 
-		"1.025" => "", 
+	) ,
+	"144" => array (
+		"1.00" => "",
+		"1.005" => "",
+		"1.010" => "",
+		"1.015" => "",
+		"1.020" => "",
+		"1.025" => "",
 		"1.030" => ""
-	) , 
-	"32" => array ( 
-		"N" => "Negative", 
-		"1" => "Low", 
-		"2" => "Moderate", 
-		"3" => "High", 
+	) ,
+	"32" => array (
+		"N" => "Negative",
+		"1" => "Low",
+		"2" => "Moderate",
+		"3" => "High",
 		"4" => "Very High"
-	) , 
-	"64" => array ( 
-		"A" => "", 
-		"B" => "", 
-		"AB" => "", 
+	) ,
+	"64" => array (
+		"A" => "",
+		"B" => "",
+		"AB" => "",
 		"O" => ""
-	) , 
-	"65" => array ( 
-		"Rh +ve" => "Positive", 
+	) ,
+	"65" => array (
+		"Rh +ve" => "Positive",
 		"Rh -ve" => "Negative"
-	) , 
-	"96" => array ( 
-		"Positive" => "Positive", 
-		"Trace" => "Trace", 
+	) ,
+	"96" => array (
+		"Positive" => "Positive",
+		"Trace" => "Trace",
 		"Negative" => "Negative"
-	) , 
-	"16" => array ( 
+	) ,
+	"16" => array (
 		"6:30" => "Normal"
-	) 
+	)
 );
 
-require_once("../lang/lang_util.php");
+include_once("../lang/lang_util.php");
 ?>
