+++ htdocs/includes/footer.php
@@ -11,9 +11,9 @@
 		}
 	});
 	}
-        
-        
-        
+
+
+
 </script>
 
 <?php
@@ -22,6 +22,8 @@
 # Closes any open database connections, and-
 # displays footer so the users know the page is done loading.
 #
+
+include_once("../lang/lang_util.php");
 include("db_close.php");
 LangUtil::setPageId("footer");
 ?>
@@ -34,7 +36,7 @@ LangUtil::setPageId("footer");
 	<hr>
 	<div class='footer_message'>
 		<small>
-		
+
 		<?php
 			echo '<a href="https://c4g.github.io/BLIS/faq/">FAQ</a>';
 		?>
@@ -93,7 +95,7 @@ LangUtil::setPageId("footer");
 			$endtime = microtime();
 			$endarray = explode(" ", $endtime);
 			$endtime = $endarray[1] + $endarray[0];
-			$totaltime = $endtime - $starttime; 
+			$totaltime = $endtime - $starttime;
 			$totaltime = round($totaltime,5);
 			$page_name = $_SERVER['PHP_SELF'];
 			$page_name_parts = explode("/", $page_name);
