+++ htdocs/includes/page_elems.php
@@ -9,7 +9,7 @@ require_once("../regn/field_htmlFactory.php");
 require_once("../regn/generate_customize_field_order_patient.php");
 require_once("db_lib.php");
 require_once("field_order_update.php");
-require_once("../lang/lang_util.php");
+include_once("../lang/lang_util.php");
 
 
 class PageElems
@@ -10758,9 +10758,9 @@ $name_list = array("yyyy_to".$count, "mm_to".$count, "dd_to".$count);
 		foreach($zip_list as $backup)
 		{
 			echo "<input type='radio' id='backup-$count' name='$field_name' value='" .
-				$backup->file_name()."'" . 
+				$backup->file_name()."'" .
 				($count == 0 ? " checked " : "") .
-				">" . 
+				">" .
 				"<label for='backup-$count'>" . $backup->friendlyDate() .
 				"</label><br>\n";
 			$count++;
