+++ htdocs/reports/reports_plugform.php
@@ -1,9 +1,12 @@
 <?php
 #
-# PHP code for a new report plug in 
-# In the file reports/reports.php: 
+# PHP code for a new report plug in
+# In the file reports/reports.php:
 # copy this code after the line saying PLUG_FORM_DIV
 # Substitute "[reportname]" by the report name (single word or with underscore)
+
+include_once("../lang/lang_util.php");
+
 ?>
 
 <?php # Report form ?>
@@ -13,11 +16,11 @@
 	<br><br>
 	<form name="[reportname]_form" id="[reportname]_form" action="reports/reports_[reportname].php" method='post' target='_blank'>
 		<table cellpadding="4px">
-			<?php 
+			<?php
 			# Form entry table: Each row <tr> is for a single field
 			# Remove unneeded rows <tr> to </tr>
 			?>
-		
+
 			<?php # Select lab faciity ?>
 			<tr>
 				<td><?php echo LangUtil::$generalTerms['FACILITY']; ?> </td>
@@ -29,7 +32,7 @@
 					</select>
 				</td>
 			</tr>
-			
+
 			<?php # Select from_date ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['FROM_DATE']; ?> </td>
@@ -42,7 +45,7 @@
 				?>
 				</td>
 			</tr>
-			
+
 			<?php # Select to_date ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['TO_DATE']; ?> </td>
@@ -55,7 +58,7 @@
 				?>
 				</td>
 			</tr>
-			
+
 			<?php # Enter specimen_id ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['SPECIMEN_ID']; ?> </td>
@@ -63,7 +66,7 @@
 					<input type='text' name='specimen_id' value=''></input>
 				</td>
 			</tr>
-			
+
 			<?php # Enter patient_id ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['PATIENT_ID']; ?> </td>
@@ -71,7 +74,7 @@
 					<input type='text' name='patient_id' value=''></input>
 				</td>
 			</tr>
-			
+
 			<?php # Enter patient_name ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['PATIENT_NAME']; ?> </td>
@@ -79,7 +82,7 @@
 					<input type='text' name='p_name' value=''></input>
 				</td>
 			</tr>
-			
+
 			<?php # Enter specimen_id ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['SPECIMEN_ID']; ?> </td>
@@ -87,7 +90,7 @@
 					<input type='text' name='specimen_id' value=''></input>
 				</td>
 			</tr>
-			
+
 			<?php # Select test_type ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['TEST_TYPE']; ?> </td>
@@ -97,7 +100,7 @@
 					</select>
 				</td>
 			</tr>
-			
+
 			<?php # Select specimen_type ?>
 			<tr valign='top'>
 				<td><?php echo LangUtil::$generalTerms['SPECIMEN_TYPE']; ?> </td>
@@ -107,7 +110,7 @@
 					</select>
 				</td>
 			</tr>
-			
+
 			<?php # Submit button ?>
 			<tr>
 				<td></td>
@@ -118,5 +121,5 @@
 			</tr>
 		</table>
 	</form>
-	
+
 </div>
