+++ htdocs/update/default.php
@@ -1,18 +1,18 @@
 <?php
-$LANG_ARRAY = array ( 
-		"general" => array ( 
-		"TIPS" => "Tips", 
-		"USERNAME" => "Username", 
-		"PWD" => "Password", 
-		"PWD_TEMP" => "Temp Password", 
-		"PWD_RESET" => "Reset Password", 
+$LANG_ARRAY = array (
+		"general" => array (
+		"TIPS" => "Tips",
+		"USERNAME" => "Username",
+		"PWD" => "Password",
+		"PWD_TEMP" => "Temp Password",
+		"PWD_RESET" => "Reset Password",
 		"CNFRMPWD" => "Confirm Password",
 		"PASSWORDRESET_ERR" => "Password reset expired. Mail c4gbackup@gmail.com with the subject 'Password' to enable password recovery",
 		"PASSWORDRESET_COMPLETE" => "Password Reset complete. Visit login.php to reset the password",
-		"NAME" => "Name", 
+		"NAME" => "Name",
 		"ADD_CURRENCY" => "Add a new currency >>",
 		"NEW_CURRENCY" => "New Currency",
-		"PHONE" => "Phone No.", 
+		"PHONE" => "Phone No.",
 		"DEFAULT_CURRENCY" => "Default Currency Name:",
 		"CURRENCY_DELIMITER" => "Currency Delimiter:",
 		"CURRENCY" => "Currency",
@@ -21,144 +21,144 @@ $LANG_ARRAY = array (
 		"UPDATED_DATE" => "Updated Date",
 		"ACTIONS" => "Actions",
 		"ADD_CURRENCY_RATE" => "Add new currency rate",
-		"EMAIL" => "Email", 
-		"CMD_LOGIN" => "Login", 
-		"CMD_SEARCH" => "Search", 
-		"CMD_SUBMIT" => "Submit", 
-		"CMD_SUBMITTING" => "Submitting", 
-		"CMD_UPDATE" => "Update", 
-		"MSG_UPDATED" => "Updated", 
-		"CMD_CANCEL" => "Cancel", 
-		"CMD_SKIP" => "Skip", 
-		"CMD_HIDE" => "Hide", 
-		"CMD_SHOW" => "Show", 
-		"REPORT_TYPE" => "Report Type", 
-		"REPORT_TO" => "Report To", 
-		"MSG_REQDFIELD" => "This field is required", 
-		"REPORTED" => "Reported", 
-		"REPORTED_NOT" => "Not Yet Reported", 
-		"SKIPPED" => "Skipped", 
-		"CMD_BACK" => "Back", 
-		"CMD_GETREPORT" => "Get Report", 
+		"EMAIL" => "Email",
+		"CMD_LOGIN" => "Login",
+		"CMD_SEARCH" => "Search",
+		"CMD_SUBMIT" => "Submit",
+		"CMD_SUBMITTING" => "Submitting",
+		"CMD_UPDATE" => "Update",
+		"MSG_UPDATED" => "Updated",
+		"CMD_CANCEL" => "Cancel",
+		"CMD_SKIP" => "Skip",
+		"CMD_HIDE" => "Hide",
+		"CMD_SHOW" => "Show",
+		"REPORT_TYPE" => "Report Type",
+		"REPORT_TO" => "Report To",
+		"MSG_REQDFIELD" => "This field is required",
+		"REPORTED" => "Reported",
+		"REPORTED_NOT" => "Not Yet Reported",
+		"SKIPPED" => "Skipped",
+		"CMD_BACK" => "Back",
+		"CMD_GETREPORT" => "Get Report",
 		"CMD_GET_SPECIMEN_REPORT" => "Get Specimen Report",
 		"CMD_GET_TEST_REPORT" => "Get Test Report",
-		"MSG_SELECT_TTYPE" => "Select test type first", 
-		"CMD_EDIT" => "Edit", 
-		"CMD_EXPORTWORD" => "Export as Word Document", 
-		"CMD_EXPORTTXT" => "Export as Text File", 
-		"CMD_EXPORTPDF" => "Export as PDF", 
-		"CMD_SELECT" => "Select", 
-		"CMD_DELETE" => "Delete", 
-		"CMD_ENTER" => "Enter", 
-		"CMD_OK" => "OK", 
-		"CMD_PRINT" => "Print", 
-		"CMD_SEARCHING" => "Searching", 
-		"CMD_FETCH" => "Fetch", 
-		"CMD_FETCHING" => "Fetching", 
-		"MSG_DELETED" => "Deleted", 
-		"MSG_RESET" => "Reset", 
-		"SELECT_ONE" => "Select one", 
-		"CMD_CLOSEPAGE" => "Close This Page", 
-		"CMD_CLOSEPREVIEW" => "Close Preview", 
-		"FACILITY" => "Facility", 
-		"LOCATION" => "Location", 
-		"LAB_MGR" => "Lab Manager", 
-		"LAB_DIR" => "Country Director", 
-		"TYPE" => "Type", 
-		"LAB_STATUS" => "Lab Status", 
-		"SP_STATUS" => "Status", 
-		"MANAGE" => "Manage", 
-		"CMD_ADDNEWACCOUNT" => "Add New Account", 
-		"CMD_THISTORY" => "Test History", 
-		"DOCTOR" => "Physician", 
-		"PATIENT" => "Patient", 
-		"PATIENTS" => "Patients", 
-		"PATIENT_NAME" => "Patient Name", 
-		"PATIENT_ID" => "Patient ID", 
+		"MSG_SELECT_TTYPE" => "Select test type first",
+		"CMD_EDIT" => "Edit",
+		"CMD_EXPORTWORD" => "Export as Word Document",
+		"CMD_EXPORTTXT" => "Export as Text File",
+		"CMD_EXPORTPDF" => "Export as PDF",
+		"CMD_SELECT" => "Select",
+		"CMD_DELETE" => "Delete",
+		"CMD_ENTER" => "Enter",
+		"CMD_OK" => "OK",
+		"CMD_PRINT" => "Print",
+		"CMD_SEARCHING" => "Searching",
+		"CMD_FETCH" => "Fetch",
+		"CMD_FETCHING" => "Fetching",
+		"MSG_DELETED" => "Deleted",
+		"MSG_RESET" => "Reset",
+		"SELECT_ONE" => "Select one",
+		"CMD_CLOSEPAGE" => "Close This Page",
+		"CMD_CLOSEPREVIEW" => "Close Preview",
+		"FACILITY" => "Facility",
+		"LOCATION" => "Location",
+		"LAB_MGR" => "Lab Manager",
+		"LAB_DIR" => "Country Director",
+		"TYPE" => "Type",
+		"LAB_STATUS" => "Lab Status",
+		"SP_STATUS" => "Status",
+		"MANAGE" => "Manage",
+		"CMD_ADDNEWACCOUNT" => "Add New Account",
+		"CMD_THISTORY" => "Test History",
+		"DOCTOR" => "Physician",
+		"PATIENT" => "Patient",
+		"PATIENTS" => "Patients",
+		"PATIENT_NAME" => "Patient Name",
+		"PATIENT_ID" => "Patient ID",
 		"PATIENT_BARCODE" => "Patient Barcode",
 		"PATIENT_SIGNATURE" => "Patient Signature",
-		"ADDL_ID" => "Additional ID", 
-		"GENDER" => "Gender", 
-		"MALE" => "Male", 
-		"FEMALE" => "Female", 
-		"PATIENT_DAILYNUM" => "Patient Number", 
-		"AGE" => "Age", 
-		"OR" => "Or", 
-		"DOB" => "Date of Birth", 
-		"YEARS" => "Years", 
-		"MONTHS" => "Months", 
+		"ADDL_ID" => "Additional ID",
+		"GENDER" => "Gender",
+		"MALE" => "Male",
+		"FEMALE" => "Female",
+		"PATIENT_DAILYNUM" => "Patient Number",
+		"AGE" => "Age",
+		"OR" => "Or",
+		"DOB" => "Date of Birth",
+		"YEARS" => "Years",
+		"MONTHS" => "Months",
 		"WEEKS" => "Weeks",
-		"DAYS" => "Days", 
+		"DAYS" => "Days",
 		"HOURS" => "Hours",
-		"MINUTES" => "Minutes", 
-		"DROPDOWN" => "Drop-down Options", 
-		"FREETEXT" => "Free-text", 
-		"NUMERIC_FIELD" => "Numeric", 
-		"MULTISELECT" => "Multi-select", 
-		"ACCESSION_NUM" => "Accession No.", 
-		"SPECIMEN" => "Specimen", 
-		"IF_AVLBL" => "if available", 
-		"SPECIMENS" => "Specimens", 
-		"SPECIMEN_ID" => "Specimen ID", 
-		"ADDL_ID_S" => "Additional ID", 
-		"SPECIMEN_TYPE" => "Specimen Type", 
-		"SPECIMEN_TYPES" => "Specimen Types", 
-		"TEST" => "Test", 
-		"TESTS" => "Tests", 
-		"TEST_TYPE" => "Test Type", 
-		"TEST_TYPES" => "Test Types", 
+		"MINUTES" => "Minutes",
+		"DROPDOWN" => "Drop-down Options",
+		"FREETEXT" => "Free-text",
+		"NUMERIC_FIELD" => "Numeric",
+		"MULTISELECT" => "Multi-select",
+		"ACCESSION_NUM" => "Accession No.",
+		"SPECIMEN" => "Specimen",
+		"IF_AVLBL" => "if available",
+		"SPECIMENS" => "Specimens",
+		"SPECIMEN_ID" => "Specimen ID",
+		"ADDL_ID_S" => "Additional ID",
+		"SPECIMEN_TYPE" => "Specimen Type",
+		"SPECIMEN_TYPES" => "Specimen Types",
+		"TEST" => "Test",
+		"TESTS" => "Tests",
+		"TEST_TYPE" => "Test Type",
+		"TEST_TYPES" => "Test Types",
 		"CONTROL_TESTING_RESULTS" => "Control Results",
-		"DATE" => "Date", 
-		"DATE_FORMAT" => "Date Format", 
-		"R_DATE" => "Lab Receipt Date", 
-		"C_DATE" => "Collection Date", 
-		"C_TIME" => "Collection Time", 
-		"E_DATE" => "Results Entry Date", 
-		"RE_DATE" => "Date Reported", 
-		"G_DATE" => "Date", 
-		"CUMULATIVE" => "Cumulative", 
-		"REGD_BY" => "Registered By", 
-		"COMMENTS" => "Comments", 
-		"TECHNICIAN" => "Technician", 
-		"TECHNICIANS" => "Technicians", 
-		"RESULTS" => "Results", 
-		"RESULT_COMMENTS" => "Remarks", 
-		"RANGE" => "Range", 
-		"RANGE_OUTOF" => "Out of Range", 
-		"DETAILS" => "Details", 
-		"OPTIONAL" => "Optional", 
-		"REF_OUT" => "Referred Out", 
-		"PENDING_RESULTS" => "Results Pending", 
-		"PENDING_VER" => "Verification Pending", 
-		"DONE" => "Completed", 
-		"REF_RETURNED" => "Referred out and Returned", 
-		"MEASURES" => "Measures", 
-		"VALUES" => "Values", 
-		"OPTIONS" => "Options", 
-		"UNIT" => "Unit", 
-		"RANGE_NUMERIC" => "Numeric Range", 
-		"RANGE_ALPHANUM" => "Alphanumeric Values", 
-		"RANGE_AUTOCOMPLETE" => "Autocomplete", 
-		"YES" => "Yes", 
-		"NO" => "No", 
-		"ERROR" => "Error", 
-		"INVALID" => "Invalid", 
-		"NOTKNOWN" => "Not Known", 
-		"REGD_BY" => "Registered By", 
-		"ENTERED_BY" => "Entered By", 
-		"VERIFIED_BY" => "Verified By", 
-		"ACTION" => "Action", 
-		"RECORDS" => "Records", 
-		"RECORDS_TEST" => "Test Records", 
-		"RECORDS_SPECIMEN" => "Specimen Records", 
-		"RECORDS_PATIENT" => "Patient Records", 
-		"PATIENT_BARCODE" => "Patient Barcode", 
-		"LAB_SECTION" => "Lab Section", 
-		"CMD_VIEW" => "View", 
-		"CMD_ADD" => "Add", 
+		"DATE" => "Date",
+		"DATE_FORMAT" => "Date Format",
+		"R_DATE" => "Lab Receipt Date",
+		"C_DATE" => "Collection Date",
+		"C_TIME" => "Collection Time",
+		"E_DATE" => "Results Entry Date",
+		"RE_DATE" => "Date Reported",
+		"G_DATE" => "Date",
+		"CUMULATIVE" => "Cumulative",
+		"REGD_BY" => "Registered By",
+		"COMMENTS" => "Comments",
+		"TECHNICIAN" => "Technician",
+		"TECHNICIANS" => "Technicians",
+		"RESULTS" => "Results",
+		"RESULT_COMMENTS" => "Remarks",
+		"RANGE" => "Range",
+		"RANGE_OUTOF" => "Out of Range",
+		"DETAILS" => "Details",
+		"OPTIONAL" => "Optional",
+		"REF_OUT" => "Referred Out",
+		"PENDING_RESULTS" => "Results Pending",
+		"PENDING_VER" => "Verification Pending",
+		"DONE" => "Completed",
+		"REF_RETURNED" => "Referred out and Returned",
+		"MEASURES" => "Measures",
+		"VALUES" => "Values",
+		"OPTIONS" => "Options",
+		"UNIT" => "Unit",
+		"RANGE_NUMERIC" => "Numeric Range",
+		"RANGE_ALPHANUM" => "Alphanumeric Values",
+		"RANGE_AUTOCOMPLETE" => "Autocomplete",
+		"YES" => "Yes",
+		"NO" => "No",
+		"ERROR" => "Error",
+		"INVALID" => "Invalid",
+		"NOTKNOWN" => "Not Known",
+		"REGD_BY" => "Registered By",
+		"ENTERED_BY" => "Entered By",
+		"VERIFIED_BY" => "Verified By",
+		"ACTION" => "Action",
+		"RECORDS" => "Records",
+		"RECORDS_TEST" => "Test Records",
+		"RECORDS_SPECIMEN" => "Specimen Records",
+		"RECORDS_PATIENT" => "Patient Records",
+		"PATIENT_BARCODE" => "Patient Barcode",
+		"LAB_SECTION" => "Lab Section",
+		"CMD_VIEW" => "View",
+		"CMD_ADD" => "Add",
 		"PREVIEW" => "Preview",
-		"ADD_CURRENCY" => "Add a new currency >>",		
-		"ALL" => "All", 
+		"ADD_CURRENCY" => "Add a new currency >>",
+		"ALL" => "All",
 		"NEW_CURRENCY" => "New Currency",
 		"CURRENCY_DELIMITER" => "Currency Delimiter:",
 		"CURRENCY" => "Currency",
@@ -166,48 +166,48 @@ $LANG_ARRAY = array (
 		"UPDATED_DATE" => "Updated Date",
 		"ACTIONS" => "Actions",
 		"ADD_CURRENCY_RATE" => "Add new currency rate",
-		"WIDTH" => "Width", 
-		"HEIGHT" => "Height", 
-		"FONT_SIZE" => "Font Size", 
+		"WIDTH" => "Width",
+		"HEIGHT" => "Height",
+		"FONT_SIZE" => "Font Size",
 		"COL_COUNT" => "Column-Count",
-		"ALL_TESTS" => "All Tests", 
-		"CMD_TRACK" => "Track Actions", 
-		"MSG_NOTFOUND" => "Not Found", 
-		"RELATED_TESTS_NOTFOUND" => "Related Tests Not Found", 
-		"MSG_SIMILARNOTFOUND" => "No similar entries found", 
-		"REF_TO" => "Referred To", 
-		"CMD_REMOVE" => "Remove", 
-		"CMD_VERIFY" => "Verify", 
+		"ALL_TESTS" => "All Tests",
+		"CMD_TRACK" => "Track Actions",
+		"MSG_NOTFOUND" => "Not Found",
+		"RELATED_TESTS_NOTFOUND" => "Related Tests Not Found",
+		"MSG_SIMILARNOTFOUND" => "No similar entries found",
+		"REF_TO" => "Referred To",
+		"CMD_REMOVE" => "Remove",
+		"CMD_VERIFY" => "Verify",
 		"CMD_REFRESH" => "Refresh",
-		"ENTER_RESULTS" => "Enter Results", 
-		"MSG_ADDED" => "Added", 
-		"FROM_DATE" => "From", 
-		"TO_DATE" => "To", 
-		"DESCRIPTION" => "Description", 
-		"PANEL_TEST" => "Panel Test", 
-		"MSG_MANDATORYFIELD" => "Mandatory Field", 
-		"SWITCH_TOTECH" => "Work as Technician", 
-		"TIPS_SELECTSITE" => "Please select site location", 
-		"TIPS_INCOMPLETEINFO" => "Incomplete information. Please enter and submit again.", 
-		"TIPS_DATEINVALID" => "Error: Invalid date entered", 
-		"ADDNEW" => "Add New", 
-		"ADDANOTHER" => "Add Another", 
-		"COMPATIBLE_SPECIMENS" => "Compatible Specimens", 
-		"COMPATIBLE_TESTS" => "Compatible Tests", 
-		"TIPS_ACC_CONFIRMDELETE" => "Are you sure you want to delete this account?", 
-		"LAB_TECH" => "Lab Technician", 
-		"LAB_RECEPTIONIST" => "Lab Receptionist", 
-		"MSG_ACC_ADDED" => "Account Added", 
-		"MSG_ACC_DELETED" => "Account Deleted", 
-		"MSG_ACC_UPDATED" => "Account Updated", 
-		"TAT" => "Turnaround Time", 
-		"TAT_AVG" => "Average TAT", 
-		"TAT_TARGET" => "Target TAT", 
-		"TAT_ILE" => "90th percentile TAT", 
-		"PREVALENCE_RATE" => "Prevalence Rate", 
-		"LANGUAGE" => "Language", 
-		"USER_FUNCTIONS" => "Functions", 
-		"APPROX" => "approximate", 
+		"ENTER_RESULTS" => "Enter Results",
+		"MSG_ADDED" => "Added",
+		"FROM_DATE" => "From",
+		"TO_DATE" => "To",
+		"DESCRIPTION" => "Description",
+		"PANEL_TEST" => "Panel Test",
+		"MSG_MANDATORYFIELD" => "Mandatory Field",
+		"SWITCH_TOTECH" => "Work as Technician",
+		"TIPS_SELECTSITE" => "Please select site location",
+		"TIPS_INCOMPLETEINFO" => "Incomplete information. Please enter and submit again.",
+		"TIPS_DATEINVALID" => "Error: Invalid date entered",
+		"ADDNEW" => "Add New",
+		"ADDANOTHER" => "Add Another",
+		"COMPATIBLE_SPECIMENS" => "Compatible Specimens",
+		"COMPATIBLE_TESTS" => "Compatible Tests",
+		"TIPS_ACC_CONFIRMDELETE" => "Are you sure you want to delete this account?",
+		"LAB_TECH" => "Lab Technician",
+		"LAB_RECEPTIONIST" => "Lab Receptionist",
+		"MSG_ACC_ADDED" => "Account Added",
+		"MSG_ACC_DELETED" => "Account Deleted",
+		"MSG_ACC_UPDATED" => "Account Updated",
+		"TAT" => "Turnaround Time",
+		"TAT_AVG" => "Average TAT",
+		"TAT_TARGET" => "Target TAT",
+		"TAT_ILE" => "90th percentile TAT",
+		"PREVALENCE_RATE" => "Prevalence Rate",
+		"LANGUAGE" => "Language",
+		"USER_FUNCTIONS" => "Functions",
+		"APPROX" => "approximate",
 		"LANDSCAPE" => "Landscape mode",
 		"TIPS_PATIENTID_EXISTS"=>"Patient ID already exist. Look up for patient to confirm and register specimen",
         "LAB_STAFF" => "Lab Staff",
@@ -215,55 +215,55 @@ $LANG_ARRAY = array (
             "ALL_SITES" => "All Sites",
             "DISTRICT" => "District",
             "REGION" => "Region"
-	) , 
-	"header" => array ( 
-		"TITLE" => "", 
-		"LOGGEDINAS" => "Logged in as", 
-		"EDITPROFILE" => "Edit Profile", 
-		"LOGOUT" => "Logout", 
-		"COMMENTS" => "Comments", 
-		"SWITCH_TOMGR" => "Work as Manager", 
-		"MENU_BACKUP" => "Backup Data", 
-		"SWITCH_TOTECH" => "Work as Technician", 
+	) ,
+	"header" => array (
+		"TITLE" => "",
+		"LOGGEDINAS" => "Logged in as",
+		"EDITPROFILE" => "Edit Profile",
+		"LOGOUT" => "Logout",
+		"COMMENTS" => "Comments",
+		"SWITCH_TOMGR" => "Work as Manager",
+		"MENU_BACKUP" => "Backup Data",
+		"SWITCH_TOTECH" => "Work as Technician",
 		"SWITCH_TODIR" => "Work as Director"
-	) , 
-	"footer" => array ( 
-		"TITLE" => "", 
-		"FOOTER_MSG" => "A joint initiative of C4G @ Georgia Tech, the CDC and participating countries", 
+	) ,
+	"footer" => array (
+		"TITLE" => "",
+		"FOOTER_MSG" => "A joint initiative of C4G @ Georgia Tech, the CDC and participating countries",
 		"SWITCHLANG" => "Commutez au Francais"
-	) , 
-	"login" => array ( 
-		"TITLE" => "", 
-		"MSG_PLSLOGIN" => "Please log in first", 
-		"MSG_TIMED_OUT" => "Session timed out due to inactivity", 
-		"MSG_ERR_PWD" => "Incorrect username/password", 
-		"MSG_ERR_PWDRST" => "Password cannot be reset now. Contact your admin or mail c4gbackup@gmail.com with the subject 'Password'", 
-		"MSG_CAPSLOCK" => "Caps Lock is On", 
-		"MSG_NEWPWD" => "Request a New Password", 
+	) ,
+	"login" => array (
+		"TITLE" => "",
+		"MSG_PLSLOGIN" => "Please log in first",
+		"MSG_TIMED_OUT" => "Session timed out due to inactivity",
+		"MSG_ERR_PWD" => "Incorrect username/password",
+		"MSG_ERR_PWDRST" => "Password cannot be reset now. Contact your admin or mail c4gbackup@gmail.com with the subject 'Password'",
+		"MSG_CAPSLOCK" => "Caps Lock is On",
+		"MSG_NEWPWD" => "Request a New Password",
 		"MSG_RESETPWD" => "Reset Password",
 		"TIPS_NEWPWD" => "If you forgot your password, click on the Request a New Password link and a new, temporary password will be sent to your email address."
-	) , 
-	"home" => array ( 
-		"TITLE" => "Home", 
-		"WELCOME" => "Welcome", 
-		"TIPS_BLISINTRO" => "The Basic Laboratory Information System (BLIS) tracks patient specimens and laboratory results.", 
+	) ,
+	"home" => array (
+		"TITLE" => "Home",
+		"WELCOME" => "Welcome",
+		"TIPS_BLISINTRO" => "The Basic Laboratory Information System (BLIS) tracks patient specimens and laboratory results.",
 		"TIPS_PWD" => "You can update your profile and password by clicking on Edit Profile."
-	) , 
-	"regn" => array ( 
+	) ,
+	"regn" => array (
 		"TITLE" => "Registration"
-	) , 
-	"find_patient" => array ( 
-		"TITLE" => "Patient Look-up", 
-		"ADD_NEW_PATIENT" => "Add New Patient", 
+	) ,
+	"find_patient" => array (
+		"TITLE" => "Patient Look-up",
+		"ADD_NEW_PATIENT" => "Add New Patient",
 		"TIPS_REGISTRATION_1" => "This page allows us to register new patients or lookup existing patients based on name, patient ID or number.",
 		"TIPS_REGISTRATION_2" => "Once a patient has been registered, we can use this page to view,edit the patients profile.",
-		"TIPS_PATIENT_LOOKUP" => "To register a new specimen, enter Patient by ID/Name or add a new Patient first.", 
-		"MSG_NOMATCH" => "No match found", 
-		"CMD_REGISTERSPECIMEN" => "Register Specimen", 
+		"TIPS_PATIENT_LOOKUP" => "To register a new specimen, enter Patient by ID/Name or add a new Patient first.",
+		"MSG_NOMATCH" => "No match found",
+		"CMD_REGISTERSPECIMEN" => "Register Specimen",
 		"CMD_VIEWPROFILE" => "View Profile",
 		"CMD_DELPROFILE" => "Delete Profile",
 		"CMD_UPDPROFILE" => "Update Profile"
-	) , 
+	) ,
 		"doctor_register" => array (
 				"TITLE" => "Patient Look-up",
 				"ADD_NEW_PATIENT" => "Add New Patient",
@@ -276,185 +276,185 @@ $LANG_ARRAY = array (
 				"CMD_DELPROFILE" => "Delete Profile",
 				"CMD_UPDPROFILE" => "Update Profile"
 		) ,
-	"new_patient" => array ( 
-		"TITLE" => "New Patient", 
-		"MSG_BACKTOLOOKUP" => "Back to Patient Look-up", 
-		"MSG_ADDNEWENTRY" => "Please add new entry", 
-		"CMD_SELECTFORREGN" => "Select for Registration", 
-		"SIMILAR_NAMES" => "Similar Names", 
-		"MSG_PID_INVALID" => "Patient ID invalid/already exists", 
+	"new_patient" => array (
+		"TITLE" => "New Patient",
+		"MSG_BACKTOLOOKUP" => "Back to Patient Look-up",
+		"MSG_ADDNEWENTRY" => "Please add new entry",
+		"CMD_SELECTFORREGN" => "Select for Registration",
+		"SIMILAR_NAMES" => "Similar Names",
+		"MSG_PID_INVALID" => "Patient ID invalid/already exists",
 		"TIPS_REGN_NEW" => "Patient ID, Name and Gender are mandatory fields",
-		"TIPS_REGN" => "You can move to the next field by pressing <i>Tab</i> key or clicking on it using the mouse.", 
+		"TIPS_REGN" => "You can move to the next field by pressing <i>Tab</i> key or clicking on it using the mouse.",
 		"TIPS_DOB_AGE" => "Only one of <i>Age</i> or <i>Date of Birth</i> is required for entry."
-	) , 
-	"new_specimen" => array ( 
-		"TITLE" => "Specimen Registration", 
-		"ADD_ANOTHER_SPECIMEN" => "Add Another Specimen", 
-		"MSG_SELECT_STYPE" => "Select specimen type first", 
-		"MSG_SURETO_REMOVE" => "Are you sure you want to remove this specimen form?", 
-		"MSG_ADD_PATIENT" => "Please select or add a patient first.", 
-		"MSG_STYPE_MISSING" => "Specimen type missing", 
-		"MSG_NOTESTS_SELECTED" => "No tests selected", 
-		"MSG_SID_MISSING" => "Specimen ID missing", 
-		"MSG_SID_INVALID" => "Specimen ID invalid/already exists", 
-		"MSG_RDATE_INVALID" => "Lab receipt date invalid", 
-		"MSG_CDATE_INVALID" => "Collection date invalid", 
-		"MSG_NOTESTMATCH" => "No compatible tests found", 
+	) ,
+	"new_specimen" => array (
+		"TITLE" => "Specimen Registration",
+		"ADD_ANOTHER_SPECIMEN" => "Add Another Specimen",
+		"MSG_SELECT_STYPE" => "Select specimen type first",
+		"MSG_SURETO_REMOVE" => "Are you sure you want to remove this specimen form?",
+		"MSG_ADD_PATIENT" => "Please select or add a patient first.",
+		"MSG_STYPE_MISSING" => "Specimen type missing",
+		"MSG_NOTESTS_SELECTED" => "No tests selected",
+		"MSG_SID_MISSING" => "Specimen ID missing",
+		"MSG_SID_INVALID" => "Specimen ID invalid/already exists",
+		"MSG_RDATE_INVALID" => "Lab receipt date invalid",
+		"MSG_CDATE_INVALID" => "Collection date invalid",
+		"MSG_NOTESTMATCH" => "No compatible tests found",
 		"TIPS_SURETOABORT" => "Are you sure you want to abort this registration?",
 		"TIPS_REGISTRATION_SPECIMEN" => "Please select a specimen type to see the tests for that specimen.",
 		"TIPS_REGISTRATION_SPECIMEN_1" => "To add more than one specimen click on Add Another Specimen option."
-	) , 
-	"specimen_added" => array ( 
-		"TITLE" => "Specimens Registered", 
+	) ,
+	"specimen_added" => array (
+		"TITLE" => "Specimens Registered",
 		"MSG_NEXTSPECIMEN" => "Register Next Specimen"
-	) , 
-	"results_entry" => array ( 
-		"TITLE" => "Results", 
-		"MENU_SINGLESPECIMEN" => "Single Specimen Results", 
+	) ,
+	"results_entry" => array (
+		"TITLE" => "Results",
+		"MENU_SINGLESPECIMEN" => "Single Specimen Results",
 		"MENU_LABSECTION" => "Lab Section-wise Results",
-		"MENU_BATCHRESULTS" => "Batch Results", 
-		"MENU_VERIFYRESULTS" => "Verify Results", 
+		"MENU_BATCHRESULTS" => "Batch Results",
+		"MENU_VERIFYRESULTS" => "Verify Results",
 		"CONTROL_TESTING_RESULTS" => "Control Results",
 		"RELATED_RESULTS" => "Related Test Results Entry",
 		"RELATED_RESULTS_EMPTY" => "No related tests for this specimen",
-		"MENU_REPORTRESULTS" => "Report Results", 
-		"MENU_WORKSHEET" => "Worksheet", 
-		"BLANK_WORKSHEET" => "Keep Blank", 
-		"NUM_ROWS" => "Number of Rows", 
-		"MSG_PENDINGNOTFOUND" => "No pending specimens found", 
-		"MSG_ALREADYENTERED" => "Results Present", 
-		"MSG_RESULTSUBMITTED" => "Results Submitted", 
-		"MSG_RESULTMISSING" => "Result values missing", 
-		"TIPS_VERIFYCONFIRM" => "Final results. Are you sure?", 
-		"TIPS_VERIFYNOTFOUND" => "No unverified results found", 
-		"TIPS_VERIFYDONE" => "Results marked as verified", 
-		"TIPS_NOSPECIMENSTOTEST" => "No pending specimens to test", 
-		"TIPS_NOUNREPORTEDFOUND" => "No unreported specimens found.", 
-		"CMD_MARKASREPORTED" => "Mark as Reported", 
-		"TIPS_MARKEDREPORTED" => "Selected specimens marked as Reported", 
-		"MSG_RESULTSUBMITTEDALL" => "All results submitted for this patient", 
+		"MENU_REPORTRESULTS" => "Report Results",
+		"MENU_WORKSHEET" => "Worksheet",
+		"BLANK_WORKSHEET" => "Keep Blank",
+		"NUM_ROWS" => "Number of Rows",
+		"MSG_PENDINGNOTFOUND" => "No pending specimens found",
+		"MSG_ALREADYENTERED" => "Results Present",
+		"MSG_RESULTSUBMITTED" => "Results Submitted",
+		"MSG_RESULTMISSING" => "Result values missing",
+		"TIPS_VERIFYCONFIRM" => "Final results. Are you sure?",
+		"TIPS_VERIFYNOTFOUND" => "No unverified results found",
+		"TIPS_VERIFYDONE" => "Results marked as verified",
+		"TIPS_NOSPECIMENSTOTEST" => "No pending specimens to test",
+		"TIPS_NOUNREPORTEDFOUND" => "No unreported specimens found.",
+		"CMD_MARKASREPORTED" => "Mark as Reported",
+		"TIPS_MARKEDREPORTED" => "Selected specimens marked as Reported",
+		"MSG_RESULTSUBMITTEDALL" => "All results submitted for this patient",
 		"CUSTOM_WORKSHEET" => "Custom Worksheet"
-	) , 
-	"search" => array ( 
-		"TITLE" => "Search", 
+	) ,
+	"search" => array (
+		"TITLE" => "Search",
 		"TIPS_SEARCH_1" => "You can use this page to search for registered patients. Enter a partial name to generate a list of patients containing the string.",
 		"TIPS_SEARCH_2" => "Once you select the patient, you can view the patient's profile, look up the tests he has undergone and the results of those tests.",
 		"TIPS_SEARCH" => "Search for a Specimen by ID or Patients by ID/Name"
-	) , 
-	"patient_profile" => array ( 
-		"TITLE" => "Patient Profile", 
-		"MSG_REGNEWSPECIMEN" => "Register New Specimen", 
-		"MSG_UPDATEPROFILE" => "Update Profile", 
-		"MSG_PRINTHISTORY" => "Print Patient Report", 
+	) ,
+	"patient_profile" => array (
+		"TITLE" => "Patient Profile",
+		"MSG_REGNEWSPECIMEN" => "Register New Specimen",
+		"MSG_UPDATEPROFILE" => "Update Profile",
+		"MSG_PRINTHISTORY" => "Print Patient Report",
 		"TIPS_AGEORDOB" => "Please enter either Age or Date of Birth"
-	) , 
-	"specimen_info" => array ( 
-		"TITLE" => "Specimen Info", 
-		"REGDTESTS" => "Registered Tests", 
-		"TIPS_NOTESTSREGD" => "No registered tests found for this specimen", 
-		"TIPS_VERIFYDONE" => "Specimen results verified and updated", 
-		"TIPS_ENTRYDONE" => "Specimen results entered", 
-		"TITLE_VERIFY" => "Verify Specimen Results", 
-		"TITLE_ENTRY" => "Specimen Results Entry", 
-		"TIPS_VERIFY" => "Verify results for this specimen and make corrections if any", 
+	) ,
+	"specimen_info" => array (
+		"TITLE" => "Specimen Info",
+		"REGDTESTS" => "Registered Tests",
+		"TIPS_NOTESTSREGD" => "No registered tests found for this specimen",
+		"TIPS_VERIFYDONE" => "Specimen results verified and updated",
+		"TIPS_ENTRYDONE" => "Specimen results entered",
+		"TITLE_VERIFY" => "Verify Specimen Results",
+		"TITLE_ENTRY" => "Specimen Results Entry",
+		"TIPS_VERIFY" => "Verify results for this specimen and make corrections if any",
 		"TIPS_ENTRY" => "Enter results for this specimen and make corrections to already existing results if any"
-	) , 
-	"lab_configs" => array ( 
-		"TITLE" => "Lab Configurations", 
-		"CMD_ADDNEWLAB" => "Add New Lab", 
-		"CMD_VIEWALL" => "View All", 
-		"NEW_LAB_CONFIGURATION" => "New Lab Configuration", 
-		"MENU_SITEINFO" => "Site Information", 
-		"MENU_FIELDS" => "Confirmation", 
-		"TIPS_CUSTOM" => "Note: Custom fields can be added after the lab has been configured, using Manage lab configuration option.", 
-		"TIPS_CONFIRMNEWLAB" => "Please review and confirm all settings and click on Add to create lab configuration.", 
-		"TIPS_CREATINGLAB" => "Please wait while the system creates lab configuration. This may take a few minutes.", 
-		"TIPS_MISSING_FACILITY" => "Error: Facility Name missing", 
-		"TIPS_MISSING_LOCATION" => "Error: Facility Name missing", 
-		"TIPS_MISSING_MGR" => "Error: Lab Manager not selected", 
-		"TIPS_MISSING_STYPES" => "Error: No specimen types selected", 
+	) ,
+	"lab_configs" => array (
+		"TITLE" => "Lab Configurations",
+		"CMD_ADDNEWLAB" => "Add New Lab",
+		"CMD_VIEWALL" => "View All",
+		"NEW_LAB_CONFIGURATION" => "New Lab Configuration",
+		"MENU_SITEINFO" => "Site Information",
+		"MENU_FIELDS" => "Confirmation",
+		"TIPS_CUSTOM" => "Note: Custom fields can be added after the lab has been configured, using Manage lab configuration option.",
+		"TIPS_CONFIRMNEWLAB" => "Please review and confirm all settings and click on Add to create lab configuration.",
+		"TIPS_CREATINGLAB" => "Please wait while the system creates lab configuration. This may take a few minutes.",
+		"TIPS_MISSING_FACILITY" => "Error: Facility Name missing",
+		"TIPS_MISSING_LOCATION" => "Error: Facility Name missing",
+		"TIPS_MISSING_MGR" => "Error: Lab Manager not selected",
+		"TIPS_MISSING_STYPES" => "Error: No specimen types selected",
 		"TIPS_MISSING_TTYPES" => "Error: No test types selected"
-	) , 
-	"lab_config_home" => array ( 
-		"TITLE" => "Lab Configuration", 
-		"MENU_SUMMARY" => "Summary", 
+	) ,
+	"lab_config_home" => array (
+		"TITLE" => "Lab Configuration",
+		"MENU_SUMMARY" => "Summary",
 		"TIPS_SUMMARY_1" => "Gives a summary of lab setup including the lab personnel accounts, test and specimen types.",
 		"TIPS_SUMMARY_2" => "-",
 		"TIPS_SUMMARY_3" => "-",
-		"MENU_BACKUP_REVERT" => "Revert To Backup", 
+		"MENU_BACKUP_REVERT" => "Revert To Backup",
 		"TIPS_REVERT" => "In case of system failure, if we want to revert to a previously backedup copy of the data we can use this option. You can choose to backup current version before doing this. You may also choose to include to revert back the language settings.",
-		"BACKUP_LOCATION" => "Backup Version", 
-		"INCLUDE_LANGUAGE_SETTINGS" => "Include Language Settings", 
+		"BACKUP_LOCATION" => "Backup Version",
+		"INCLUDE_LANGUAGE_SETTINGS" => "Include Language Settings",
 		"MENU_ST_TYPES" => "Specimen/Test Types",
 		"TIPS_SPECIMENTESTTYPES" => "Specimen / Test Types - This option allows us to configure the Specimens can be collected and Tests that are performed at the particular lab.",
 		"TIPS_TARGETTAT" => "Target TAT - This option list the turnaround time for each test type. We can modify the TAT using the Edit option and changing the value.",
 		"TIPS_RESULTINTERPRETATION" => "Result Interpretation - This option allows us to specify the interpretation for multiple ranges of values for each test type.",
-		"MENU_USERS" => "User Accounts", 
-		"ADD_USER_TYPE" => "Add New User Type", 
+		"MENU_USERS" => "User Accounts",
+		"ADD_USER_TYPE" => "Add New User Type",
 		"TIPS_USERACCOUNTS_1" => "This page gives a summary of the various users given access to the system.",
 		"TIPS_USERACCOUNTS_2" => "It also allows for creation of new user accounts.",
 		"TIPS_USERACCOUNTS_3" => "It allows to edit account settings, delete accounts and monitor account activity.",
-		"MENU_CUSTOM" => "Registration Fields", 
+		"MENU_CUSTOM" => "Registration Fields",
 		"TIPS_REGISTRATIONFIELDS_1" => "This page allows for the configuration of the patient registration page. It allows for creating mandatory fields and hides certain fields if they are not desired to be filled.",
 		"TIPS_REGISTRATIONFIELDS_2" => "It also allows for creation of certain custom fields for Patient registration and new Specimen addition which may be needed by certain labs only.",
 		"TIPS_REGISTRATIONFIELDS_3" => "-",
-		"MENU_TAT" => "Target TAT", 
+		"MENU_TAT" => "Target TAT",
 		"MENU_INFECTION" => "Infection Report",
 		"TIPS_INFECTIONREPORT" => "This is an editable page for setting the various parameters associaated with each individual test type.",
 		"MENU_GENERAL" => "General Settings",
-		"Reports" => "Reports",		
+		"Reports" => "Reports",
 		"Tests"=>"Tests",
 		"New Stock"=>"New Stock",
 		"Edit Stock"=>"Edit Stock",
 		"Inventory"=>"Inventory",
-		"MENU_REPORTCONFIG" => "Daily Report Settings", 
+		"MENU_REPORTCONFIG" => "Daily Report Settings",
 		"TIPS_DAILYREPORTSETTINGS" => "This is an editable page for setting the layout of the Patient Report, Daily log of Specimens and Daily log of Patients.",
-		"MENU_WORKSHEETCONFIG" => "Worksheet", 
+		"MENU_WORKSHEETCONFIG" => "Worksheet",
 		"TIPS_WORKSHEETS" => "This is an editable page for creating custom worksheets.",
-		"MENU_MGR" => "Change Manager", 
-		"MENU_DEL" => "Delete Configuration", 
-		"MENU_EXPORTCONFIG" => "Export Configuration", 
-		"RECENT_ACTIVITY" => "Recent Activity", 
-		"SP_REGNS" => "Specimen Registrations", 
-		"RE_ENTRIES" => "Result Entries", 
-		"TIPS_RECENTACT" => "No recent user activity found", 
-		"NEW_LAB_USER" => "New Lab User", 
-		"EDIT_LAB_USER" => "New Lab User", 
-		"TIPS_CUSTOMFIELD_NOTFOUND" => "No custom fields exist", 
-		"CUSTOMFIELDS" => "Custom Fields", 
-		"NEW_CUSTOMFIELD" => "New Custom Field", 
-		"EDIT_CUSTOMFIELD" => "Edit Custom Field", 
-		"ASSIGN_TO" => "Assign To", 
-		"TIPS_LABDELETE" => "Are you sure you want to delete this lab configuration? All data and technician accounts will be deleted permanently.", 
-		"TIPS_SPECIMENSNOTSELECTED" => "Error: No specimen types selected", 
-		"TIPS_TESTSNOTSELECTED" => "Error: No test types selected", 
-		"TIPS_USERACC" => "Edit user account details or reset password by entering a new one.", 
-		"GROUP_BYGENDER" => "Group By Gender", 
-		"GROUP_BYAGE" => "Group By Age", 
-		"RANGE_AGE" => "Age Range", 
-		"RANGE_NOTREQD" => "No range configuration required.", 
-		"MSG_PENDINGTAT" => "Default TAT for pending tests", 
-		"MODIFYLANG" => "Modify Language", 
-		"MSG_LANGUPDATED" => "Language strings updated.", 
-		"OTHER_FIELDS" => "Other Fields", 
-		"MARGINS" => "Margins", 
-		"COLUMN_WIDTH" => "Column Width", 
-		"TOP" => "Top", 
-		"BOTTOM" => "Bottom", 
-		"LEFT" => "Left", 
-		"RIGHT" => "Right", 
-		"DAILY" => "Daily", 
-		"WEEKLY" => "Weekly", 
-		"MONTHLY" => "Monthly", 
-		"YEARLY" => "Yearly", 
-		"USE_PNAME_RESULTS" => "Display Name at Results Entry", 
-		"CUSTOM_WORKSHEETS" => "Custom Worksheets", 
-		"NEW_CUSTOMWORKSHEET" => "Add Custom Worksheet", 
-		"ADDED_CUSTOMWORKSHEET" => "Custom Worksheet Added", 
-		"EDIT_CUSTOMWORKSHEET" => "Edit Custom Worksheet", 
-		"UPDATED_CUSTOMWORKSHEET" => "Custom Worksheet Updated", 
-		"BACK_TOCONFIG" => "Back to Configuration", 
-		"BACKUP_CURRENT_VERSION" => "Backup current version before reverting?", 
+		"MENU_MGR" => "Change Manager",
+		"MENU_DEL" => "Delete Configuration",
+		"MENU_EXPORTCONFIG" => "Export Configuration",
+		"RECENT_ACTIVITY" => "Recent Activity",
+		"SP_REGNS" => "Specimen Registrations",
+		"RE_ENTRIES" => "Result Entries",
+		"TIPS_RECENTACT" => "No recent user activity found",
+		"NEW_LAB_USER" => "New Lab User",
+		"EDIT_LAB_USER" => "New Lab User",
+		"TIPS_CUSTOMFIELD_NOTFOUND" => "No custom fields exist",
+		"CUSTOMFIELDS" => "Custom Fields",
+		"NEW_CUSTOMFIELD" => "New Custom Field",
+		"EDIT_CUSTOMFIELD" => "Edit Custom Field",
+		"ASSIGN_TO" => "Assign To",
+		"TIPS_LABDELETE" => "Are you sure you want to delete this lab configuration? All data and technician accounts will be deleted permanently.",
+		"TIPS_SPECIMENSNOTSELECTED" => "Error: No specimen types selected",
+		"TIPS_TESTSNOTSELECTED" => "Error: No test types selected",
+		"TIPS_USERACC" => "Edit user account details or reset password by entering a new one.",
+		"GROUP_BYGENDER" => "Group By Gender",
+		"GROUP_BYAGE" => "Group By Age",
+		"RANGE_AGE" => "Age Range",
+		"RANGE_NOTREQD" => "No range configuration required.",
+		"MSG_PENDINGTAT" => "Default TAT for pending tests",
+		"MODIFYLANG" => "Modify Language",
+		"MSG_LANGUPDATED" => "Language strings updated.",
+		"OTHER_FIELDS" => "Other Fields",
+		"MARGINS" => "Margins",
+		"COLUMN_WIDTH" => "Column Width",
+		"TOP" => "Top",
+		"BOTTOM" => "Bottom",
+		"LEFT" => "Left",
+		"RIGHT" => "Right",
+		"DAILY" => "Daily",
+		"WEEKLY" => "Weekly",
+		"MONTHLY" => "Monthly",
+		"YEARLY" => "Yearly",
+		"USE_PNAME_RESULTS" => "Display Name at Results Entry",
+		"CUSTOM_WORKSHEETS" => "Custom Worksheets",
+		"NEW_CUSTOMWORKSHEET" => "Add Custom Worksheet",
+		"ADDED_CUSTOMWORKSHEET" => "Custom Worksheet Added",
+		"EDIT_CUSTOMWORKSHEET" => "Edit Custom Worksheet",
+		"UPDATED_CUSTOMWORKSHEET" => "Custom Worksheet Updated",
+		"BACK_TOCONFIG" => "Back to Configuration",
+		"BACKUP_CURRENT_VERSION" => "Backup current version before reverting?",
 		"TIPS_REVERTDONE" => "Backup reverted successfully. Please shutdown the system and restart to complete the reverting process.",
 		"TIPS_SETUPNETWORK_3" => "You will now be able to access BLIS by clicking on BlisSetup.html and entering your username and password.",
 		"TIPS_SETUPNETWORK_4" => "In case of a computer restart or network failure click on Setup Network option again.",
@@ -484,37 +484,37 @@ $LANG_ARRAY = array (
         "PRINT_UNVERIFIED" => "Allow printing of unverified tests",
 		"IP_SAVED" => "IP address saved."
 	) ,
-	"lab_admins" => array ( 
-		"TITLE" => "Lab Managers", 
-		"LABS_ASSIGNED" => "Labs Assigned", 
-		"TIPS_MANAGELAB" => "To edit technician accounts, use Manage option in Lab Configurations page.", 
-		"TIPS_ACC_LABSEXIST" => "Lab Configurations exist under this admin account. Please delete or re-assign them before deleting this account.", 
-		"ADD_NEW_ADMIN" => "Add New Manager", 
+	"lab_admins" => array (
+		"TITLE" => "Lab Managers",
+		"LABS_ASSIGNED" => "Labs Assigned",
+		"TIPS_MANAGELAB" => "To edit technician accounts, use Manage option in Lab Configurations page.",
+		"TIPS_ACC_LABSEXIST" => "Lab Configurations exist under this admin account. Please delete or re-assign them before deleting this account.",
+		"ADD_NEW_ADMIN" => "Add New Manager",
 		"EDIT_ADMIN_ACC" => "Edit Lab Manager"
-	) , 
-	"catalog" => array ( 
-		"TITLE" => "Test Catalog", 
-		"MENU_REMOVEDATA" => "Delete Catalog Data", 
-		"TIPS_CATALOG" => "You can add, remove or modify specimen types, test types and valid ranges on this page.", 
-		"TIPS_TESTSNOTFOUND" => "No Test Types found in Catalog", 
-		"TIPS_SPECIMENSNOTFOUND" => "No Specimen Types found in Catalog", 
-		"TIPS_REMOVEDATA" => "Sure to remove catalog data? All specimen/test types will be deleted and would have to be entered again.", 
-		"NEW_TEST_TYPE" => "New Test Type", 
-		"NEW_SPECIMEN_TYPE" => "New Specimen Type", 
-		"EDIT_TEST_TYPE" => "Edit Test Type", 
-		"EDIT_SPECIMEN_TYPE" => "Edit Specimen Type", 
-		"NEW_LAB_SECTION" => "New Lab Section", 
-		"TIPS_MISSING_TESTNAME" => "Error: Test name missing", 
-		"TIPS_MISSING_SPECIMENNAME" => "Error: Specimen name missing", 
-		"TIPS_MISSING_CATNAME" => "Error: Lab Section missing", 
-		"TIPS_MISSING_SELECTEDSPECIMEN" => "Error: No compatible specimens selected ", 
-		"TIPS_MISSING_SELECTEDTESTS" => "Error: No compatible tests selected ", 
-		"TIPS_MISSING_SELECTEDMEASURES" => "Error: No measures selected for this test type", 
-		"TEST_TYPE_ADDED" => "Test Type Added", 
-		"SPECIMEN_TYPE_ADDED" => "Specimen Type Added", 
-		"TEST_TYPE_UPDATED" => "Test Type Updated", 
-		"SPECIMEN_TYPE_UPDATED" => "Specimen Type Updated", 
-		"CMD_BACK_TOCATALOG" => "Back to Catalog", 
+	) ,
+	"catalog" => array (
+		"TITLE" => "Test Catalog",
+		"MENU_REMOVEDATA" => "Delete Catalog Data",
+		"TIPS_CATALOG" => "You can add, remove or modify specimen types, test types and valid ranges on this page.",
+		"TIPS_TESTSNOTFOUND" => "No Test Types found in Catalog",
+		"TIPS_SPECIMENSNOTFOUND" => "No Specimen Types found in Catalog",
+		"TIPS_REMOVEDATA" => "Sure to remove catalog data? All specimen/test types will be deleted and would have to be entered again.",
+		"NEW_TEST_TYPE" => "New Test Type",
+		"NEW_SPECIMEN_TYPE" => "New Specimen Type",
+		"EDIT_TEST_TYPE" => "Edit Test Type",
+		"EDIT_SPECIMEN_TYPE" => "Edit Specimen Type",
+		"NEW_LAB_SECTION" => "New Lab Section",
+		"TIPS_MISSING_TESTNAME" => "Error: Test name missing",
+		"TIPS_MISSING_SPECIMENNAME" => "Error: Specimen name missing",
+		"TIPS_MISSING_CATNAME" => "Error: Lab Section missing",
+		"TIPS_MISSING_SELECTEDSPECIMEN" => "Error: No compatible specimens selected ",
+		"TIPS_MISSING_SELECTEDTESTS" => "Error: No compatible tests selected ",
+		"TIPS_MISSING_SELECTEDMEASURES" => "Error: No measures selected for this test type",
+		"TEST_TYPE_ADDED" => "Test Type Added",
+		"SPECIMEN_TYPE_ADDED" => "Specimen Type Added",
+		"TEST_TYPE_UPDATED" => "Test Type Updated",
+		"SPECIMEN_TYPE_UPDATED" => "Specimen Type Updated",
+		"CMD_BACK_TOCATALOG" => "Back to Catalog",
 		"TIPS_MEASURES" => "Only numbers and alphabets should be used while adding measures.For the range option please specify the lower and the upper bound.",
 		"TIPS_TC_TESTTYPE_1" => "Gives a complete listing of the available test types.",
 		"TIPS_TC_TESTTYPE_2" => "Use this page to edit existing test types.",
@@ -522,74 +522,74 @@ $LANG_ARRAY = array (
 		"TIPS_TC_SPECIMENTYPE_1" => "Gives a complete listing of the available specimen types.",
 		"TIPS_TC_SPECIMENTYPE_2" => "Use this page to edit existing specimen types. Tests compatible with a particular specimen type can added/deleted by checking the appropriate boxes located here.",
 		"TIPS_TC_SPECIMENTYPE_3" => "You can also add new Specimen types using the 'Add New' option located near the top of the Page."
-	) , 
-	"reports" => array ( 
-		"AGE_UNIT" => "Unit of Age", 
-		"GROUP_BYGENDER" => "Group By Gender", 
-		"GROUP_BYAGE" => "Group By Age", 
-		"TITLE" => "Reports", 
-		"MENU_DAILY" => "Daily Reports", 
-		"MENU_PATIENT" => "Patient Report", 
-		"MENU_SPECIMEN" => "Individual Specimen", 
-		"MENU_PHISTORY" => "Patient History", 
-		"MENU_TESTRECORDS" => "Test Records", 
-		"MENU_PENDINGTESTS" => "Pending Tests", 
-		"MENU_DAILYLOGS" => "Daily Log", 
-		"MENU_AGGREPORTS" => "Aggregate Reports", 
-		"MENU_INFECTIONSUMMARY" => "Prevalence Rate", 
+	) ,
+	"reports" => array (
+		"AGE_UNIT" => "Unit of Age",
+		"GROUP_BYGENDER" => "Group By Gender",
+		"GROUP_BYAGE" => "Group By Age",
+		"TITLE" => "Reports",
+		"MENU_DAILY" => "Daily Reports",
+		"MENU_PATIENT" => "Patient Report",
+		"MENU_SPECIMEN" => "Individual Specimen",
+		"MENU_PHISTORY" => "Patient History",
+		"MENU_TESTRECORDS" => "Test Records",
+		"MENU_PENDINGTESTS" => "Pending Tests",
+		"MENU_DAILYLOGS" => "Daily Log",
+		"MENU_AGGREPORTS" => "Aggregate Reports",
+		"MENU_INFECTIONSUMMARY" => "Prevalence Rate",
 		"MENU_TEST_SPECIMEN_AGGREGATE_REPORT_CONFIG" => "Test/Specimen Aggregate Configuration",
-		"MENU_SPECIMEN_AGGREGATE_REPORT_FORM" => "Specimen Aggregate Report", 
-		"MENU_TEST_AGGREGATE_REPORT_FORM" => "Test Aggregate Report", 
-		"MENU_COUNTS" => "Counts", 
+		"MENU_SPECIMEN_AGGREGATE_REPORT_FORM" => "Specimen Aggregate Report",
+		"MENU_TEST_AGGREGATE_REPORT_FORM" => "Test Aggregate Report",
+		"MENU_COUNTS" => "Counts",
 		"MENU_CONTROLREPORT" => "Control Report",
-		"MENU_TAT" => "Turnaround Time", 
-		"MENU_INFECTIONREPORT" => "Infection Report", 
-		"COUNT_TYPE" => "Count Type", 
-		"COUNT_TEST" => "Test Count", 
-		"COUNT_SPECIMEN" => "Specimen Count", 
-		"TIPS_PHISTORY" => "Select Site and Patient ID to print Patient Test History.", 
-		"TIPS_SPECIMEN" => "Select Specimen to view Report of all tests done on it", 
-		"TIPS_TESTRECORDS" => "Select Site, Test Type, Date Interval and Fields to print Test Records.", 
-		"TIPS_PENDINGTESTS" => "Select Site and Test Type to view tests pending at the site.", 
-		"TIPS_DAILYLOGS" => "Print all records handled on a given day.", 
-		"TIPS_INFECTIONSUMMARY" => "Select facility and date range to view infection graph and prevalence rates.", 
-		"TIPS_COUNTS" => "Select Site and Date Interval to view number of Specimens or Tests Handled over the specified duration.", 
-		"TIPS_TAT" => "Select Site and Date Interval to view average test-wise Turn around Times over the specified duration.", 
-		"TIPS_INFECTIONREPORT" => "Select tests to generate age and gender-wise Infection Report.", 
-		"TIPS_GETACCESSIONREPORT" => "Get Entire Accession Report", 
-		"RANGE_NORMAL" => "Normal Range", 
-		"REPORT_SAMPLE" => "Complete Sample Report", 
-		"TIPS_RECNOTFOUND" => "No records found", 
-		"TOTAL" => "Total", 
-		"TOTAL_TESTS" => "Total Tests", 
-		"TOTAL_PATIENTS" => "Total Patients", 
-		"TOTAL_SPECIMENS" => "Total Specimens", 
-		"TESTED_N" => "Tested Negative", 
-		"TESTED_P" => "Tested Positive", 
-		"TOTAL_MF" => "M/F Total", 
-		"RANGE_AGE" => "Age Ranges", 
-		"MSG_BACKTOREPORTS" => "Back To Reports", 
-		"TIPS_NONEWTESTS" => "No new tests registered on this date.", 
-		"TIPS_NONEWPATIENTS" => "No new patients registered on this date.", 
-		"MSG_SHOWTABLE" => "Show Table", 
-		"MSG_HIDETABLE" => "Hide Table", 
-		"MSG_SHOWGRAPH" => "Show Graph", 
-		"MSG_HIDEGRAPH" => "Hide Graph", 
-		"TIPS_NODISCRETE" => "Error: No alphanumeric value tests found", 
-		"PROGRESSION_M" => "Monthly", 
-		"PROGRESSION_W" => "Weekly", 
-		"PROGRESSION_D" => "Daily", 
-		"REPORT_PERIOD" => "Reporting Period", 
-		"TIPS_NOTATTESTS" => "No test entries found for the given date range.", 
-		"TIPS_TATEXCEED" => "No specimens exceeded target TAT.", 
-		"TIPS_TATEXCEEDNUM" => "Specimens exceeded target TAT.", 
-		"TIPS_TATNOPENDING" => "No specimens pending in the supplied date interval.", 
-		"TIPS_TATPENDING" => "Specimens pending in the supplied date interval.", 
-		"TIPS_CONFIGINFECTION" => "ERROR: Infection Report not configured for this facility. Please contact lab manager.", 
-		"TIPS_NOMATCHSPECIMENS" => "No matching specimens found", 
-		"TIPS_MULTPATIENTS" => "Selected specimens belong to different patients. Please try again.", 
-		"MSG_INCLUDEPENDING" => "Include Pending Tests", 
-		"TIPS_TABLESORT" => "Click on column heading to sort the table by that field.", 
+		"MENU_TAT" => "Turnaround Time",
+		"MENU_INFECTIONREPORT" => "Infection Report",
+		"COUNT_TYPE" => "Count Type",
+		"COUNT_TEST" => "Test Count",
+		"COUNT_SPECIMEN" => "Specimen Count",
+		"TIPS_PHISTORY" => "Select Site and Patient ID to print Patient Test History.",
+		"TIPS_SPECIMEN" => "Select Specimen to view Report of all tests done on it",
+		"TIPS_TESTRECORDS" => "Select Site, Test Type, Date Interval and Fields to print Test Records.",
+		"TIPS_PENDINGTESTS" => "Select Site and Test Type to view tests pending at the site.",
+		"TIPS_DAILYLOGS" => "Print all records handled on a given day.",
+		"TIPS_INFECTIONSUMMARY" => "Select facility and date range to view infection graph and prevalence rates.",
+		"TIPS_COUNTS" => "Select Site and Date Interval to view number of Specimens or Tests Handled over the specified duration.",
+		"TIPS_TAT" => "Select Site and Date Interval to view average test-wise Turn around Times over the specified duration.",
+		"TIPS_INFECTIONREPORT" => "Select tests to generate age and gender-wise Infection Report.",
+		"TIPS_GETACCESSIONREPORT" => "Get Entire Accession Report",
+		"RANGE_NORMAL" => "Normal Range",
+		"REPORT_SAMPLE" => "Complete Sample Report",
+		"TIPS_RECNOTFOUND" => "No records found",
+		"TOTAL" => "Total",
+		"TOTAL_TESTS" => "Total Tests",
+		"TOTAL_PATIENTS" => "Total Patients",
+		"TOTAL_SPECIMENS" => "Total Specimens",
+		"TESTED_N" => "Tested Negative",
+		"TESTED_P" => "Tested Positive",
+		"TOTAL_MF" => "M/F Total",
+		"RANGE_AGE" => "Age Ranges",
+		"MSG_BACKTOREPORTS" => "Back To Reports",
+		"TIPS_NONEWTESTS" => "No new tests registered on this date.",
+		"TIPS_NONEWPATIENTS" => "No new patients registered on this date.",
+		"MSG_SHOWTABLE" => "Show Table",
+		"MSG_HIDETABLE" => "Hide Table",
+		"MSG_SHOWGRAPH" => "Show Graph",
+		"MSG_HIDEGRAPH" => "Hide Graph",
+		"TIPS_NODISCRETE" => "Error: No alphanumeric value tests found",
+		"PROGRESSION_M" => "Monthly",
+		"PROGRESSION_W" => "Weekly",
+		"PROGRESSION_D" => "Daily",
+		"REPORT_PERIOD" => "Reporting Period",
+		"TIPS_NOTATTESTS" => "No test entries found for the given date range.",
+		"TIPS_TATEXCEED" => "No specimens exceeded target TAT.",
+		"TIPS_TATEXCEEDNUM" => "Specimens exceeded target TAT.",
+		"TIPS_TATNOPENDING" => "No specimens pending in the supplied date interval.",
+		"TIPS_TATPENDING" => "Specimens pending in the supplied date interval.",
+		"TIPS_CONFIGINFECTION" => "ERROR: Infection Report not configured for this facility. Please contact lab manager.",
+		"TIPS_NOMATCHSPECIMENS" => "No matching specimens found",
+		"TIPS_MULTPATIENTS" => "Selected specimens belong to different patients. Please try again.",
+		"MSG_INCLUDEPENDING" => "Include Pending Tests",
+		"TIPS_TABLESORT" => "Click on column heading to sort the table by that field.",
 		"MSG_ALLTESTSCOMPLETED" => "All tests completed",
         "MENU_TESTWISE_REPORTS" => "Test Specific Reports",
         "SITE" => "Site",
@@ -603,140 +603,140 @@ $LANG_ARRAY = array (
         "INTERPRETATION" => "Interpretation"
 
     ) ,
-	"edit_profile" => array ( 
-		"LINK_CHANGEPWD" => "Change Password", 
-		"TIPS_CASEPWD" => "Please note that passwords are case-sensitive.", 
-		"MSG_NOFIELDUPDATE" => "No profile fields to update", 
-		"CURRPWD" => "Current Password", 
-		"NEWPWD" => "New Password", 
-		"REENTERPWD" => "Re-enter Password", 
-		"MSG_REQDFIELD" => "Required field", 
-		"MSG_PWDNOMATCH" => "Passwords do not match", 
-		"MSG_PWDMINLENGTH" => "Min password length is 3", 
-		"MSG_PROFILEUPDATED" => "Profile Updated", 
-		"MSG_PWDUPDATEERROR" => "Error while updating Password", 
-		"MSG_PWDMATCHERROR" => "Error: Old password did not match", 
+	"edit_profile" => array (
+		"LINK_CHANGEPWD" => "Change Password",
+		"TIPS_CASEPWD" => "Please note that passwords are case-sensitive.",
+		"MSG_NOFIELDUPDATE" => "No profile fields to update",
+		"CURRPWD" => "Current Password",
+		"NEWPWD" => "New Password",
+		"REENTERPWD" => "Re-enter Password",
+		"MSG_REQDFIELD" => "Required field",
+		"MSG_PWDNOMATCH" => "Passwords do not match",
+		"MSG_PWDMINLENGTH" => "Min password length is 3",
+		"MSG_PROFILEUPDATED" => "Profile Updated",
+		"MSG_PWDUPDATEERROR" => "Error while updating Password",
+		"MSG_PWDMATCHERROR" => "Error: Old password did not match",
 		"MSG_PWDUPDATED" => "Password updated"
-	) , 
-	"help" => array ( 
+	) ,
+	"help" => array (
 		"TITLE" => "Help"
-	) , 
-	"stocks" => array ( 
-		"Reagent" => "Item", 
-		"Lot_Number" => "Lot Number", 
-		"Expiry_Date" => "Expiry Date", 
-		"Manufacturer" => "Manufacturer", 
-		"Supplier" => "Supplier", 
-		"Quantity_Supplied" => "Quantity Supplied", 
-		"Units" => "Units", 
-		"Quantity" => "Quantity", 
-		"Edit" => "Edit", 
-		"Inventory" => "Inventory", 
-		"Current_Inventory" => "Current Inventory", 
-		"Quantity_Signed_Out" => "Quantity Signed Out", 
-		"Receiver" => "Receiver", 
-		"Remarks" => "Remarks", 
-		"View_Current_Inventory" => "View Current Inventory", 
-		"Current_Stock" => "Current Stock", 
-		"Add_Stock" => "Add Stock", 
-		"Edit_Stock" => "Edit Stock", 
-		"Total_Quantity" => "Total Quantity", 
-		"Quantity_In_Stock" => "Quantity In Stock", 
-		"TIPS_ADD" => "Add the details of the stock in the given form. To add more that one stock items you can select Add Another.", 
+	) ,
+	"stocks" => array (
+		"Reagent" => "Item",
+		"Lot_Number" => "Lot Number",
+		"Expiry_Date" => "Expiry Date",
+		"Manufacturer" => "Manufacturer",
+		"Supplier" => "Supplier",
+		"Quantity_Supplied" => "Quantity Supplied",
+		"Units" => "Units",
+		"Quantity" => "Quantity",
+		"Edit" => "Edit",
+		"Inventory" => "Inventory",
+		"Current_Inventory" => "Current Inventory",
+		"Quantity_Signed_Out" => "Quantity Signed Out",
+		"Receiver" => "Receiver",
+		"Remarks" => "Remarks",
+		"View_Current_Inventory" => "View Current Inventory",
+		"Current_Stock" => "Current Stock",
+		"Add_Stock" => "Add Stock",
+		"Edit_Stock" => "Edit Stock",
+		"Total_Quantity" => "Total Quantity",
+		"Quantity_In_Stock" => "Quantity In Stock",
+		"TIPS_ADD" => "Add the details of the stock in the given form. To add more that one stock items you can select Add Another.",
 		"TIPS_UPDATE" => "Enter the quantity of the stock being checked out. The table below gives account of the current inventory.",
 		"Cost_per_Unit" => "Cost per Unit",
 		"Date_received" => "Date received"
-	) , 
-	"lang_modify" => array ( 
-		"TITLE" => "Modify Language", 
+	) ,
+	"lang_modify" => array (
+		"TITLE" => "Modify Language",
 		"CATEGORY" => "Category",
 		"TIPS_MODIFYLANG_1" => "This page allows to toggle between languages. It also allows you to change the content being displayed.",
 		"TIPS_MODIFYLANG_2" => "You can also choose to change the language for a few pages using this option. The pages are listed as a drop-down menu."
-	) , 
-	"blis_help_page" => array ( 
-		"TITLE" => "Help page", 
-		"MENU_LABCONFIG" => "Lab Configuration", 
-		"MENU_SUMMARY" => "Summary", 
-		"TIPS_SUMMARY_1" => "Gives a summary of lab setup including the lab personnel accounts, test and specimen types.", 
-		"TIPS_SUMMARY_2" => "-", 
-		"TIPS_SUMMARY_3" => "-", 
-		"MENU_TESTS" => "Tests", 
-		"TIPS_SPECIMENTESTTYPES" => "Specimen / Test Types - This option allows us to configure the Specimens can be collected and Tests that are performed at the particular lab.", 
-		"TIPS_TARGETTAT" => "Target TAT - This option list the turnaround time for each test type. We can modify the TAT using the Edit option and changing the value.", 
-		"TIPS_RESULTINTERPRETATION" => "Result Interpretation - This option allows us to specify the interpretation for multiple ranges of values for each test type.", 
-		"MENU_L_REPORTS" => "Reports", 
-		"MENU_R_INFECTIONREPORT" => "Infection Report", 
-		"TIPS_INFECTIONREPORT" => "Infection Report: This option generates an aggregate infection report for a particular period for one or all Lab sections. It also provides an option to create a Word document of the generated report.", 
-		"MENU_R_DRS" => "Daily Report Settings", 
-		"TIPS_DAILYREPORTSETTINGS" => "This is an editable page for setting the layout of the Patient Report, Daily log of Specimens and Daily log of Patients.", 
-		"MENU_R_WORKSHEETS" => "Worksheets", 
-		"TIPS_WORKSHEETS" => "This is an editable page for creating custom worksheets.", 
-		"MENU_USERACCOUNTS" => "User Accounts", 
-		"TIPS_USERACCOUNTS_1" => "This page gives a summary of the various users given access to the system.", 
-		"TIPS_USERACCOUNTS_2" => "It also allows for creation of new user accounts.", 
-		"TIPS_USERACCOUNTS_3" => "It allows to edit account settings, delete accounts and monitor account activity.", 
-		"MENU_REGISTRATIONFIELDS" => "Registration Fields", 
-		"TIPS_REGISTRATIONFIELDS_1" => "This page allows for the configuration of the patient registration page. It allows for creating mandatory fields and hides certain fields if they are not desired to be filled.", 
-		"TIPS_REGISTRATIONFIELDS_2" => "It also allows for creation of certain custom fields for Patient registration and new Specimen addition which may be needed by certain labs only.", 
-		"TIPS_REGISTRATIONFIELDS_3" => "-", 
-		"MENU_MODIFYLANG" => "Modify Language", 
-		"TIPS_MODIFYLANG_1" => "This page allows to toggle between languages.", 
-		"TIPS_MODIFYLANG_2" => "You can also choose to change the language for a few pages using this option. The pages are listed as a drop-down menu.", 
-		"MENU_SETUPNETWORK" => "Setup Network", 
-		"TIPS_SETUPNETWORK_1" => "Login as Admin.", 
-		"TIPS_SETUPNETWORK_2" => "Click on the Setup Network option in the Lab Configuration tab.", 
-		"TIPS_SETUPNETWORK_3" => "You will now be able to access BLIS by clicking on BlisSetup.html and entering your username and password.", 
-		"TIPS_SETUPNETWORK_4" => "In case of a computer restart or network failure repeat above steps again.", 
-		"MENU_EXPORTCONFIG" => "Export Configuration", 
-		"TIPS_EXPORTCONFIG" => "This option gives the configuration summary for the entire system as a Microsoft Word document.", 
-		"MENU_BACKUP" => "Backup Data", 
-		"TIPS_BACKUP" => "This option allows to take a backup of the entire system to prevent loss of data.", 
-		"MENU_REVERT" => "Revert to Backup", 
-		"TIPS_REVERT" => "In case of system failure, if we want to revert to a previously backedup copy of the data we can use this option.", 
-		"MENU_TESTCATALOG" => "Test Catalog", 
-		"MENU_TC_TESTTYPE" => "Test Type", 
-		"TIPS_TC_TESTTYPE_1" => "Gives a complete listing of the available test types.", 
-		"TIPS_TC_TESTTYPE_2" => "We can use this page to edit existing present test types.", 
-		"TIPS_TC_TESTTYPE_3" => "New test types can also be added using the Add New option located near the top of the Page.", 
-		"MENU_TC_SPECIMENTYPE" => "Specimen Type", 
-		"TIPS_TC_SPECIMENTYPE_1" => "Gives a complete listing of the available specimen types.", 
-		"TIPS_TC_SPECIMENTYPE_2" => "We can use this page to edit existing specimen types. Tests compatible with a particular specimen type can added/deleted by checking the appropriate boxes located here.", 
-		"TIPS_TC_SPECIMENTYPE_3" => "New specimen types can also be added using the Add New option located near the top of the Page.", 
-		"MENU_BACKUPDATA" => "Backup Data", 
-		"TIPS_BACKUPDATA" => "This option takes a backup of the system and stores it on the local drive. This option is only available to the Admin.", 
-		"MENU_REGISTRATION" => "Registration", 
-		"TIPS_REGISTRATION_1" => "This page allows us to register new patients or lookup existing patients based on name, patient ID or number.", 
-		"TIPS_REGISTRATION_2" => "Once a patient has been registered, we can use this page to view,edit the patients profile.", 
-		"TIPS_REGISTRATION_3" => "We can also register a specimen the patient has provided for a particular test.", 
-		"MENU_RESULTS" => "Results", 
-		"MENU_R_SINGLESPECIMENRESULTS" => "Single Specimen Results", 
-		"TIPS_R_SINGLESPECIMENRESULTS" => "This option allows us to add results for a particular patient based on the specimens he had provided.", 
-		"MENU_R_BATCHRESULTS" => "Batch Results", 
-		"TIPS_R_BATCHRESULTS" => "This option allows us to add results for a particular Test Type.", 
-		"MENU_R_VERIFYRESULTS" => "Verify Results", 
-		"TIPS_R_VERIFYRESULTS" => "This option allows us to verify the result based on the test type. It shows the list of results for all patients whose results have not been verified. You can modify the results and enter remarks before verifying the results.", 
-		"MENU_R_WORKSHEET" => "Worksheet", 
-		"TIPS_R_WORKSHEET_1" => "This option generates a worksheet based on the Lab Section and Test Type. You can also use a custom worksheet which can be created by Admins using Lab Configuration > Tests > Reports > Worksheet.", 
-		"TIPS_R_WORKSHEET_2" => "You can create a blank worksheet by choosing the Keep Blank option and specifying the number of rows you need.", 
-		"MENU_SEARCH" => "Search", 
+	) ,
+	"blis_help_page" => array (
+		"TITLE" => "Help page",
+		"MENU_LABCONFIG" => "Lab Configuration",
+		"MENU_SUMMARY" => "Summary",
+		"TIPS_SUMMARY_1" => "Gives a summary of lab setup including the lab personnel accounts, test and specimen types.",
+		"TIPS_SUMMARY_2" => "-",
+		"TIPS_SUMMARY_3" => "-",
+		"MENU_TESTS" => "Tests",
+		"TIPS_SPECIMENTESTTYPES" => "Specimen / Test Types - This option allows us to configure the Specimens can be collected and Tests that are performed at the particular lab.",
+		"TIPS_TARGETTAT" => "Target TAT - This option list the turnaround time for each test type. We can modify the TAT using the Edit option and changing the value.",
+		"TIPS_RESULTINTERPRETATION" => "Result Interpretation - This option allows us to specify the interpretation for multiple ranges of values for each test type.",
+		"MENU_L_REPORTS" => "Reports",
+		"MENU_R_INFECTIONREPORT" => "Infection Report",
+		"TIPS_INFECTIONREPORT" => "Infection Report: This option generates an aggregate infection report for a particular period for one or all Lab sections. It also provides an option to create a Word document of the generated report.",
+		"MENU_R_DRS" => "Daily Report Settings",
+		"TIPS_DAILYREPORTSETTINGS" => "This is an editable page for setting the layout of the Patient Report, Daily log of Specimens and Daily log of Patients.",
+		"MENU_R_WORKSHEETS" => "Worksheets",
+		"TIPS_WORKSHEETS" => "This is an editable page for creating custom worksheets.",
+		"MENU_USERACCOUNTS" => "User Accounts",
+		"TIPS_USERACCOUNTS_1" => "This page gives a summary of the various users given access to the system.",
+		"TIPS_USERACCOUNTS_2" => "It also allows for creation of new user accounts.",
+		"TIPS_USERACCOUNTS_3" => "It allows to edit account settings, delete accounts and monitor account activity.",
+		"MENU_REGISTRATIONFIELDS" => "Registration Fields",
+		"TIPS_REGISTRATIONFIELDS_1" => "This page allows for the configuration of the patient registration page. It allows for creating mandatory fields and hides certain fields if they are not desired to be filled.",
+		"TIPS_REGISTRATIONFIELDS_2" => "It also allows for creation of certain custom fields for Patient registration and new Specimen addition which may be needed by certain labs only.",
+		"TIPS_REGISTRATIONFIELDS_3" => "-",
+		"MENU_MODIFYLANG" => "Modify Language",
+		"TIPS_MODIFYLANG_1" => "This page allows to toggle between languages.",
+		"TIPS_MODIFYLANG_2" => "You can also choose to change the language for a few pages using this option. The pages are listed as a drop-down menu.",
+		"MENU_SETUPNETWORK" => "Setup Network",
+		"TIPS_SETUPNETWORK_1" => "Login as Admin.",
+		"TIPS_SETUPNETWORK_2" => "Click on the Setup Network option in the Lab Configuration tab.",
+		"TIPS_SETUPNETWORK_3" => "You will now be able to access BLIS by clicking on BlisSetup.html and entering your username and password.",
+		"TIPS_SETUPNETWORK_4" => "In case of a computer restart or network failure repeat above steps again.",
+		"MENU_EXPORTCONFIG" => "Export Configuration",
+		"TIPS_EXPORTCONFIG" => "This option gives the configuration summary for the entire system as a Microsoft Word document.",
+		"MENU_BACKUP" => "Backup Data",
+		"TIPS_BACKUP" => "This option allows to take a backup of the entire system to prevent loss of data.",
+		"MENU_REVERT" => "Revert to Backup",
+		"TIPS_REVERT" => "In case of system failure, if we want to revert to a previously backedup copy of the data we can use this option.",
+		"MENU_TESTCATALOG" => "Test Catalog",
+		"MENU_TC_TESTTYPE" => "Test Type",
+		"TIPS_TC_TESTTYPE_1" => "Gives a complete listing of the available test types.",
+		"TIPS_TC_TESTTYPE_2" => "We can use this page to edit existing present test types.",
+		"TIPS_TC_TESTTYPE_3" => "New test types can also be added using the Add New option located near the top of the Page.",
+		"MENU_TC_SPECIMENTYPE" => "Specimen Type",
+		"TIPS_TC_SPECIMENTYPE_1" => "Gives a complete listing of the available specimen types.",
+		"TIPS_TC_SPECIMENTYPE_2" => "We can use this page to edit existing specimen types. Tests compatible with a particular specimen type can added/deleted by checking the appropriate boxes located here.",
+		"TIPS_TC_SPECIMENTYPE_3" => "New specimen types can also be added using the Add New option located near the top of the Page.",
+		"MENU_BACKUPDATA" => "Backup Data",
+		"TIPS_BACKUPDATA" => "This option takes a backup of the system and stores it on the local drive. This option is only available to the Admin.",
+		"MENU_REGISTRATION" => "Registration",
+		"TIPS_REGISTRATION_1" => "This page allows us to register new patients or lookup existing patients based on name, patient ID or number.",
+		"TIPS_REGISTRATION_2" => "Once a patient has been registered, we can use this page to view,edit the patients profile.",
+		"TIPS_REGISTRATION_3" => "We can also register a specimen the patient has provided for a particular test.",
+		"MENU_RESULTS" => "Results",
+		"MENU_R_SINGLESPECIMENRESULTS" => "Single Specimen Results",
+		"TIPS_R_SINGLESPECIMENRESULTS" => "This option allows us to add results for a particular patient based on the specimens he had provided.",
+		"MENU_R_BATCHRESULTS" => "Batch Results",
+		"TIPS_R_BATCHRESULTS" => "This option allows us to add results for a particular Test Type.",
+		"MENU_R_VERIFYRESULTS" => "Verify Results",
+		"TIPS_R_VERIFYRESULTS" => "This option allows us to verify the result based on the test type. It shows the list of results for all patients whose results have not been verified. You can modify the results and enter remarks before verifying the results.",
+		"MENU_R_WORKSHEET" => "Worksheet",
+		"TIPS_R_WORKSHEET_1" => "This option generates a worksheet based on the Lab Section and Test Type. You can also use a custom worksheet which can be created by Admins using Lab Configuration > Tests > Reports > Worksheet.",
+		"TIPS_R_WORKSHEET_2" => "You can create a blank worksheet by choosing the Keep Blank option and specifying the number of rows you need.",
+		"MENU_SEARCH" => "Search",
 		"TIPS_SEARCH_1" => "You can use this page to search for registered patients. Enter a partial name to generate a list of patients containing the string.",
 		"TIPS_SEARCH_2" => "Once you select the patient, you can view the patient's profile, look up the tests he has undergone and the results of those tests.",
-		"TIPS_SEARCH_3" => "-", 
-		"MENU_INVENTORY" => "Inventory", 
-		"TIPS_INVENTORY_1" => "This page displays the current item quantities in stock.", 
-		"TIPS_INVENTORY_2" => "You can update stock as you acquire new items by adding the item name,quantity received, receiver name and remarks.", 
-		"TIPS_INVENTORY_3" => "-", 
-		"MENU_REPORTS" => "Reports", 
-		"MENU_DAILYREPORTS" => "Daily Reports", 
-		"TIPS_PATIENTREPORT" => "Patient Report: This option allows us to generate the report for a particular patient based on either, the patient name, patient number or patient id.", 
-		"TIPS_DAILYLOG" => "Daily Log: This option generates a daily of either patient or test records from a particular start date to end date. You can customize the kind of report you want by choosing one option from the Section (e.g. Serology, Hematology) and/or Test type (e.g. ALT/SGPT, HGB) dropdown menus.", 
-		"MENU_AGGREGATEREPORTS" => "Aggregate Reports", 
-		"TIPS_PREVALENCERATE" => "Prevalence Rate: This report gives the prevalence of a particular infection type based on the number of tests done and the results. You can enter the specified period for which you want to check the prevalence rate of a particular infection. You can also views the Trends of the infection for the defined period, as a graph, by clicking the Trends option when the report is displayed.", 
-		"TIPS_COUNTS" => "Counts: This option generates a report for a particular period of time based on the count of tests, specimen or doctor statistics.", 
-		"TIPS_TURNAROUNDTIME" => "Turnaround Time: This option generates the turnaround time for all or chosen test for a particular period of time. It also generates a graph of the statistics. You can choose to include both pending and completed tests or just the completed tests.", 
-		"TIPS_INFECTIONREPORT" => "Infection Report: This option generates an aggregate infection report for a particular period for one or all Lab sections. It also provides an option to create a Word document of the generated report.", 
+		"TIPS_SEARCH_3" => "-",
+		"MENU_INVENTORY" => "Inventory",
+		"TIPS_INVENTORY_1" => "This page displays the current item quantities in stock.",
+		"TIPS_INVENTORY_2" => "You can update stock as you acquire new items by adding the item name,quantity received, receiver name and remarks.",
+		"TIPS_INVENTORY_3" => "-",
+		"MENU_REPORTS" => "Reports",
+		"MENU_DAILYREPORTS" => "Daily Reports",
+		"TIPS_PATIENTREPORT" => "Patient Report: This option allows us to generate the report for a particular patient based on either, the patient name, patient number or patient id.",
+		"TIPS_DAILYLOG" => "Daily Log: This option generates a daily of either patient or test records from a particular start date to end date. You can customize the kind of report you want by choosing one option from the Section (e.g. Serology, Hematology) and/or Test type (e.g. ALT/SGPT, HGB) dropdown menus.",
+		"MENU_AGGREGATEREPORTS" => "Aggregate Reports",
+		"TIPS_PREVALENCERATE" => "Prevalence Rate: This report gives the prevalence of a particular infection type based on the number of tests done and the results. You can enter the specified period for which you want to check the prevalence rate of a particular infection. You can also views the Trends of the infection for the defined period, as a graph, by clicking the Trends option when the report is displayed.",
+		"TIPS_COUNTS" => "Counts: This option generates a report for a particular period of time based on the count of tests, specimen or doctor statistics.",
+		"TIPS_TURNAROUNDTIME" => "Turnaround Time: This option generates the turnaround time for all or chosen test for a particular period of time. It also generates a graph of the statistics. You can choose to include both pending and completed tests or just the completed tests.",
+		"TIPS_INFECTIONREPORT" => "Infection Report: This option generates an aggregate infection report for a particular period for one or all Lab sections. It also provides an option to create a Word document of the generated report.",
 		"TIPS_INVENTORY" => "Inventory: This option generates a listing of the current inventory."
-		
+
 	) ,
 	"search_condition" => array (
 	"SEARCH_BEGIN_WITH"=>"Starts with",
@@ -745,5 +745,5 @@ $LANG_ARRAY = array (
 	)
 );
 
-include("../lang/lang_util.php");
+include_once("../lang/lang_util.php");
 ?>
