+++ htdocs/update/update.php
@@ -6,7 +6,7 @@
 
 require_once("redirect.php");
 require_once("../includes/script_elems.php");
-require_once("lang/lang_util.php");
+include_once("lang/lang_util.php");
 
 LangUtil::setPageId("update");
